### 2023-04-05 929d646
correctly parse string expressions for memory
