+++ docs/index.rst
@@ -1,7 +1,7 @@
 pygwb documentation
 ==============================================
 .. image:: pygwb_logo_plasma.png
-   :width: 80
+   :width: 8
 
 .. automodule:: pygwb
     :members:
