### 2023-04-14 2843bc2
Updating stat checks plotting
