### 2023-04-18 e35bae0
Changing plotting style for some plots
