### 2023-04-18 fd68635
Removing gate_pad h5 saving.
