+++ pygwb/baseline.py
@@ -1332,7 +1332,7 @@ class Baseline(object):
             save = self._json_save
             ext = ".json"
 
-        elif save_data_type == "hdf5":
+        elif save_data_type == "hdf5" or save_data_type == "h5":
             save = self._hdf5_save
             ext = ".h5"
 
@@ -1353,10 +1353,6 @@ class Baseline(object):
             self.sigma_spectrogram,
             self.badGPStimes,
             self.delta_sigmas,
-            self.interferometer_1.gates,
-            self.interferometer_1.gate_pad,
-            self.interferometer_2.gates,
-            self.interferometer_2.gate_pad,
         )
 
     def save_psds_csds(
@@ -1390,7 +1386,7 @@ class Baseline(object):
             save_csd = self._json_save_csd
             ext = ".json"
 
-        elif save_data_type == "hdf5":
+        elif save_data_type == "hdf5" or save_data_type == "h5":
             save_csd = self._hdf5_save_csd
             ext = ".h5"
 
@@ -1423,6 +1419,10 @@ class Baseline(object):
             self.interferometer_2.psd_spectrogram,
             self.interferometer_1.average_psd,
             self.interferometer_2.average_psd,
+            self.interferometer_1.gates,
+            self.interferometer_1.gate_pad,
+            self.interferometer_2.gates,
+            self.interferometer_2.gate_pad,
             coherence,
             psd_1_coh,
             psd_2_coh,
@@ -1443,10 +1443,6 @@ class Baseline(object):
         sigma_spectrogram,
         badGPStimes,
         delta_sigmas,
-        ifo_1_gates,
-        ifo_1_gate_pad,
-        ifo_2_gates,
-        ifo_2_gate_pad,
     ):
         try:
             naive_sigma_values = delta_sigmas["naive_sigmas"]
@@ -1471,10 +1467,6 @@ class Baseline(object):
             delta_sigma_values=delta_sigmas["values"],
             naive_sigma_values=naive_sigma_values,
             slide_sigma_values=slide_sigma_values,
-            ifo_1_gates=ifo_1_gates,
-            ifo_1_gate_pad=ifo_1_gate_pad,
-            ifo_2_gates=ifo_2_gates,
-            ifo_2_gate_pad=ifo_2_gate_pad,
         )
 
     def _pickle_save(
@@ -1490,10 +1482,6 @@ class Baseline(object):
         sigma_spectrogram,
         badGPStimes,
         delta_sigmas,
-        ifo_1_gates,
-        ifo_1_gate_pad,
-        ifo_2_gates,
-        ifo_2_gate_pad,
     ):
         save_dictionary = {
             "frequencies": frequencies,
@@ -1506,10 +1494,6 @@ class Baseline(object):
             "sigma_spectrogram": sigma_spectrogram,
             "badGPStimes": badGPStimes,
             "delta_sigmas": list(delta_sigmas.items()),
-            "ifo_1_gates": ifo_1_gates,
-            "ifo_1_gate_pad": ifo_1_gate_pad,
-            "ifo_2_gates": ifo_2_gates,
-            "ifo_2_gate_pad": ifo_2_gate_pad,
         }
 
         with open(filename, "wb") as f:
@@ -1528,10 +1512,6 @@ class Baseline(object):
         sigma_spectrogram,
         badGPStimes,
         delta_sigmas,
-        ifo_1_gates,
-        ifo_1_gate_pad,
-        ifo_2_gates,
-        ifo_2_gate_pad,
     ):
         list_freqs = frequencies.tolist()
[…]
