### 2023-04-23 43f5487
Corrected segment_duration to be int
