### 2023-04-24 c18edb5
adding coarse graining/welch to psd/csd
