+++ pygwb/baseline.py
@@ -35,10 +35,12 @@ class Baseline(object):
         frequencies=None,
         calibration_epsilon=0,
         notch_list_path="",
+        coarse_grain_psd=False,
+        coarse_grain_csd=True,
         overlap_factor=0.5,
         zeropad_csd=True,
         window_fftgram_dict={"window_fftgram": "hann"},
-        N_average_segments_welch_psd=2,
+        N_average_segments_psd=2,
         sampling_frequency=None,
     ):
         """
@@ -60,6 +62,10 @@ class Baseline(object):
             Calibration uncertainty for this baseline
         notch_list_path: str, optional
             File path of the baseline notch list
+        coarse_grain_psd: bool
+            Whether to apply coarse graining to obtain PSD spectra. Default is False.
+        coarse_grain_csd: bool
+            Whether to apply coarse graining to obtain CSD spectra. Default is True.
         overlap_factor: float, optional
             Factor by which to overlap the segments in the psd and csd estimation.
             Default is 1/2, if set to 0 no overlap is performed.
@@ -67,7 +73,7 @@ class Baseline(object):
             If True, applies zeropadding in the csd estimation. True by default.
         window_fftgram_dict: dictionary, optional
             Dictionary containing name and parameters describing which window to use when producing fftgrams for psds and csds. Default is \"hann\".
-        N_average_segments_welch_psd: int, optional
+        N_average_segments_psd: int, optional
             Number of segments used for PSD averaging (from both sides of the segment of interest)
             N_avg_segs should be even and >= 2.
         """
@@ -76,10 +82,12 @@ class Baseline(object):
         self.interferometer_2 = interferometer_2
         self.calibration_epsilon = calibration_epsilon
         self.notch_list_path = notch_list_path
+        self.coarse_grain_psd = coarse_grain_psd
+        self.coarse_grain_csd = coarse_grain_csd
         self.overlap_factor = overlap_factor
         self.zeropad_csd = zeropad_csd
         self.window_fftgram_dict = window_fftgram_dict
-        self.N_average_segments_welch_psd = N_average_segments_welch_psd
+        self.N_average_segments_psd = N_average_segments_psd
         self._tensor_orf_calculated = False
         self._vector_orf_calculated = False
         self._scalar_orf_calculated = False
@@ -268,7 +276,7 @@ class Baseline(object):
     def csd_segment_offset(self):
         if self._duration_set:
             stride = self.duration * (1 - self.overlap_factor)
-            return int(np.ceil(self.duration / stride)) * int(self.N_average_segments_welch_psd/2)
+            return int(np.ceil(self.duration / stride)) * int(self.N_average_segments_psd/2)
         else:
             raise ValueError("Trying to calculate CSD segment offset before setting duration. Need to set duration before attempting this.")
 
@@ -680,10 +688,12 @@ class Baseline(object):
             calibration_epsilon=parameters.calibration_epsilon,
             frequencies=frequencies,
             notch_list_path=parameters.notch_list_path,
+            coarse_grain_psd=parameters.coarse_grain_psd,
+            coarse_grain_csd=parameters.coarse_grain_csd,
             overlap_factor=parameters.overlap_factor,
             zeropad_csd=parameters.zeropad_csd,
             window_fftgram_dict=parameters.window_fft_dict,
-            N_average_segments_welch_psd=parameters.N_average_segments_welch_psd,
+            N_average_segments_psd=parameters.N_average_segments_psd,
             sampling_frequency=parameters.new_sample_rate,
         )
 
@@ -728,6 +738,7 @@ class Baseline(object):
         try:
             self.interferometer_1.set_psd_spectrogram(
[…]
