+++ test/test_data/parameters_baseline_test.ini
@@ -23,8 +23,9 @@ cluster_window: 0.5
 window_fftgram: hann
 [density_estimation]
 frequency_resolution: 0.03125
-N_average_segments_welch_psd: 2
-coarse_grain: 0
+N_average_segments_psd: 2
+coarse_grain_psd: False
+coarse_grain_csd: True
 overlap_factor: 0.5
 zeropad_csd: True
 [postprocessing]

[…]
