+++ pygwb/statistical_checks.py
@@ -136,10 +136,12 @@ class StatisticalChecks(object):
             self.running_sigmas,
         ) = self.compute_running_quantities()
 
+        self.time_tag = "{self.params.t0}-{self.params.tf}"
+
         if file_tag:
             self.file_tag = file_tag
         else:
-            self.file_tag = {self.sliding_times_all[0]}-{self.sliding_times_all[-1]}
+            self.file_tag = self.time_tag
 
         self.legend_fontsize = legend_fontsize
         self.axes_labelsize = legend_fontsize + 2
@@ -319,6 +321,7 @@ class StatisticalChecks(object):
         plt.ylabel(r"Point estimate $\pm 1.65 \sigma$", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
+        plt.title(f'Running point estimate in {self.time_tag}')
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_point_estimate.png", bbox_inches='tight'
         )
@@ -339,6 +342,7 @@ class StatisticalChecks(object):
         plt.ylabel(r"$\sigma$", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
+        plt.title(r'Running $\sigma$' + f'in {self.time_tag}')
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_sigma.png", bbox_inches = 'tight'
         )
@@ -354,9 +358,10 @@ class StatisticalChecks(object):
         plt.grid(True)
         plt.xlim(t_array[0], t_array[-1])
         plt.xlabel("Lag (s)", size=self.axes_labelsize)
-        plt.ylabel("IFFT of Integrand of Pt Estimate", size=self.axes_labelsize)
+        plt.ylabel(r"$\Omega$ integrand IFFT", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
+        plt.title(f"IFFT of the point estimate integrand in {self.time_tag}")
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-IFFT_point_estimate_integrand.png", bbox_inches='tight'
         )
@@ -377,6 +382,7 @@ class StatisticalChecks(object):
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
         plt.xscale("log")
+        plt.title(f"Absolute SNR in {self.time_tag}")
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-abs_point_estimate_integrand.png",
             bbox_inches="tight",
@@ -397,6 +403,7 @@ class StatisticalChecks(object):
         plt.xscale("log")
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
+        plt.title(f"Cumulative SNR in {self.time_tag}")
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-cumulative_SNR_spectrum.png",
             bbox_inches="tight",
@@ -417,6 +424,7 @@ class StatisticalChecks(object):
         plt.xscale("log")
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
+        plt.title(f"Real SNR in {self.time_tag}")
[…]
