+++ pygwb/statistical_checks.py
@@ -136,15 +136,16 @@ class StatisticalChecks(object):
             self.running_sigmas,
         ) = self.compute_running_quantities()
 
-        self.time_tag = "{self.params.t0}-{self.params.tf}"
+        self.time_tag = f"{int(self.params.t0)}"+"$-$"+f"{int(self.params.tf)}"
 
         if file_tag:
             self.file_tag = file_tag
         else:
-            self.file_tag = self.time_tag
+            self.file_tag = f"{self.params.t0}-{self.params.tf}"
 
         self.legend_fontsize = legend_fontsize
         self.axes_labelsize = legend_fontsize + 2
+        self.title_fontsize = legend_fontsize + 4
         self.annotate_fontsize = legend_fontsize - 4
 
     def get_data_after_dsc(self):
@@ -321,7 +322,7 @@ class StatisticalChecks(object):
         plt.ylabel(r"Point estimate $\pm 1.65 \sigma$", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
-        plt.title(f'Running point estimate in {self.time_tag}')
+        plt.title(f'Running point estimate in {self.time_tag}', fontsize=self.title_fontsize)
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_point_estimate.png", bbox_inches='tight'
         )
@@ -342,7 +343,7 @@ class StatisticalChecks(object):
         plt.ylabel(r"$\sigma$", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
-        plt.title(r'Running $\sigma$' + f'in {self.time_tag}')
+        plt.title(r'Running $\sigma$ ' + f'in {self.time_tag}', fontsize=self.title_fontsize)
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_sigma.png", bbox_inches = 'tight'
         )
@@ -361,7 +362,7 @@ class StatisticalChecks(object):
         plt.ylabel(r"$\Omega$ integrand IFFT", size=self.axes_labelsize)
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
-        plt.title(f"IFFT of the point estimate integrand in {self.time_tag}")
+        plt.title(r"$\Omega$ integrand IFFT" + f" in {self.time_tag}", fontsize=self.title_fontsize)
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-IFFT_point_estimate_integrand.png", bbox_inches='tight'
         )
@@ -382,7 +383,7 @@ class StatisticalChecks(object):
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
         plt.xscale("log")
-        plt.title(f"Absolute SNR in {self.time_tag}")
+        plt.title(f"Absolute SNR in {self.time_tag}", fontsize=self.title_fontsize)
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-abs_point_estimate_integrand.png",
             bbox_inches="tight",
@@ -403,7 +404,7 @@ class StatisticalChecks(object):
         plt.xscale("log")
         plt.xticks(fontsize=self.legend_fontsize)
         plt.yticks(fontsize=self.legend_fontsize)
-        plt.title(f"Cumulative SNR in {self.time_tag}")
[…]
