### 2023-04-25 bbc9046
Add missing string formatter
