### 2023-04-25 f2a6597
adding dsc statement into the stats
