### 2023-04-29 f44ab72
fixing issue with overlapping/non-overlapping segs
