+++ pygwb_pipe/parameters.ini
@@ -51,6 +51,7 @@ frequency_resolution: 0.03125
 N_average_segments_psd: 2
 coarse_grain_psd: False
 coarse_grain_csd: True
+overlap_factor_welch: 0.5
 overlap_factor: 0.5
 zeropad_csd: True
 [postprocessing]
