+++ test/test_data/parameters_detector_test.ini
@@ -32,6 +32,7 @@ N_average_segments_psd: 2
 coarse_grain_psd: False
 coarse_grain_csd: True
 overlap_factor: 0.5
+overlap_factor_welch: 0.5
 zeropad_csd: True
 [postprocessing]
 polarization: tensor 
