+++ test/test_data/parameters_baseline_test.ini
@@ -27,6 +27,7 @@ N_average_segments_psd: 2
 coarse_grain_psd: False
 coarse_grain_csd: True
 overlap_factor: 0.5
+overlap_factor_welch: 0.5
 zeropad_csd: True
 [postprocessing]
 polarization: tensor 

[…]
