### 2023-04-29 13c5346
linking up with bias factor calculation
