+++ pygwb_pipe/parameters.ini
@@ -44,6 +44,8 @@ cluster_window: 0.5
 [window_fft_specs]
 ; Window used for fft (used CSD and PSD estimation)
 window_fftgram: hann
+[window_fft_welch_specs]
+window_fftgram: hann
 [density_estimation]
 ; Final frequency resolution of CSD and PSD 
 frequency_resolution: 0.03125
@@ -53,7 +55,6 @@ coarse_grain_psd: False
 coarse_grain_csd: True
 overlap_factor_welch: 0.5
 overlap_factor: 0.5
-zeropad_csd: True
 [postprocessing]
 polarization: tensor 
 ; Power law index
