+++ test/spectral_test.py
@@ -68,7 +68,7 @@ class TestSpectralDensities(unittest.TestCase):
             self.segment_duration,
             frequency_resolution,
             overlap_factor=overlap_factor,
-            window_fftgram_dict_welch_psd={"window_fftgram":"hann"},
+            window_fftgram_dict={"window_fftgram":"hann"},
             overlap_factor_welch= 0.5,
         )
         N_avg_segs = 2
