+++ test/test_data/parameters_baseline_test.ini
@@ -21,6 +21,8 @@ gate_threshold: 50
 cluster_window: 0.5
 [window_fft_specs]
 window_fftgram: hann
+[window_fft_welch_specs]
+window_fftgram: hann
 [density_estimation]
 frequency_resolution: 0.03125
 N_average_segments_psd: 2
