### 2023-05-01 fcf86df
fixed window for window factor estimation to be boxcar when coarse-graining is used to estimate PSDs
