### 2023-05-02 209c15b
updated logic in bias factor for ovl=0
