### 2023-05-03 cd0b8e2
first attempt at fixing window factor calculation across the board
