### 2023-05-04 905f900
Update plot_coherence. All frequencies, log scale plot, low frequency lin scale plot.
