### 2023-05-04 bf38be9
Add function plot_hist_coherence
