### 2023-05-04 291c166
fix formatting style for pre-commits of statistical checks
