### 2023-05-05 2e4161d
update plot_all function to include new hist_coherence
