### 2023-05-05 af77f2a
update parameter names to be consistent with attributes of statistical checks module
