### 2023-05-08 cdf5698
solve some minor bugs in plot_hist_coherence
