### 2023-05-08 c08779b
revert changes to run_statistical_checks_baseline_pickle implemented for testing purposes.
