### 2023-05-08 c55f200
Update spectral.py
