### 2023-05-08 #5454 @editorialbot
[REVIEW]: `pygwb`: a Python-based library for gravitational-wave background searches
**Submitting author:** @a-renzini (<a href="http://orcid.org/0000-0002-4589-3987">Arianna Renzini</a>)
**Repository:** https://github.com/a-renzini/pygwb/
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** v1.4.0
