### 2023-05-09 a606559
modifying overlap factor calculation
