### 2023-05-09 bdbfb6b
adding a warning in baseline re. overlap factors>0.5
