### 2023-05-09 #5454 @Sbozzolo
Hi @a-renzini, I will be reviewing your package.

I wanted to clear some expectations regarding the timeline: I will soon leave for some travel, so I won't be able to work on the review for the next 3 weeks or so. I am sorry about that since I see that you've already waited for a month.

Even if the second reviewer hasn't been found yet, I would invite you to look at the [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) and make sure that your project checks all the boxes.

From my cursory first look, I can already see some areas for improvement. In particular, the documentation seems particularly bare bones. Critically, the available documentation does not tell me what the code does nor how to use it.
Here are some of the issues I've seen:
- The readme does not say what the package is about
- The readme has broken links to your CI badges
- The readme does not provide installation instructions
- The "about" page in the documentation consist of a single diagram (the meaning of which is unclear to me)
- The "installation" page points to a repository that is not GitHub
- It is unclear what version of Python is supported, but the installation instruction recommends Python 3.8 and points users to a link to a version of Anaconda that is 2 years old.
- The "pip install" method does no longer work for Debian 12+ or Ubuntu 23.04+ (see PEP 668). More distributions will probably follow soon.
- The "installation" page also claims that there are no released versions.
- The "contributing" page points users to LIGO servers
- The "contributing" page does not mention `pre-commit` or `coverage`, but mentions `pytest` (which is not listed among the dependencies).
- Why is there a pickle file in the repo? (Pickle file are not portable).
- Most tutorials do not contain any explanation, are not well formatted, and some have warnings/errors.
- Acronyms are used but not defined (e.g., SGWB, CBC).
- The documentation for the APIs is also bare bones. For example, there are no explanations for functions/arguments, or types.

My personal rule of thumb to identify good documentation is to see if I could use the code just looking at the documentation. Currently, this is not possible. I will open an issue in your repo regarding this.

I am looking forward to diving into your package!
