### 2023-05-09 a-renzini/pygwb#1 @Sbozzolo
[JOSS review] Improve documentation
(Main review for JOSS happening [here](https://github.com/openjournals/joss-reviews/issues/5454))

The current documentation is very very minimal and needs to be expanded to describe capabilities and usage.

More specifically, here are some of the issues I've seen:

- [ ] The readme does not say what the package is about
- [ ] The readme has broken links to your CI badges
- [ ] The readme does not provide installation instructions
- [ ] The "about" page in the documentation consist of a single diagram (the meaning of which is unclear to me)
- [ ] The "installation" page points to a repository that is not GitHub
- [ ] It is unclear what version of Python is supported, but the installation instruction recommends Python 3.8 and points users to a link to a version of Anaconda that is 2 years old.
- [ ] The "pip install" method does no longer work for Debian 12+ or Ubuntu 23.04+ (see PEP 668). More distributions will probably follow soon.
- [ ] The "installation" page also claims that there are no released versions.
- [ ] The "contributing" page points users to LIGO servers
- [ ] The "contributing" page does not mention pre-commit or coverage, but mentions pytest (which is not listed among the dependencies).
- [ ] Why is there a pickle file in the repo? (Pickle file are not portable).
- [ ] Most tutorials do not contain any explanation, are not well formatted, and some have warnings/errors.
- [ ] Acronyms are used but not defined (e.g., SGWB, CBC).
- [ ] The documentation for the APIs is also bare bones. For example, there are no explanations for functions/arguments, or types.

This list is not comprehensive and I would recommend designing the documentation from scratch.

Here is a good resource for a possible way to write effective documentation: https://documentation.divio.com/
