### 2023-05-10 3d34890
adding gating plot to statistical checks
