+++ pygwb/statistical_checks.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 import matplotlib
 import matplotlib.pyplot as plt
+from matplotlib.patches import Rectangle 
 
 matplotlib.rcParams['figure.figsize'] = (8,6)
 matplotlib.rcParams['axes.grid'] = True
@@ -48,6 +49,8 @@ class StatisticalChecks(object):
         plot_dir,
         baseline_name,
         param_file,
+        gates_ifo1 = None,
+        gates_ifo2 = None,
         file_tag = None,
         legend_fontsize = 16
     ):
@@ -101,6 +104,8 @@ class StatisticalChecks(object):
         self.sliding_deviate_all = (
             self.sliding_omega_all - np.nanmean(self.sliding_omega_all)
         ) / self.sliding_sigmas_all
+        self.gates_ifo1 = gates_ifo1
+        self.gates_ifo2 = gates_ifo2
 
         self.coherence_spectrum = coherence_spectrum
 
@@ -520,8 +525,6 @@ class StatisticalChecks(object):
         Generates and saves a histogram of the coherence distribution. The plot shows the data after the delta-sigma cut (bad GPS times) was applied. This function does not require any input parameters, as it accesses the data through the attributes of the class.
         Furthermore, it also saves a text file which contains the frequencies at which outliers of the coherence distribution were identified, i.e. spectral artefacts.
         """
-
-
         coherence = self.coherence_spectrum
         frequencies = self.freqs
         total_bins = 1000
@@ -534,7 +537,6 @@ class StatisticalChecks(object):
         n_segs = len(self.sliding_omega_cut) * int(np.floor(self.params.segment_duration/(fftlength))-1)
         predicted = alpha * n_frequencies * delta_coherence * n_segs * np.exp(-alpha * n_segs * coherence)
         threshold = np.log(alpha * n_segs * n_frequencies * delta_coherence) / (n_segs * alpha)
-        print(predicted,threshold)
 
         fig, axs = plt.subplots(nrows=1, ncols=1, figsize=(10, 8))
    
@@ -629,7 +631,6 @@ class StatisticalChecks(object):
         Generates and saves a plot of the cumulative sensitivity. This function does not require any input parameters, as it accesses the data through the attributes of the class (e.g. `sigma_spectrum`).
 
         """
-
         cumul_sens = integrate.cumtrapz((1 / self.sigma_spectrum ** 2), self.freqs)
         cumul_sens = cumul_sens / cumul_sens[-1]
         plt.figure(figsize=(10, 8))
@@ -1158,12 +1159,46 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-sigma_time_fit.png", bbox_inches = 'tight'
         )
 
+    def plot_gates_in_time(self):
+        if self.gates_ifo1 is None and self.gates_ifo2 is None:
+            return
+        
+        fig, ax = plt.subplots(nrows=1, ncols=1, figsize=(10, 8))
+        if self.gates_ifo1 is not None:
+            self.total_gated_time_ifo1 = np.sum(self.gates_ifo1[:,1]-self.gates_ifo1[:,0])
+            self.total_gated_percent_ifo1 = self.total_gated_time_ifo1/(int(self.params.tf)- int(self.sliding_times_all[0]))*100
+            gate_times_in_days_ifo1 = (np.array(self.gates_ifo1[:,0]) - self.sliding_times_all[0]) / 86400.0
[…]
