+++ pygwb/statistical_checks.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 import matplotlib
 import matplotlib.pyplot as plt
-from matplotlib.patches import Rectangle 
+from matplotlib.patches import Rectangle
 
 matplotlib.rcParams['figure.figsize'] = (8,6)
 matplotlib.rcParams['axes.grid'] = True
