### 2023-05-11 35ea59f
fixed a few ifs in stat checks when data is missing
