### 2023-05-11 4009123
fixed a lot of bugs in stat checks when data is missing
