+++ pygwb/statistical_checks.py
@@ -300,6 +300,8 @@ class StatisticalChecks(object):
             Maximum value on the y-axis.
 
         """
+        if not self.days_cut:
+            return
         fig = plt.figure(figsize=(10, 8))
         plt.plot(
             self.days_cut,
@@ -341,6 +343,8 @@ class StatisticalChecks(object):
         Generates and saves a plot of the running sigma. The plotted values are the ones after the delta sigma cut. This function does not require any input parameters, as it accesses the data through the attributes of the class (e.g. `days_cut`).
 
         """
+        if not self.days_cut:
+            return
         fig = plt.figure(figsize=(10, 8))
         plt.plot(
             self.days_cut, self.running_sigmas, '.', markersize=2, color=sea[0], label=self.baseline_name
@@ -384,6 +388,8 @@ class StatisticalChecks(object):
         Generates and saves a plot of the point estimate integrand. This function does not require any input parameters, as it accesses the data through the attributes of the class (e.g. `point_estimate_integrand`).
 
         """
+        if np.isnan(self.point_estimate_spectrum).all():
+            return
         plt.figure(figsize=(10, 8))
         plt.semilogy(
             self.freqs,
@@ -468,6 +474,9 @@ class StatisticalChecks(object):
         """
         Generates and saves a plot of the sigma spectrum. This function does not require any input parameters, as it accesses the data through the attributes of the class (e.g. `sigma_spectrum`).
         """
+        if np.isinf(self.sigma_spectrum).all():
+            return
+
         plt.figure(figsize=(10, 8))
         plt.plot(self.freqs, self.sigma_spectrum, color=sea[0])
         plt.xlabel("Frequency (Hz)", size=self.axes_labelsize)
@@ -486,6 +495,9 @@ class StatisticalChecks(object):
         """
         Generates and saves a plot of the coherence spectrum, if present. This function does not require any input parameters, as it accesses the data through the attributes of the class (e.g. `coherence_spectrum`).
         """
+        if not self.coherence_spectrum:
+            return
+
         flow = flow or self.flow
         fhigh = fhigh or self.fhigh
 
@@ -529,6 +541,8 @@ class StatisticalChecks(object):
         Generates and saves a histogram of the coherence distribution. The plot shows the data after the delta-sigma cut (bad GPS times) was applied. This function does not require any input parameters, as it accesses the data through the attributes of the class.
         Furthermore, it also saves a text file which contains the frequencies at which outliers of the coherence distribution were identified, i.e. spectral artefacts.
         """
+        if not self.coherence_spectrum:
+            return
         coherence = self.coherence_spectrum
         frequencies = self.freqs
         total_bins = 1000
@@ -761,8 +775,12 @@ class StatisticalChecks(object):
         axs[0].tick_params(axis="y", labelsize=self.legend_fontsize)
         axs[0].yaxis.offsetText.set_fontsize(self.legend_fontsize)
 
-        minx1 = min(self.sliding_sigma_cut)
-        maxx1 = max(self.sliding_sigma_cut)
+        if not self.sliding_sigma_cut:
+            minx1 = min(self.sliding_sigmas_all)
+            maxx1 = max(self.sliding_sigmas_all)
+        else:
[…]
