### 2023-05-11 90cc1d5
fixed more bugs in stat checks when data is missing
