### 2023-05-11 a8bef03
fixed even more bugs in stat checks when data is missing
