### 2023-05-15 45227f2
Close plots to reduce memory load
