+++ pygwb/statistical_checks.py
@@ -337,6 +337,7 @@ class StatisticalChecks(object):
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_point_estimate.png", bbox_inches='tight'
         )
+        plt.close()
 
     def plot_running_sigma(self):
         """
@@ -360,6 +361,7 @@ class StatisticalChecks(object):
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-running_sigma.png", bbox_inches = 'tight'
         )
+        plt.close()
 
     def plot_IFFT_point_estimate_integrand(self):
         """
@@ -382,6 +384,7 @@ class StatisticalChecks(object):
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-IFFT_point_estimate_integrand.png", bbox_inches='tight'
         )
+        plt.close()
 
     def plot_SNR_spectrum(self):
         """
@@ -406,6 +409,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-abs_point_estimate_integrand.png",
             bbox_inches="tight",
         )
+        plt.close()
 
     def plot_cumulative_SNR_spectrum(self):
         """
@@ -427,6 +431,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-cumulative_SNR_spectrum.png",
             bbox_inches="tight",
         )
+        plt.close()
 
     def plot_real_SNR_spectrum(self):
         """
@@ -448,6 +453,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-real_SNR_spectrum.png",
             bbox_inches="tight",
         )
+        plt.close()
 
     def plot_imag_SNR_spectrum(self):
         """
@@ -469,6 +475,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-imag_SNR_spectrum.png",
             bbox_inches="tight",
         )
+        plt.close()
 
     def plot_sigma_spectrum(self):
         """
@@ -490,6 +497,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-sigma_spectrum.png",
             bbox_inches="tight",
         )
+        plt.close()
 
     def plot_coherence_spectrum(self, flow=None, fhigh=None):
         """
@@ -519,6 +527,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-coherence_spectrum.png",
             bbox_inches="tight",
         )
+        plt.close()
 
         plt.figure(figsize=(10, 8))
         plt.plot(self.freqs, self.coherence_spectrum, color=sea[0])
@@ -534,6 +543,7 @@ class StatisticalChecks(object):
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-coherence_spectrum_zoom.png",
             bbox_inches="tight",
         )
+        plt.close()
 
 
     def plot_hist_coherence(self):
@@ -595,6 +605,7 @@ class StatisticalChecks(object):
         plt.savefig(
             f"{self.plot_dir / self.baseline_name}-{self.file_tag}-histogram_coherence.png", bbox_inches = 'tight'
         )
+        plt.close()
 
         fig, axs = plt.subplots(nrows=1, ncols=1, figsize=(10, 8))      
 
@@ -635,6 +646,7 @@ class StatisticalChecks(object):
         plt.savefig(
[…]
