### 2023-05-15 4f2e85c
Address case where there are no gates
