### 2023-05-15 4d56eb0
fixing more little bugs due to data irregularities
