### 2023-05-19 #5454 @cmbiwer
I added some comments to the issue @Sbozzolo started here: https://github.com/a-renzini/pygwb/issues/1

I reviewed it without looking at @Sbozzolo review. And ended up with many of the same issues. I worked around some of the issues with the installation to try the test suite. I added a couple more things to that already existing issue on pygwb.
