### 2023-05-19 a-renzini/pygwb#1 @cmbiwer
- [ ] State of the field in the JOSS paper submission.

One more with the paper itself. The state of the field is a requirement in the paper. But there wasn't really any mention of the state of the field (eg. LALStochastic) and other packages, and what separates pygwb from those packages.
