### 2023-05-19 fdf0c35
fixing window factors in calc_Y_sigma_spectra
