+++ .github/workflows/documentation.yaml
@@ -14,7 +14,7 @@ jobs:
           pip install sphinx numpydoc nbsphinx sphinx_rtd_theme sphinx-tabs autodoc ipython-genutils
       - name: Sphinx build
         run: |
-          sphinx-build doc _build
+          sphinx-build docs _build
           
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
