+++ .github/workflows/documentation.yaml
@@ -10,8 +10,6 @@ jobs:
       - uses: actions/setup-python@v3
       - name: Install dependencies
         run: |
-          apt-get -yqq update
-          apt-get -yqq install pandoc
           pip install --upgrade pip setuptools
           pip install ipykernel ipython==8.12.0 jupyter
           pip install sphinx numpydoc nbsphinx sphinx_rtd_theme sphinx-tabs autodoc ipython-genutils
