+++ .github/workflows/documentation.yaml
@@ -8,6 +8,8 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v3
+        with:
+          python-version: 3.8
       - name: Install dependencies
         run: |
           pip install --upgrade pip setuptools
