+++ .github/workflows/documentation.yaml
@@ -18,13 +18,12 @@ jobs:
           pip install .
       - name: Sphinx build
         run: |
-          sphinx-build docs _build
-          
+          sphinx-build docs docs/_build
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
         with:
           publish_branch: gh-pages
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: _build/
+          publish_dir: docs/_build/
           force_orphan: true
