+++ .github/workflows/documentation.yaml
@@ -16,12 +16,15 @@ jobs:
           pip install ipykernel ipython==8.12.0 jupyter
           pip install sphinx numpydoc nbsphinx sphinx_rtd_theme sphinx-tabs autodoc ipython-genutils
           pip install .
-      - name: Sphinx build
+      - name: Setup tutorials
         run: |
+          ls
           cp ../tutorials/*.ipynb ./
           cp ../tutorials/*.dat ./
           cp -r ../tutorials/input ./
           rm integrationTest*.ipynb
+      - name: Sphinx build
+        run: |
           sphinx-build docs docs/_build
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
