+++ .github/workflows/documentation.yaml
@@ -14,6 +14,7 @@ jobs:
         run: |
           pip install --upgrade pip setuptools
           pip install ipykernel ipython==8.12.0 jupyter
+          pip install pandoc
           pip install sphinx numpydoc nbsphinx sphinx_rtd_theme sphinx-tabs autodoc ipython-genutils
           pip install .
       - name: Setup tutorials
