+++ .github/workflows/documentation.yaml
@@ -27,7 +27,10 @@ jobs:
           ls docs
       - name: Sphinx build
         run: |
-          sphinx-build docs docs/_build
+          cd docs
+          make clean
+          make html
+          cd -
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
