+++ .github/workflows/documentation.yaml
@@ -13,8 +13,8 @@ jobs:
       - name: Install dependencies
         run: |
           pip install --upgrade pip setuptools
-          pip install ipykernel ipython==8.12.0 jupyter
           pip install pandoc
+          pip install ipykernel ipython==8.12.0 jupyter
           pip install sphinx numpydoc nbsphinx sphinx_rtd_theme sphinx-tabs autodoc ipython-genutils
           pip install .
       - name: Setup tutorials
