+++ .github/workflows/documentation.yaml
@@ -12,7 +12,6 @@ jobs:
           python-version: 3.8
       - name: Install dependencies
         run: |
-          sudo apt-get -yqq udpdate
           sudo apt-get -yqq install pandoc
           pip install --upgrade pip setuptools
           pip install ipykernel ipython==8.12.0 jupyter
