+++ .github/workflows/documentation.yaml
@@ -37,5 +37,5 @@ jobs:
         with:
           publish_branch: gh-pages
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: docs/_build/
+          publish_dir: _build/
           force_orphan: true
