+++ .github/workflows/documentation.yaml
@@ -46,8 +46,8 @@ jobs:
           force_orphan: true
   deploy:
     environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
+    name: github-pages
+    url: ${{ steps.deployment.outputs.page_url }}
     runs-on: ubuntu-latest
     needs: build
     steps:
