+++ .github/workflows/documentation.yaml
@@ -37,22 +37,10 @@ jobs:
           path: docs/_build/
       - name: Deploy
         uses: peaceiris/actions-gh-pages@v3
-        if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
+        if: ${{ github.event_name == 'push' }}
         with:
           publish_branch: joss_submission
           keep_files: true
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: docs/_build/
           force_orphan: true
-  deploy:
-    environment:
-    name: github-pages
-    url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v2
-        with:
-          path: docs/_build/
