+++ .github/workflows/documentation.yaml
@@ -39,7 +39,7 @@ jobs:
         uses: peaceiris/actions-gh-pages@v3
         if: ${{ github.event_name == 'push' }}
         with:
-          publish_branch: joss_submission
+          publish_branch: gh-pages
           keep_files: true
           github_token: ${{ secrets.GITHUB_TOKEN }}
           publish_dir: docs/_build/
