+++ .github/workflows/documentation.yaml
@@ -42,5 +42,5 @@ jobs:
           publish_branch: gh-pages
           keep_files: true
           github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: docs/_build/
+          publish_dir: docs/_build/html/
           force_orphan: true
