### 2023-05-23 a-renzini/pygwb#1 @a-renzini
@Sbozzolo and @cmbiwer : thanks for your feedback! Several of the issues you point out are easy to resolve and we will update the documentation accordingly; there are a couple though that I am unsure how to address so let me post questions here.
