### 2023-05-23 a-renzini/pygwb#1 @a-renzini
> I went through before looking at @Sbozzolo review. I wrote down basically the same issues and agree with @Sbozzolo . There is a lot missing from the documentation/tutorial and things currently point to URLs that require LIGO credentials.
>
> I'll add a few more things though.
>
> * [ ]  Test suite seems to fail.
>
> I installed pytest myself (see the comments above, its not there) and tried the test suite. I see two test fail; `TestInterferometer.test_set_timeseries_from_channel_name` and `TestInterferometer.test_from_parameters`. The message (truncated):
>
> ```
> _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
>
> fobj = '/var/folders/7b/9dh6yr6n6d9dvwlc33g1_3rw0000gp/T/astropy-download-92331-u335cim9', args = (), kwargs = {'end': LIGOTimeGPS(1247644588, 0), 'start': LIGOTimeGPS(1247644169, 0), 'tag': 'C00'}
> h5f = <Closed HDF5 file>
>
>     @wraps(func)
>     def decorated_func(fobj, *args, **kwargs):
>         # pylint: disable=missing-docstring
>         if not isinstance(fobj, h5py.HLObject):
>             if isinstance(fobj, FILE_LIKE):
>                 fobj = fobj.name
>             with h5py.File(fobj, 'r') as h5f:
> >               return func(h5f, *args, **kwargs)
> E               TypeError: Failed to read GWOSC data from 'https://gwosc.org/archive/data/O3a_16KHZ_R1/1246756864/H-H1_GWOSC_O3a_16KHZ_R1-1247641600-4096.hdf5': read_gwosc_hdf5() got an unexpected keyword argument 'tag'
>
> ../../../opt/anaconda3/envs/pygwb/lib/python3.8/site-packages/gwpy/io/hdf5.py:59: TypeError
> ```
>
> * [ ]  Installation instructions seem incomplete.
>
> Aside from the issues already mentioned. It says to do:
>
> ```
> conda install .
> ```
>
> From the dir above the repo.
>
> That didn't work either there. And even `cd` into in the repo, it didn't work. Conda couldn't find the package; eg.:
>
> ```
> $ conda install .
> Collecting package metadata (current_repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
> Collecting package metadata (repodata.json): done
> Solving environment: failed with initial frozen solve. Retrying with flexible solve.
>
> PackagesNotFoundError: The following packages are not available from current channels:
>
>   - .
>
> Current channels:
>
>   - https://conda.anaconda.org/conda-forge/osx-64
>   - https://conda.anaconda.org/conda-forge/noarch
>   - https://repo.anaconda.com/pkgs/main/osx-64
>   - https://repo.anaconda.com/pkgs/main/noarch
>   - https://repo.anaconda.com/pkgs/r/osx-64
>   - https://repo.anaconda.com/pkgs/r/noarch
>
> To search for alternate channels that may provide the conda package you're
> looking for, navigate to
>
>     https://anaconda.org
> ```
>
> I ended up using `python -m pip install .` to try using the package.
>
> It may just be my Anaconda version. But its not stated anywhere what version is required.

The failure of the test suite here seems to be due to a gwpy issue which (at least to my knowledge) was solved a while ago. Could you upload the full logs so I know what version you are running, etc?

Also, we run the testing suite in the pygwb gitlab repo, so that is why the CI badges are broken in the github version. Should I duplicate the testing also in github? It seems like a bit of a waste of computational resources. The gitlab repo is now public so you can see the test coverage there.
