+++ .github/workflows/documentation.yaml
@@ -20,9 +20,10 @@ jobs:
       - name: Setup tutorials
         run: |
           ls
-          cp tutorials/make_notchlist.ipynb ./docs
+          cp tutorials/*.ipynb ./docs
           cp tutorials/*.dat ./docs
           cp -r tutorials/input ./docs
+          rm docs/integrationTest*.ipynb
           ls docs
       - name: Sphinx build
         run: |
