+++ docs/about.rst
@@ -5,7 +5,8 @@ About ``pygwb``
 A gravitational-wave background (GWB) is expected from the superposition of all gravitational waves (GWs) too faint to be detected individually, or by the incoherent overlap of a large number of signals in the same band `(Renzini+ 2022) <https://www.mdpi.com/2075-4434/10/1/34>`_. A GWB is primarily characterized by its spectral emission, usually parameterized by the GW fractional energy density spectrum, which is the target for stochastic GW searches `(Allen & Romano 1999) <https://journals.aps.org/prd/abstract/10.1103/PhysRevD.59.102001>`_,
 
 .. math:: 
-\Omega_{\rm GW}(f) = \frac{1}{\rho_c}\frac{d\rho_{\rm GW}(f)}{d\ln f},
+
+   \Omega_{\rm GW}(f) = \frac{1}{\rho_c}\frac{d\rho_{\rm GW}(f)}{d\ln f},
 
 
 where :math:`d\rho_{\rm GW}` is the energy density of GWs in the frequency band :math:`f` to :math:`df` , and :math:`\rho_c` is the critical energy density of the Universe. Different categories of GW sources may be identified by the unique spectral shape of their background emission; hence, the detection of a GWB will provide invaluable information about the evolution of the Universe and the population of GW sources within it.
