### 2023-05-23 a-renzini/pygwb#1 @a-renzini
- [ ] The "about" page in the documentation consist of a single diagram (the meaning of which is unclear to me)

Now addressed here: 7ea65ea8621ae952e368bd8b2dd81e88abeefca6
