### 2023-05-23 213fa0c
fixing issue with broken gwosc `tag` argument
