+++ pygwb/parameters.py
@@ -88,8 +88,6 @@ class Parameters:
         Cluster window. Default is 0.5.
     gate_whiten: bool
         Whether to whiten when gating. Default is True.
-    tag: str
-        Hint for the read_data function to retrieve one specific type of data, e.g.: C00, C01
     return_naive_and_averaged_sigmas: bool
         option to return naive and sliding sigmas from delta sigma cut
     window_fftgram_dict: dictionary, optional
@@ -132,7 +130,6 @@ class Parameters:
     gate_threshold: float = 50.0
     cluster_window: float = 0.5
     gate_whiten: bool = True
-    tag: str = "C00"
     return_naive_and_averaged_sigmas: bool = False
 
 
@@ -388,7 +385,6 @@ class ParametersHelp(enum.Enum):
     gate_threshold = "Gate threshold. Default is 50."
     cluster_window = "Cluster window. Default is 0.5."
     gate_whiten = "Whether to whiten when gating. Default is True."
-    tag = "Hint for the read_data function to retrieve one specific type of data, e.g.: C00, C01"
     return_naive_and_averaged_sigmas = "option to return naive and sliding sigmas from delta sigma cut. Default value: False"
 
     @property
