+++ setup.py
@@ -39,7 +39,7 @@ setup(
         "matplotlib",
         "scipy>=1.8.0",
         "bilby>=1.4",
-        "gwpy>=3.0.1",
+        "gwpy>=3.0.2",
         "astropy>=5.2",
         "gwdetchar",
         "gwsumm",
