### 2023-05-23 a-renzini/pygwb#1 @a-renzini
@cmbiwer re:
- [ ]  Test suite seems to fail.

I tracked this issue back and it is due to a backwards-incompatible change in the `gwosc` package; this was fixed in pygwb v.1.1, but the version you see here is the one associated to the ApJ paper, and thus does not have this change. I have gone ahead and updated (and cleaned up) the code here, removing the `tag` argument which was deprecated by `gwosc`. We will merge these changes into `pygwb` once the review is over. The documentation should now run.

See 213fa0cd7bbba3b2acc682977cd043209ca1a01d

(I could also have fixed the version of gwosc to an older one, to make the old v1.0.0 work; however, as I expect the changes we make here will eventually be merged with the current version of the code, I thought it would be more useful to update the code this way. I am also happy to go the other way.)
