### 2023-05-23 1597224
Adding ability to load gates from file and apply them.
