### 2023-05-23 a-renzini/pygwb#1 @cmbiwer
> The failure of the test suite here seems to be due to a gwpy issue which (at least to my knowledge) was solved a while ago. Could you upload the full logs so I know what version you are running, etc?

This is the version it installed for me:
```
(pygwb) pn1618376:pygwb cmbiwer$ python
Python 3.8.16 | packaged by conda-forge | (default, Feb  1 2023, 16:05:36)
[Clang 14.0.6 ] on darwin
>>> import pygwb
>>> pygwb.__version__
'1.0.1.dev4'
```
