### 2023-05-23 a-renzini/pygwb#1 @a-renzini
> > The failure of the test suite here seems to be due to a gwpy issue which (at least to my knowledge) was solved a while ago. Could you upload the full logs so I know what version you are running, etc?
>
> This is the version it installed for me:
>
> ```
> (pygwb) pn1618376:pygwb cmbiwer$ python
> Python 3.8.16 | packaged by conda-forge | (default, Feb  1 2023, 16:05:36)
> [Clang 14.0.6 ] on darwin
> Type "help", "copyright", "credits" or "license" for more information.
> >>> import pygwb
> >>> pygwb.__version__
> '1.0.1.dev4'
> ```

sorry - I meant `gwpy` version; in any case, I addressed this issue now in 213fa0cd7bbba3b2acc682977cd043209ca1a01d as mentioned above - please let me know whether the test suite now works with these changes :)

(note that the issue was coming from the fact that `gwpy` depends on `gwosc`, and although the version of `gwpy` installed by `pygwb v1.0.0` is correct, `gwpy` doesn't fix the version of `gwosc` it needs and the break was coming from updates in `gwosc`)
