+++ pygwb/detector.py
@@ -399,7 +399,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
         )
         self.gates = self.gates | new_gates
         self.gate_pad = gate_tpad
-        
+
     def load_gates(self, loaded_object, index, **kwargs):
         """
         Load gates from a pygwb output file and apply them to the Interferometer object. 
@@ -412,16 +412,16 @@ class Interferometer(bilby.gw.detector.Interferometer):
         index ; int
             Integer representing the correct ifo object in the baseline
         """
-        
+
         gates = loaded_object[f"ifo_{index}_gates"]
         gate_tpad = kwargs.pop("gate_tpad")
-        
+
         self.timeseries, new_gates = self_gate_data(
             self.timeseries,
             tpad=gate_tpad,
             gates=gates,
         )
-        
+
         self.gates = self.gates | new_gates
         self.gate_pad = None
 
