### 2023-05-24 5a1c6fe
changing coherence calculation: using now naive PSDs
