### 2023-05-24 e32895f
Update statistical_checks.py
