### 2023-05-25 102f53c
adding some annotations to the statchecks
