### 2023-05-25 feef05b
fixed possible ifo parameters to increse robustness
