### 2023-05-25 36e6550
added run locally option to isotropic workflow
