+++ pygwb/detector.py
@@ -400,7 +400,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
         self.gates = self.gates | new_gates
         self.gate_pad = gate_tpad
 
-    def load_gates(self, loaded_object, index, **kwargs):
+    def apply_gates_from_file(self, loaded_object, index, **kwargs):
         """
         Load gates from a pygwb output file and apply them to the Interferometer object. 
         The gated times are stored as a property of the object.
@@ -411,6 +411,9 @@ class Interferometer(bilby.gw.detector.Interferometer):
             Object that represents the data in the output file, e.g. a loaded npzobject.
         index ; int
             Integer representing the correct ifo object in the baseline
+        gate_tpad : float
+            half-width time duration (seconds) in which the Planck window
+            is tapered
         """
         gates = loaded_object[f"ifo_{index}_gates"]
         gate_tpad = kwargs.pop("gate_tpad")
