+++ pygwb_pipe/parameters.ini
@@ -29,6 +29,8 @@ window_downsampling: hamming
 ftype: fir
 [gating]
 ; gating parameters
+; Path to file containing gates to load
+path_gate_data: 
 ; Running self-gating or not
 gate_data: False
 ; Whitening the data before gating
