### 2023-05-26 428eafe
Adding some stuff for testing, but not finished.
