+++ test/detector_test.py
@@ -137,6 +137,26 @@ class TestInterferometer(unittest.TestCase):
         self.assertTrue((_known_gate in ifo.gates), True)
         self.assertTrue(abs(ifo.gates), 2*gate_tzero)
         self.assertTrue(ifo.gate_pad, gate_tpad)
+        
+    def test_gated_times_from_file(self):
+        gwpy_timeseries = self.testdata["original_timeseries"]
+        ifo = detector.Interferometer.get_empty_interferometer(self.ifo)
+        ifo.set_timeseries_from_gwpy_timeseries(
+            gwpy_timeseries=gwpy_timeseries, **self.kwargs
+        )
+        gate_tzero = 1.0
+        gate_tpad = 0.5
+        gate_threshold = 50.0
+        cluster_window = 0.5
+        gate_whiten = True
+        ifo.gate_data_apply(
+            gate_tzero = gate_tzero, gate_tpad = gate_tpad,
+            gate_threshold = gate_threshold, cluster_window = cluster_window,
+            gate_whiten = gate_whiten,
+        )
+        
+        
+        
 
 if __name__ == "__main__":
     unittest.main()
