+++ test/test_data/parameters_detector_test.ini
@@ -18,6 +18,7 @@ number_cropped_seconds: 2
 window_downsampling: hamming
 ftype: fir
 [gating]
+path_gate_data: /home/max.lalleman/public_html/Code/New_Clone_pyGWB/pygwb/test/test_data
 gate_data: False
 gate_whiten: True
 gate_tzero: 1
