### 2023-05-28 a-renzini/pygwb#1 @Sbozzolo
> Thank you! We are hard at work addressing this issue. Could this be turned into a checklist by any chance?

Sorry for the delay. I changed the list into a checklist. In this, let me remark again that this list summarizes my first impressions, so I'd encourage you to also think about the documentation beyond these points.

> @Sbozzolo the reason there are pickle files as a part of the repository is because some of our tests depend on them; they are not part of the codebase itself (when packaging the code for distribution the test suite is omitted). Our objects are designed as data containers, at different stages of a GWB analysis, and the test suite is based on pre-compiled pickled versions of these objects. We have found this an effective way of keeping track of changes in the modules and objects, without having the test suite re-run the analysis every time.

I was under the impression that pickle files were not portable across versions/OSs, but it turns out that they are, as long as the same protocol is being used. Nonetheless, I think that sharing pickles is not the best idea. As the [documentation](https://docs.python.org/3/library/pickle.html#module-pickle) clearly remarks, pickles are not safe. Moreover, they are opaque and can only be read by Python. If it is possible and not too difficult, I'd suggest moving to something like JSON. Also, if used only for tests, the file should not be at the root level of your repository.

> Now addressed here: https://github.com/a-renzini/pygwb/commit/07926e35901c09130752054cab414bc18a352a0c

What branch should I look at when reviewing the code/documentation?
