### 2023-05-28 #5454 @a-renzini
> @a-renzini
>
> The existence of two nearly identical public repositories (https://github.com/a-renzini/pygwb/, https://git.ligo.org/pygwb/pygwb) is a little bit confusing. Which one is the "real" one?
>
> If you want to continue the development on GitLab but also want to be on GitHub, maybe you can consider mirroring the repository instead of creating a separate one?

Hello @Sbozzolo - that is exactly what I've done; the GitHub is a mirror of the GitLab! We are currently developing on GitLab.
