### 2023-05-28 #5454 @Sbozzolo
> > @a-renzini
> > The existence of two nearly identical public repositories (https://github.com/a-renzini/pygwb/, https://git.ligo.org/pygwb/pygwb) is a little bit confusing. Which one is the "real" one?
> > If you want to continue the development on GitLab but also want to be on GitHub, maybe you can consider mirroring the repository instead of creating a separate one?
>
> Hello @Sbozzolo - that is exactly what I've done; the GitHub is a mirror of the GitLab! We are currently developing on GitLab.

It doesn't look like that the mirroring is working properly: the latest commit on GitLab is 36e6550eb5622d4456861b46c10c047163d394a5, but this commit does not exist on GitHub.

Also, if the repo on GitHub is just a mirror, I'd recommend pointing users to the GitLab repo in the readme (and maybe renaming the repo to pygwb-mirror, see, e.g., https://github.com/emacs-mirror/emacs). Also, your JOSS submission says that the link to the repo is the GitHub one, so maybe we should amend that too.
