### 2023-05-28 a-renzini/pygwb#1 @Sbozzolo
Addendum: I realized that there are more pickle files in the tests folder. I had "checkpoint.pkl" in mind when writing the message.
