### 2023-05-28 a-renzini/pygwb#1 @a-renzini
> Addendum: I realized that there are more pickle files in the tests folder. I had "checkpoint.pkl" in mind when writing the message.

oops - that file is NOT meant to be there. All the pickles used for checks should be in `test/test_data`! We must have forgotten to gitignore it...
