### 2023-05-28 #5454 @a-renzini
@Sbozzolo I realise this wasn't obvious: the new documentation is produced as a pages job in github, and lives here:

https://a-renzini.github.io/pygwb/

once we are happy with it, I will merge it into gitlab and it will be produced as a gitlab pages job.
