### 2023-05-28 f8727af
fixing the data segment finding -- applied same fix as in stochmon
