### 2023-05-30 f2cce5f
Adding tests for applying gates to file.
