### 2023-05-30 a4d526d
Changing loading path of file with gates.
