### 2023-05-30 4c12e3c
adding notches to stat checks
