+++ pygwb/statistical_checks.py
@@ -1,4 +1,5 @@
 import json
+from loguru import logger
 import warnings
 from os import listdir
 from os.path import isfile, join
@@ -6,7 +7,7 @@ from pathlib import Path
 
 import matplotlib
 import matplotlib.pyplot as plt
-from matplotlib.patches import Rectangle
+import matplotlib.transforms as mt
 
 matplotlib.rcParams['figure.figsize'] = (8,6)
 matplotlib.rcParams['axes.grid'] = True
@@ -30,6 +31,7 @@ from scipy import integrate, stats
 from scipy.optimize import curve_fit
 
 from pygwb.baseline import Baseline
+from pygwb.notch import StochNotchList
 from pygwb.parameters import Parameters
 from pygwb.util import StatKS, calc_bias
 
@@ -45,12 +47,13 @@ class StatisticalChecks(object):
         coherence_n_segs,
         point_estimate_spectrum,
         sigma_spectrum,
-        freqs,
+        frequencies,
         badGPSTimes,
         delta_sigmas,
         plot_dir,
         baseline_name,
         param_file,
+        frequency_mask = None,
         gates_ifo1 = None,
         gates_ifo2 = None,
         file_tag = None,
@@ -70,14 +73,14 @@ class StatisticalChecks(object):
         naive_sigmas_all: array
             Array of naive sigmas before the bad GPS times from the delta sigma cut are applied.
         coherence_spectrum: array
-            Array containing a coherence spectrum. Each entry in this array corresponds to the 2-detector coherence spectrum evaluated at the corresponding frequency in the freqs array.
+            Array containing a coherence spectrum. Each entry in this array corresponds to the 2-detector coherence spectrum evaluated at the corresponding frequency in the frequencies array.
         coherence_n_segs: int
             Number of segments used for coherence calculation.
         point_estimate_spectrum: array
-            Array containing the point estimate spectrum. Each entry in this array corresponds to the point estimate spectrum evaluated at the corresponding frequency in the freqs array.
+            Array containing the point estimate spectrum. Each entry in this array corresponds to the point estimate spectrum evaluated at the corresponding frequency in the frequencies array.
         sigma_spectrum: array
-            Array containing the sigma spectrum. Each entry in this array corresponds to the sigma spectrum evaluated at the corresponding frequency in the freqs array.
-        freqs: array
+            Array containing the sigma spectrum. Each entry in this array corresponds to the sigma spectrum evaluated at the corresponding frequency in the frequencies array.
+        frequencies: array
             Array containing the frequencies.
         badGPStimes: array
             Array of bad GPS times, i.e. times that do not pass the delta sigma cut.
@@ -89,6 +92,14 @@ class StatisticalChecks(object):
             Name of the baseline under consideration.
         param_file: str
             String with path to the file containing the parameters that were used for the analysis run.
+        frequency_mask: array
+            Boolean mask applied to the specrtra in broad-band analyses. 
+        gates_ifo1/gates_ifo2: list
+            List of gates applied to interferometer 1/2.
+        file_tag: str
+            Tag to be used in file naming convention.
+        legend_fontsize: int
[…]
