+++ test/detector_test.py
@@ -2,6 +2,7 @@ import pickle
 import unittest
 
 import numpy as np
+from gwpy.segments import SegmentList, Segment
 
 from pygwb import detector, parameters
 
@@ -139,23 +140,6 @@ class TestInterferometer(unittest.TestCase):
         self.assertTrue(ifo.gate_pad, gate_tpad)
         
     def test_gated_times_from_file(self):
-        gwpy_timeseries = self.testdata["original_timeseries"]
-        ifo = detector.Interferometer.get_empty_interferometer(self.ifo)
-        ifo.set_timeseries_from_gwpy_timeseries(
-            gwpy_timeseries=gwpy_timeseries, **self.kwargs
-        )
-        gate_tzero = 1.0
-        gate_tpad = 0.5
-        gate_threshold = 50.0
-        cluster_window = 0.5
-        gate_whiten = True
-        ifo.gate_data_apply(
-            gate_tzero = gate_tzero, gate_tpad = gate_tpad,
-            gate_threshold = gate_threshold, cluster_window = cluster_window,
-            gate_whiten = gate_whiten,
-        )
-        gates_from_function = ifo.gates
-        print(gates_from_function)
         npzobject = np.load("./test/test_data/point_estimate_sigma_1247644138-1247645038.npz")
         ifo_for_loading = detector.Interferometer.get_empty_interferometer(self.ifo)
         ifo_for_loading.set_timeseries_from_gwpy_timeseries(
@@ -168,9 +152,9 @@ class TestInterferometer(unittest.TestCase):
             gate_tpad = gate_tpad
         )
         gates_applied_from_file = ifo_for_loading.gates
-        print(gates_applied_from_file)
+        gates_we_know = SegmentList(Segment(1247644445.8190918, 1247644447.8190918))
         for index, gates in enumerate(gates_applied_from_file):
-            self.assertEqual(gates, gates_from_function[index])
+            self.assertEqual(gates, gates_we_know[index])
         
 
 if __name__ == "__main__":
