### 2023-05-31 #5454 @Sbozzolo
> I'll look into pointing/renaming as suggested. This does bring up a more top-level question though: during initial submission (which was a while ago now) we fixed the version for review to v1.0.0, which is the one associated to the ApJ paper. Shall we review this version for the JOSS paper, or shall we review the most up-to-date version? This will make a difference to the documentation. We don't have strong feelings either way; I don't know what the standard for JOSS papers is (whether the paper is linked to a specific version of the code, or whether it is meant to be somewhat general).

This is probably a question for @plaplant, but I'd personally say that it's a better use of your time to upgrade the most recent version as opposed to an outdated one.

> @Sbozzolo I realise this wasn't obvious: the new documentation is produced as a pages job in github, and lives here:
>
> https://a-renzini.github.io/pygwb/
>
> once we are happy with it, I will merge it into gitlab and it will be produced as a gitlab pages job.

Thank you!

I'd also recommend changing the repository in your JOSS paper to your GitLab one.
