### 2023-06-02 faaf8f8
Fixing test for frequency saving change
