+++ pygwb/detector.py
@@ -17,7 +17,25 @@ from .spectral import before_after_average, power_spectral_density
 
 class Interferometer(bilby.gw.detector.Interferometer):
 
-    """Subclass of bilby's Interferometer class"""
+    """
+    Subclass of bilby's Interferometer class which is charged with handling, storing and saving all relevant interferometer data.
+    It handles all data analysis parts relating to the individual detectors in a baseline of a network.
+    An example would be loading in data from a certain channel and computing the psd of the detector.
+    
+    Examples
+    --------
+    
+    >>> from pygwb.detector import Interferometer
+
+    To load in some data, first make an empty detector object.
+    
+    >>> ifo_1 : Interferometer.get_empty_interferometer("H1")
+    
+    Then, load in data using one of the provided `set` functions.
+    
+    >>> 
+    
+    """
 
     def __init__(self, *args, **kwargs):
         """Instantiate an Interferometer class
@@ -28,46 +46,46 @@ class Interferometer(bilby.gw.detector.Interferometer):
         **kwargs : keyword arguments passed to the (parent) bilby's Inteferometer class.
 
         Nominally, the bilby's Interferometer class takes the following arguments.
-        name: str
+        name: `str`
             Interferometer name, e.g. H1.
-        power_spectral_density: bilby.gw.detector.PowerSpectralDensity
+        power_spectral_density: `bilby.gw.detector.PowerSpectralDensity`
             Power spectral density determining the sensitivity of the detector.
-        minimum_frequency: float
+        minimum_frequency: `float`
             Minimum frequency to analyse for detector.
-        maximum_frequency: float
+        maximum_frequency: `float`
             Maximum frequency to analyse for detector.
-        length: float
+        length: `float`
             Length of the interferometer in km.
-        latitude: float
+        latitude: `float`
             Latitude North in degrees (South is negative).
-        longitude: float
+        longitude: `float`
             Longitude East in degrees (West is negative).
-        elevation: float
+        elevation: `float`
             Height above surface in metres.
-        xarm_azimuth: float
+        xarm_azimuth: `float`
             Orientation of the x arm in degrees North of East.
-        yarm_azimuth: float
+        yarm_azimuth: `float`
             Orientation of the y arm in degrees North of East.
-        xarm_tilt: float, optional
+        xarm_tilt: `float`, optional
             Tilt of the x arm in radians above the horizontal defined by
             ellipsoid earth model in LIGO-T980044-08.
-        yarm_tilt: float, optional
+        yarm_tilt: `float`, optional
             Tilt of the y arm in radians above the horizontal.
         calibration_model: Recalibration
             Calibration model, this applies the calibration correction to the
             template, the default model applies no correction.
 
-        See https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.interferometer.Interferometer.html#bilby.gw.detector.interferometer.Interferometer
+        See `docs of bilby <https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.interferometer.Interferometer.html#bilby.gw.detector.interferometer.Interferometer>`_
         for the detailed docs of the parent class.
 
         Additional attributes
-        timeseries : gwpy timeseries
+        timeseries : `gwpy.timeseries.TimeSeries`
             timeseries object with resampling/high-pass filter applied.
-        psd_spectrogram : gwpy spectrogram
+        psd_spectrogram : `gwpy.spectrogram.Spectrogram`
             gwpy spectrogram of power spectral density
-        gates: gwpy SegmentList
+        gates: `gwpy.segments.SegmentList`
             List of segments that have been gated, not including any additional padding.
-        gate_pad: float
+        gate_pad: `float`
[…]
