### 2023-06-02 0307bf1
Adding more info to doc strings delta sigma cut
