+++ pygwb/delta_sigma_cut.py
@@ -5,6 +5,14 @@ from pygwb.postprocessing import calculate_point_estimate_sigma_spectra
 from pygwb.util import calc_bias
 
 
+"""
+In general, the noise level in ground-based detectors changes slowly on time-scales of tens of minutes to hours. The
+variance associated to each segment is an indicator of that level of noise, which typically changes
+at roughly the percent level from one data segment to the next. However, there are occasional very loud disturbances
+to the detectors, such as glitches, which violate the Gaussianity of the noise. To avoid biases due to these noise events, 
+the delta-sigma cut module is implemented in `pygwb` which flags specific segments to be cut from the analyzed set.
+"""
+
 def dsc_cut(
     naive_sigma: np.ndarray,
     slide_sigma: np.ndarray,
@@ -13,34 +21,36 @@ def dsc_cut(
     bf_ns: float = 1,
 ):
     """
-    Function that performs the delta sigma cut, a veto that marks certain GPS times as unusable if the estimations of
+    Function that performs the delta sigma cut, a veto that marks certain GPS times as unusable if the estimation of
     the PSD in the naive (estimating sigma in bin J) and sliding (estimating sigma in bins J \pm 1) differ by more than
-    a certain factor (default: dsc=0.2)
+    a certain threshold:
+    .. math:: 
+        \frac{|\bar{\sigma}_{t, \alpha} b_{\rm avg} - \sigma_{t, \alpha} b_{\rm nav} |} {\bar{\sigma}_{t, \alpha} b_{\rm avg}}>{\rm threshold}
 
     Parameters
     ==========
-    naive_sigma: array
+    naive_sigma: array_like
         Naive sigma
 
-    slide_sigma: array
+    slide_sigma: array_like
         Sliding sigma
 
-    dsc: float
-        Threshold to perform the delta sigma cut
+    dsc: float, optional
+        Threshold used for the delta sigma cut. Default is 0.2.
 
-    bf_ss: float
-        Sliding bias factor
+    bf_ss: float, optional
+        Bias factor for sliding sigmas. Default is 1.
 
-    bf_ns: float
-        Naive bias factor
+    bf_ns: float, optional
+        Bias factor for naive sigmas. Default is 1.
 
     Returns
     =======
     dsigma >= dsc: bool
-        True: the segment's delta sigma exceeds the threshold value, thus making its corresponding GPStime BAD
-        False:  the segment's delta sigma is less than the threshold value, thus making its corresponding GPStime GOOD
-    dsigma: np.array
-        values of the difference between sliding sigma and naive sigma, i.e.: real value of the delta sigma cut per segment
+        Mask containing bools indicating whether the segment's delta sigma exceeds the threshold value or not. True indicates that the 
+        corresponding GPS times were bad, whereas False denotes good GPS times.
+    dsigma: array_like
+        Values of the difference between sliding sigma and naive sigma, i.e., the actual value of the delta sigma per segment.
     """
 
     dsigma = np.abs(slide_sigma * bf_ss - naive_sigma * bf_ns) / (slide_sigma * bf_ss)
@@ -62,7 +72,7 @@ def run_dsc(
     frequency_mask: np.array = True,
     window_fftgram_dict: dict = {"window_fftgram": "hann"},
     overlap_factor: float=0.5,
-    N_average_segments_welch_psd = 2,
+    N_average_segments_welch_psd: int = 2,
     return_naive_and_averaged_sigmas: bool = False,
 ):
 
@@ -72,44 +82,57 @@ def run_dsc(
     Parameters
     ==========
     dsc: float
-        The value of the delta sigma cut to use
+        Threshold used for the delta sigma cut.
 
     segment_duration: int
-        Duration of each segment
+        Duration of each segment.
 
-    psd1_naive; psd2_naive: np.array
[…]
