### 2023-06-02 7dbbda0
Remove dsc tutorial from docs page
