+++ pygwb/pe.py
@@ -53,7 +53,7 @@ class GWBModel(bilby.Likelihood):
 
     @abstractmethod
     def parameters(self):
-        """Parameters. Should return a dictionary"""
+        """Parameters to be inferred from the data. Should return a dictionary"""
         pass
 
     @abstractmethod
@@ -127,15 +127,45 @@ class PowerLawModel(GWBModel):
     .. math::
         \Omega(f) = \Omega_{\text{ref}} \left(\frac{f}{f_{\text{ref}}}\right)^{\alpha}
     Parameters:
-    -----------
+    ===========
     fref : float
-        reference frequency for defining the model (:math:`f_{\text{ref}}`)
+        Reference frequency for defining the model (:math:`f_{\text{ref}}`)
     omega_ref : float
-        amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
+        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
     alpha : float
-        spectral index of the power law (:math:`\alpha`)
+        Spectral index of the power law (:math:`\alpha`)
     frequencies : numpy.ndarray
-        array of frequencies at which to evaluate the model
+        Array of frequencies at which to evaluate the model
+        
+    Example
+    -------
+    
+    To illustrate how to construct a likelihood with this model, we start by importing the relevant packages:
+    
+    >>> from pygwb.pe import PowerLawModel
+    >>> from pygwb.baseline import Baseline
+    >>> import bilby.gw.detector as bilbydet
+    >>> impot numpy as np
+    
+    The model requires a list of baselines on which to run the parameter estimation. For the sake of the example,
+    we consider the Hanford-Livingston baseline, and create a baseline object:
+    
+    >>> H1 = bilbydet.get_empty_interferometer('H1')
+    >>> L1 = bilbydet.get_empty_interferometer('L1')
+    >>> HL = Baseline('H1L1', H1, L1)
+    
+    The frequencies, point estimate spectrum and sigma spectrum of the baseline need to be set as well. For this
+    example and for simplicity, we populate them with random numbers
+    
+    >>> HL.frequencies = np.linspace(0,100,1000)
+    >>> HL.point_estimate_spectrum = np.random.rand(1000)
+    >>> HL.sigma_spectrum = np.random.rand(1000)
+    
+    The baseline is then added to the other kwargs needed for the model to be instantiated:
+    
+    >>> kwargs_pl = {"baselines":[HL], "model_name":'PL', "fref":25}
+    >>> model_pl = PowerLawModel(**kwargs_pl)
+    
     """
 
     def __init__(self, **kwargs):
@@ -180,17 +210,17 @@ class BrokenPowerLawModel(GWBModel):
             \Omega_{\text{ref}} \left( \frac{f}{f_{\text{ref}}} \right) ^ {\alpha_2}, f > f_{\text{ref}}
         \end{cases}
     Parameters:
-    -----------
+    ===========
     omega_ref : float
-        amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
+        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
     alpha_1 : float
-        spectral index of the broken power law (:math:`\alpha_1`)
+        Spectral index of the broken power law (:math:`\alpha_1`)
     alpha_2 : float
-        spectral index of the broken power law (:math:`\alpha_2`)
+        Spectral index of the broken power law (:math:`\alpha_2`)
     fbreak : float
-        break frequency for the broken power law (:math:`f_{\text{ref}}`)
+        Break frequency for the broken power law (:math:`f_{\text{ref}}`)
     frequencies : numpy.ndarray
[…]
