+++ pygwb/delta_sigma_cut.py
@@ -53,6 +53,7 @@ def dsc_cut(
         corresponding GPS times were bad, whereas False denotes good GPS times.
     dsigma: array_like
         Values of the difference between sliding sigma and naive sigma, i.e., the actual value of the delta sigma per segment.
+    
     """
 
     dsigma = np.abs(slide_sigma * bf_ss - naive_sigma * bf_ns) / (slide_sigma * bf_ss)
