### 2023-06-02 4aa085c
Adding docstrings to util and detector.
