+++ pygwb/util.py
@@ -7,11 +7,20 @@ from scipy.signal import get_window
 
 from pygwb.constants import H0
 
+"""
+The util module collects all useful functions of pygwb that are used in multiple other modules or do not belong contextually in one of the other modules.
+These functions mainly performs computations that are relatively easy and at multiple steps of the pipeline analysis.
+"""
 
 def parse_window_dict(window_dict):
-    '''
-    Pasrse the window dictionary properly for scipy compatibility.
-    '''
+    """
+    Parse the window dictionary properly for scipy compatibility.
+    
+    Parameters
+    ==========
+    window_dict: `dictionary`
+        Dictionary containing the window characteristics.
+    """
     bools = ['sym', 'norm']
     floats = ['center', 'tau', 'alpha', 'beta', 'nbar', 'sll', 'std', 'p', 'sig', 'at']
     for key in window_dict.keys():
@@ -27,20 +36,21 @@ def window_factors(N, window_fftgram_dict={"window_fftgram": "hann"}, overlap_fa
     """
     Calculate window factors. By default, for a hann window with 50% overlap.
 
-    Parameters:
+    Parameters
     ===========
-    window_fftgram_dict: dictionary, optional
+    window_fftgram_dict: ``dictionary``, optional
         Dictionary with window characteristics. Default is `(window_fftgram_dict={"window_fftgram": "hann"}`
-    overlap_factor: float, optional
+    overlap_factor: ``float``, optional
         Defines the overlap between consecutive data chunks used in the calculation. Default is 0.5.
         
 
     Returns:
     ========
-    w1w2bar: float
-    w1w2squaredbar: float
-    w1w2ovlbar: float
-    w1w2squaredovlbar: float
+    This functions returns the four window factors which correct the analysis for the effect of the window factor.
+    w1w2bar: ``float``
+    w1w2squaredbar: ``float``
+    w1w2ovlbar: ``float``
+    w1w2squaredovlbar: ``float``
     """
     window_tuple = get_window_tuple(window_fftgram_dict)
     w = get_window(window_tuple, N, fftbins=False)
@@ -58,16 +68,16 @@ def window_factors(N, window_fftgram_dict={"window_fftgram": "hann"}, overlap_fa
 
 def get_window_tuple(window_fftgram_dict={"window_fftgram": "hann"}):
     """
-    Unpack the `window_fft_dict` dictionary into a `tuple` that may be read by scipy.get_window.
+    Unpack the `window_fft_dict` dictionary into a `tuple` that may be read by ``scipy.get_window``.
 
     Parameters:
     ===========
-    window_fftgram_dict: dictionary, optional
+    window_fftgram_dict: ``dictionary``, optional
         Dictionary with window characteristics. Default is `(window_fftgram_dict={"window_fftgram": "hann"}`.
 
     Returns:
     ========
-    window_tuple: tuple
+    window_tuple: ``tuple``
         A tuple containing the window_fft name as the first entry, followed by optional entries of the window_fft_dict.
 
     Notes:
@@ -91,16 +101,16 @@ def calc_rho1(N, window_fftgram_dict={"window_fftgram": "hann"}, overlap_factor=
 
     Parameters:
     ===========
-    N: int
+    N: ``int``
         Length of the window.
-    window_fftgram_dict: dictionary, optional
+    window_fftgram_dict: ``dictionary``, optional
         Dictionary with window characteristics. Default is `(window_fftgram_dict={"window_fftgram": "hann"}`.
-    overlap_factor: float, optional
+    overlap_factor: ``float``, optional
         Defines the overlap between consecutive data chunks used in the calculation. Default is 0.5.
 
     Returns:
     ========
-    rho1: float
+    rho1: ``float``
[…]
