### 2023-06-02 #5454 @plaplant
@Sbozzolo @a-renzini thanks for both of your contributions to the review process! I think it's best to focus on the most recent version of the code, given that the initial submission was several months ago and development has progressed since then. Most software repositories are understandably still under development (even post-JOSS publication!). We will make and tag a new release when the review process is completed, but it need not be the version that was initially submitted.

As for the question about GitLab vs. a GitHub mirror: [per JOSS's submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements), the software must be hosted at a place where users can open issues using accounts without manual approval. It seems that it not possible on [the GitLab version of the code](https://git.ligo.org/pygwb) to do that (though please correct me if that's not the case!), so the JOSS review and paper should be for the GitHub mirrored version. I agree that a name change clearly indicating it as a mirror might be helpful, but I don't think it's a requirement for publication if that is not desirable for some particular reason.
