### 2023-06-05 abc2b83
Fixing math environment in docstring
