+++ pygwb/orfs.py
@@ -5,6 +5,26 @@ from .constants import speed_of_light
 
 
 def Tplus(alpha, beta):
+    """
+    Function used in the computation of the tensor ORF, as given by 
+    Eq. (34) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    Tplus: 
+    
+    
+    """
     return (
         -(
             3.0 / 8 * spherical_jn(0, alpha)
@@ -27,6 +47,23 @@ def Tplus(alpha, beta):
 
 
 def Tminus(alpha, beta):
+    """
+    Function used in the computation of the tensor ORF, as given by 
+    Eq. (35) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return (
         spherical_jn(0, alpha)
         + 5.0 / 7 * spherical_jn(2, alpha)
@@ -35,6 +72,23 @@ def Tminus(alpha, beta):
 
 
 def Vplus(alpha, beta):
+    """
+    Function used in the computation of the vector ORF, as given by 
+    Eq. (37) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return (
         -(
             3.0 / 8 * spherical_jn(0, alpha)
@@ -57,6 +111,23 @@ def Vplus(alpha, beta):
 
 
 def Vminus(alpha, beta):
+    """
+    Function used in the computation of the vector ORF, as given by 
+    Eq. (38) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return (
         spherical_jn(0, alpha)
         - 5.0 / 14 * spherical_jn(2, alpha)
@@ -65,6 +136,23 @@ def Vminus(alpha, beta):
 
 
 def Splus(alpha, beta):
+    """
+    Function used in the computation of the scalar ORF, as given by 
+    Eq. (40) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return (
         -(
             3.0 / 8 * spherical_jn(0, alpha)
@@ -87,6 +175,23 @@ def Splus(alpha, beta):
 
 
 def Sminus(alpha, beta):
+    """
+    Function used in the computation of the scalar ORF, as given by 
+    Eq. (41) of https://arxiv.org/pdf/0903.0528.pdf
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return (
         spherical_jn(0, alpha)
         - 5.0 / 7 * spherical_jn(2, alpha)
@@ -95,6 +200,21 @@ def Sminus(alpha, beta):
 
 
 def T_right_left(alpha, beta):
+    """
+    
+    Parameters
+    ==========
+    
+    alpha:
+    
+    beta:
+    
+    
+    
+    Returns
+    =======
+    
+    """
     return -np.sin(beta / 2) * (
         (-spherical_jn(1, alpha) + 7.0 / 8 * spherical_jn(3, alpha))
[…]
