+++ pygwb/pe.py
@@ -136,7 +136,7 @@ class PowerLawModel(GWBModel):
     >>> from pygwb.pe import PowerLawModel
     >>> from pygwb.baseline import Baseline
     >>> import bilby.gw.detector as bilbydet
-    >>> impot numpy as np
+    >>> import numpy as np
     
     The model requires a list of baselines on which to run the parameter estimation. For the sake of the example,
     we consider the LIGO Hanford-Livingston baseline, and create a baseline object:
@@ -212,15 +212,15 @@ class BrokenPowerLawModel(GWBModel):
             \Omega_{\text{ref}} \left( \frac{f}{f_{\text{ref}}} \right) ^ {\alpha_2}, f > f_{\text{ref}}
         \end{cases}
         
-    Example
-    -------
+    Examples
+    --------
     
     To illustrate how to construct a likelihood with this model, we start by importing the relevant packages:
     
     >>> from pygwb.pe import BrokenPowerLawModel
     >>> from pygwb.baseline import Baseline
     >>> import bilby.gw.detector as bilbydet
-    >>> impot numpy as np
+    >>> import numpy as np
     
     The model requires a list of baselines on which to run the parameter estimation. For the sake of the example,
     we consider the LIGO Hanford-Livingston baseline, and create a baseline object:
@@ -305,32 +305,16 @@ class TripleBrokenPowerLawModel(GWBModel):
             \Omega_{\text{ref}} \left( \frac{f}{f_1} \right) ^ {\alpha_2}, f_1 < f \leqslant f_2  \\
             \Omega_{\text{ref}} \left( \frac{f_2}{f_1} \right) ^{\alpha_2} \left( \frac{f}{f_2} \right)^{\alpha_3}, f > f_2
         \end{cases}
-    Parameters:
-    ===========
-    omega_ref : float
-        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`) 
-    alpha_1 : float
-        Spectral index of the broken power law (:math:`\alpha_1`)
-    alpha_2 : float
-        Spectral index of the broken power law (:math:`\alpha_2`)
-    alpha_3 : float
-        Spectral index of the broken power law (:math:`\alpha_3`)
-    fbreak1 : float
-        First break frequency for the triple broken power law (:math:`f_1`)
-    fbreak2 : float
-        Second break frequency for the triple broken power law (:math:`f_2`)
-    frequencies : numpy.ndarray
-        Array of frequencies at which to evaluate the model
         
-    Example
-    -------
+    Examples
+    --------
     
     To illustrate how to construct a likelihood with this model, we start by importing the relevant packages:
     
     >>> from pygwb.pe import TripleBrokenPowerLawModel
     >>> from pygwb.baseline import Baseline
     >>> import bilby.gw.detector as bilbydet
-    >>> impot numpy as np
+    >>> import numpy as np
     
     The model requires a list of baselines on which to run the parameter estimation. For the sake of the example,
     we consider the LIGO Hanford-Livingston baseline, and create a baseline object:
@@ -350,6 +334,23 @@ class TripleBrokenPowerLawModel(GWBModel):
     
     >>> kwargs_tbpl = {"baselines":[HL], "model_name":'TBPL'}
     >>> model_tbpl = TripleBrokenPowerLawModel(**kwargs_tbpl)
+    
+    Parameters:
+    ===========
+    omega_ref : float
+        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`) 
+    alpha_1 : float
+        Spectral index of the broken power law (:math:`\alpha_1`)
+    alpha_2 : float
[…]
