+++ pygwb/pe.py
@@ -18,14 +18,15 @@ class GWBModel(bilby.Likelihood):
     where :math:`\Omega_{\rm M}(f_k|\mathbf{\Theta})` is the model being fit to data, and :math:`\mathbf{\Theta}` are the model's parameters.
 
     The noise likelihood is given by setting :math:`\Omega_{\rm M}(f_k|\mathbf{\Theta})=0`.
-    
-    See also
-    --------
-    bilby.Likelihood : The parent class used for the implementation.
 
     """
 
     def __init__(self, baselines=None, model_name=None, polarizations=None):
+        """
+        See also
+        --------
+        bilby.Likelihood : The parent class used for the implementation.
+        """
         super(GWBModel, self).__init__()
         # list of baselines
         if baselines is None:
@@ -183,24 +184,25 @@ class PowerLawModel(GWBModel):
     >>> kwargs_pl = {"baselines":[HL], "model_name":'PL', "fref":25}
     >>> model_pl = PowerLawModel(**kwargs_pl)
     
-    Parameters:
-    ===========
-    fref : float
-        Reference frequency for defining the model (:math:`f_{\text{ref}}`)
-    omega_ref : float
-        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
-    alpha : float
-        Spectral index of the power law (:math:`\alpha`)
-    frequencies : numpy.ndarray
-        Array of frequencies at which to evaluate the model
-        
-    See also
-    --------
-    pygwb.GWBModel : The parent class used for the implementation.
-    
     """
 
     def __init__(self, **kwargs):
+        """
+        Parameters:
+        ===========
+        fref : float
+            Reference frequency for defining the model (:math:`f_{\text{ref}}`)
+        omega_ref : float
+            Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
+        alpha : float
+            Spectral index of the power law (:math:`\alpha`)
+        frequencies : numpy.ndarray
+            Array of frequencies at which to evaluate the model
+        
+        See also
+        --------
+        pygwb.GWBModel : The parent class used for the implementation.
+        """
         try:
             fref = kwargs.pop("fref")
         except KeyError:
@@ -270,27 +272,28 @@ class BrokenPowerLawModel(GWBModel):
     
     >>> kwargs_bpl = {"baselines":[HL], "model_name":'BPL'}
     >>> model_bpl = BrokenPowerLawModel(**kwargs_bpl)
-        
-    Parameters:
-    ===========
-    omega_ref : float
-        Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
-    alpha_1 : float
-        Spectral index of the broken power law (:math:`\alpha_1`)
-    alpha_2 : float
-        Spectral index of the broken power law (:math:`\alpha_2`)
-    fbreak : float
-        Break frequency for the broken power law (:math:`f_{\text{ref}}`)
-    frequencies : numpy.ndarray
-        Array of frequencies at which to evaluate the model
-        
-    See also
-    --------
-    pygwb.GWBModel : The parent class used for the implementation.
     
     """
 
     def __init__(self, **kwargs):
+        """
+        Parameters:
+        ===========
+        omega_ref : float
+            Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
+        alpha_1 : float
+            Spectral index of the broken power law (:math:`\alpha_1`)
+        alpha_2 : float
+            Spectral index of the broken power law (:math:`\alpha_2`)
+        fbreak : float
+            Break frequency for the broken power law (:math:`f_{\text{ref}}`)
+        frequencies : numpy.ndarray
+            Array of frequencies at which to evaluate the model
+
+        See also
[…]
