+++ pygwb/simulator.py
@@ -11,6 +11,44 @@ from pygwb.util import interpolate_frequency_series
 
 
 class Simulator(object):
+    """
+    This module provides an easy-to-use tool to simulate data, both a stochastic gravitational-wave 
+    background, as well as individual compact binary coalescence waveforms.
+    
+    Examples
+    --------
+    
+    To illustrate how to create a ``simulator`` object, we start by 
+    importing the relevant packages:
+    
+    >>> import numpy as np
+    >>> from pygwb.simulator import Simulator
+    >>> import bilby.gw.detector as bilbydet
+    
+    For concreteness, we will simulate a signal for the LIGO Hanford-Livingston
+    baseline. We create the two detectors as follows:
+    
+    >>> H1 = bilbydet.get_empty_interferometer('H1')
+    >>> L1 = bilbydet.get_empty_interferometer('L1')
+    
+    As an example, we consider a flat signal PSD:
+    
+    >>> freqs = np.linspace(0, 1000, 10000)
+    >>> signal = gwpy.frequencyseries.FrequencySeries(1e-42*np.ones(10000), frequencies=freqs)
+    
+    The simulator object is then created as follows:
+    
+    >>> simulator_object = Simulator(interferometers = [H1,L1], N_segments = 1, duration = 64, 
+        sampling_frequency = 1024, intensity_GW = signal)
+        
+    The actual data is then simulated by calling the ``generate_data`` method:
+    
+    >>> data = simulator_object.generate_data()
+    
+    For more information, we refer the reader to the ``simulator`` tutorials.
+    
+    """
+    
     def __init__(
         self,
         interferometers,
@@ -26,36 +64,42 @@ class Simulator(object):
         continuous=False,
     ):
         """
-        Class that simulates an isotropic stochastic background.
+        Create an instance of the the ``simulator`` class. 
 
         Parameters
         ==========
-        interferometers: list of bilby interferometer objects
-        intensity_GW: gwpy.frequencyseries.FrequencySeries
+        
+        interferometers: list
+            List of bilby interferometer objects
+        intensity_GW: gwpy.frequencyseries.FrequencySeries, optional
             A gwpy.frequencyseries.FrequencySeries containing the desired strain power spectrum
-            which needs to be simulated. Note: A range of spectral indices (from -3 to 3) was
-            tested. However, one should be careful for spectral indices outside of this range,
-            as the splicing procedure implemented in this module is known to introduce a bias for
-            some values of the spectral index (usually large negative numbers).
-        injection_dict: dict
-            Dictionary containing the parameters of CBC injections.
+            which needs to be simulated. Defaults to None (then ``injection_dict`` should be passed).
+        injection_dict: dict, optional
+            Dictionary containing the parameters of CBC injections. Defaults to None (then 
+            ``intensity_GW`` should be passed). 
         N_segments: int
             Number of segments that needs to be generated for the simulation
         duration: float
-            Duration of a simulated data segment
+            Duration of a simulated data segment in seconds.
         sampling_frequency: float
-            Sampling frequency
-        start_time: float
-            Start time of the simulation
-        no_noise: boolean
-            Flag that sets the noise_PSDs to 0
-        seed: int
-            Seed to use for random data generation
-        continuous: boolean
-            Flag to generate continuous segments. If True, a seed must also be provided 
-
-        Returns
-        =======
+            Sampling frequency in Hz.
+        start_time: float, optional
+            Start time (in seconds) of the simulation. Defaults to 0.
+        no_noise: boolean, optional
+            Flag that sets the noise_PSDs to 0. Defaults to False.
+        seed: int, optional
+            Seed to use for random data generation. Defaults to None.
+        continuous: boolean, optional
[…]
