### 2023-06-05 78f47c3
Adding docstrings information in detector, preprocessing, spectral, util.
