+++ pygwb/preprocessing.py
@@ -1,3 +1,20 @@
+"""Preprocessing module collects all the functions that handle the preprocessing (what's in a name) of the data used in the analysis.
+
+The functions of the preprocessing module can handle anything related with preparing the data for the pygwb analysis run.
+Functions are present that read data from frame files, locally or publicly. Other functions are able to resample the data, apply a high-pass filter to it or apply a timeshift to the data.
+These functionalities comes together in the triplet of ``preprocessing_data`` functions which read in data and resample and/or high-pass it immediately.
+The triplet can work from a given ``gwpy.timeseries.TimeSeries``, a normal array or using a gravitational wave channel that will read in that channel from the provided local or public frame files. 
+
+The last function can gate the data based on the gate function in gwpy, `gwpy.timeseries.TimeSeries.gate`.
+
+Examples
+--------
+
+As an example, we will read in some data from a certain channel and then resample, high-pass and apply gating to the data.
+
+>>> from pygwb.preprocessing import *
+"""
+
 import copy
 import os
 import warnings
@@ -20,28 +37,29 @@ def set_start_time(
     either with or without sidereal option
 
     Parameters
-    ==========
-    job_start_GPS: int
+    ========
+    job_start_GPS: `int`
         Integer indicating the start time (in GPS)
-        of the data
+        of the data.
 
-    job_end_GPS: int
+    job_end_GPS: `int`
         Integer indicating the end time (in GPS)
-        of the data
+        of the data.
 
-    buffer_secs: int
-        Number of cropped seconds
+    buffer_secs: `int`
+        Number of cropped seconds.
 
-    segment_duration: int
-        Duration of each segment
+    segment_duration: `int`
+        Duration of each segment.
 
-    do_sidereal: bool
-        ---
+    do_sidereal: `bool`, optional
+        When this option is turned on, the code 
+        considers sidereal days instead of terrestial days.
 
     Returns
     =======
-    centered_start_time: int
-        Integer with the initial time of the segment to be pre-processed
+    centered_start_time: ``int``
+        Integer with the initial time of the segment to be pre-processed.
     """
     if not do_sidereal:
         job_duration = job_end_GPS - job_start_GPS
@@ -76,36 +94,37 @@ def read_data(
     stochastic pipeline
 
     Parameters
-    ==========
-    IFO: string
-        Interferometer from which to retrieve the data
+    ========
+    IFO: ``str``
+        Interferometer from which to retrieve the data.
 
-    data_type: string
+    data_type: ``str``
         String indicating the type of data to be read,
         either:
         - 'public' : data from GWOSC (https://www.gw-openscience.org/)
         - 'private' : data from the LIGO-Virgo servers restricted to members of the collaboration
         - 'local' (if 'local_data_path'): data the user has saved locally
 
-    channel: string
-        Name of the channel (e.g.: "L1:GWOSC-4KHZ_R1_STRAIN")
+    channel: ``str``
+        Name of the channel (e.g.: "L1:GWOSC-4KHZ_R1_STRAIN").
 
-    t0: int
-        GPS time of the start of the data taking
+    t0: ``int``
+        GPS time of the start of the data taking.
 
-    tf: int
-        GPS time of the end of the data taking
+    tf: ``int``
+        GPS time of the end of the data taking.
 
-    local_data_path: str, optional
-        path where local gwf is stored
+    local_data_path: ``str``, optional
+        Path where local gwf frame file is stored.
 
-    input_sample_rate: int
-        Sampling rate of the timeseries to be read
+    input_sample_rate: ``int``, optional
+        Sampling rate of the timeseries to be read;
+        default is 16384 Hz.
 
[…]
