### 2023-06-05 e35f555
Docstrings in spectral.
