+++ pygwb/postprocessing.py
@@ -1,3 +1,10 @@
+"""The postprocessing modules groups all functions which are useful in the end stages of the analysis. In particular when combining spectrograms into spectra and spectra into one overall point estimate for the gravitational-wave background.
+
+When computing an estimator for the gravitational-wave background in the isotropic search as it is ran by pygwb, we go through different steps of the analysis. Postprocessing is the last stage of the analysis and the functions utilised in that step are collected in this module. This step happens after the computation of the PSDs and CSDs. It contains functions such as :code:`postprocess_Y_sigma` which combines even and odd segments from point estimate and sigma spectrograms using a ingenious method where odd and even segments are treated differently if the data is overlapping. To achieve that, it uses another function of the postprocessing module, :code:`odd_even_segment_postprocessing`.
+These spectrograms have to be computed from the CSDs and PSDs. The postprocessing module also contains the function responsible for this calculation, :code:`calculate_point_estimate_sigma_spectra`. Starting from a set of CSD and PSD spectrograms, one can compute the point estimate and sigma spectrograms, objects that contain both frequency and segment data information. 
+Then the spectrograms are combined into spectra with the functions mentioned above. These spectra then have to be combined into one single point estimate and its variance. To achieve that goal another function, :code:`calc_Y_sigma_from_Yf_sigmaf`, is utilised.
+"""
+
 import numpy as np
 import scipy.ndimage as ndi
 from loguru import logger
@@ -24,35 +31,38 @@ def postprocess_Y_sigma(
     """Run postprocessing of point estimate and sigma spectrograms, combining even and
     odd segments in the case of overlapping data. For more details see - https://dcc.ligo.org/public/0027/T040089/000/T040089-00.pdf
 
-    Parameters:
-    -----------
-    Y_fs : array-like
-        2D array of point estimates with Ntimes x Nfreqs with overlapping segments
-    var_fs : array-like
-        2D array of variances or 2D with dimensions Ntimes x Nfreqs with overlapping time segments
-    segment_duration : float
-        Duration of each time segment
-    deltaF : float
-        Frequency resolution
-    new_sample_rate : float
-        sample rate of timeseries after resampling
+    Parameters
+    --------
+    Y_fs: array-like
+        2D array of point estimates with Ntimes x Nfreqs with overlapping segments.
+    var_fs: array-like
+        2D array of variances or 2D with dimensions Ntimes x Nfreqs with overlapping time segments.
+    segment_duration: `float`
+        Duration of each time segment.
+    deltaF: `float`
+        Frequency resolution.
+    new_sample_rate: `float`
+        Sample rate of timeseries after resampling.
     frequency_mask: array-like, optional
         Boolean mask to apply to frequencies for the calculation. Defaults to True which includes all frequencies in the analysis.
     badtimes_mask: array-like, optional
         Boolean mask to apply to GPStimes in the calculation. Defaults to None such that all times are included.
-    do_overlap: bool, optional
+    do_overlap: `bool`, optional
         Indicates whether to proceed with the odd/even postprocessing calculation which assumes a 50\% overlap between consecutive segments. Default is True.
-    window_fftgram_dict: dictionary, optional
+    window_fftgram_dict: `dictionary`, optional
         Dictionary with window characteristics used in PSD estimation. Default is `(window_fftgram_dict={"window_fftgram": "hann"}`
-    overlap_factor: float, optional
+    overlap_factor: `float`, optional
[…]
