+++ pygwb/postprocessing.py
@@ -123,6 +123,7 @@ def odd_even_segment_postprocessing(
     frequency_mask=True,
     window_fftgram_dict={"window_fftgram": "hann"},
     overlap_factor=0.5,
+    N_avg_segs=2,
 ):
     """Perform averaging which combines even and odd segments for overlapping data. 
 
