### 2023-06-06 543f7f8
Second exception test.
