+++ pygwb/notch.py
@@ -1,4 +1,4 @@
-"""The notch module handles all things considering notches. It bookkeeps and can calculate them with the provided classes and functions.
+"""The notch module handles all things considering notches. It bookkeeps and can calculates them with the provided classes and functions.
 
 This module has two main classes and a couple of functions all dealing with the createn of a notchlist used in the analysis to get rid of bad behaving frequencies.
 First of all there is the class StochNotch, based on the Notch class from ``bilby.gw.detector.strain_data``. This class stores a single Notch object containing a small description of the notch and the corresponding minimum and maximum frequency. 
@@ -77,7 +77,8 @@ class StochNotchList(list):
         notch_list: `list`
             A list of length-3 tuples of the (min, max) frequency; description for the notches.
 
-        :raises ValueError: If the list is malformed.
+        :raises:
+            ValueError: If the list is malformed.
         """
 
         if notch_list is not None:
@@ -89,17 +90,17 @@ class StochNotchList(list):
                     raise ValueError(msg)
 
     def check_frequency(self, freq):
-        """Check if freq is inside the notch list
+        """Check if freq is inside the notch list.
 
         Parameters
         ==========
-        freq: float
-            The frequency to check
+        freq: `float`
+            The frequency to check.
 
         Returns
         =======
         True/False:
-            If freq inside any of the notches, return True, else False
+            If freq inside any of the notches, return True, else False.
         """
 
         for notch in self:
@@ -108,21 +109,21 @@ class StochNotchList(list):
         return False
 
     def get_notch_mask(self, frequency_array, save_file_flag=False, filename=""):
-        """Get a boolean mask for the frequencies in frequency_array in the notch list
+        """Get a boolean mask for the frequencies in frequency_array in the notch list.
 
         Parameters
         ==========
-        frequency_array: np.ndarray
-            An array of frequencies
-        save_file_flag: boolean
-            A boolean flag indicating whether to save the notch mask in a file or not
-        filename: str
-            The name of the file where to store the notch mask if save_file_flag is true
+        frequency_array: `np.ndarray`
+            An array of frequencies.
+        save_file_flag: `bool`
+            A boolean flag indicating whether to save the notch mask in a file or not.
+        filename: `str`
+            The name of the file where to store the notch mask if save_file_flag is true.
 
         Returns
         =======
-        notch_mask: np.ndarray
-            An array of booleans that are False for frequencies in the notch
+        notch_mask: `np.ndarray`
+            An array of booleans that are False for frequencies in the notch.
 
         Notes
         =====
@@ -139,15 +140,15 @@ class StochNotchList(list):
         return notch_mask
 
     def save_notch_mask(self, frequency_array, filename):
-        """Saves a boolean mask for the frequencies in frequency_array in the notch list
+        """Saves a boolean mask for the frequencies in frequency_array in the notch list.
 
         Parameters
         ==========
-        frequency_array: np.ndarray
-            An array of frequencies
+        frequency_array: `np.ndarray`
+            An array of frequencies.
 
-        filename: str
-            Name of the target file
+        filename: `str`
+            Name of the target file.
 
         Notes
         =====
@@ -161,11 +162,11 @@ class StochNotchList(list):
         )
 
     def save_to_txt(self, filename):
-        """Save the nocth list to a txt-file (after sorting)
+        """Save the notch list to a txt-file (after sorting).
 
         Parameters
         ==========
-        filename: str
+        filename: `str`
             Name of the target file
 
         """
@@ -186,7 +187,7 @@ class StochNotchList(list):
         )
 
[…]
