### 2023-06-06 acd8343
Adding docstrings in notch and parameters.
