+++ pygwb/parameters.py
@@ -15,83 +15,83 @@ class Parameters:
     Parameters class: a dataclass which contains all parameters required for initialising a pygwb Interferometer, a pygwb Baseline, and run pygwb_pipe.
 
     Attributes
-    ----------
-    t0 : float
+    --------
+    t0 : `float`
         Initial time.
-    tf: float
+    tf: `float`
         Final time.
-    data_type: str
+    data_type: `str`
         Type of data to access/download; options are private, public, local. Default is public.
-    channel: str
-        Channel name; needs to match an existing channel. Default is \"GWOSC-16KHZ_R1_STRAIN\"
-    new_sample_rate: int
+    channel: `str`
+        Channel name; needs to match an existing channel. Default is \"GWOSC-16KHZ_R1_STRAIN\".
+    new_sample_rate: `int`
         Sample rate to use when downsampling the data (Hz). Default is 4096 Hz.
-    input_sample_rate: int
+    input_sample_rate: `int`
         Sample rate of the read data (Hz). Default is 16384 Hz.
-    cutoff_frequency: int
+    cutoff_frequency: `int`
         Lower frequency cutoff; applied in filtering in preprocessing (Hz). Default is 11 Hz.
-    segment_duration: int
+    segment_duration: `int`
         Duration of the individual segments to analyse (seconds). Default is 192 seconds.
-    number_cropped_seconds: int
+    number_cropped_seconds: `int`
         Number of seconds to crop at the start and end of the analysed data (seconds). Default is 2 seconds.
-    window_downsampling: str
-        Type of window to use in preprocessing. Default is \"hamming\"
-    ftype: str
-        Type of filter to use in downsampling. Default is \"fir\"
-    frequency_resolution: float
+    window_downsampling: `str`
+        Type of window to use in preprocessing. Default is \"hamming\".
+    ftype: `str`
+        Type of filter to use in downsampling. Default is \"fir\".
+    frequency_resolution: `float`
         Frequency resolution of the final output spectrum (Hz). Default is 1\/32 Hz.
-    polarization: str
+    polarization: `str`
         Polarisation type for the overlap reduction function calculation; options are scalar, vector, tensor. Default is tensor.
-    alpha: float
+    alpha: `float`
         Spectral index to filter the data for. Default is 0.
-    fref: int
+    fref: `int`
         Reference frequency to filter the data at (Hz). Default is 25 Hz.
-    flow: int
+    flow: `int`
         Lower frequency to include in the analysis (Hz). Default is 20 Hz.
-    fhigh: int
+    fhigh: `int`
         Higher frequency to include in the analysis (Hz). Default is 1726 Hz.
-    coarse_grain: bool
+    coarse_grain: `bool`
         Whether to apply coarse graining to the spectra. Default is 0.
-    interferometer_list: list
-        List of interferometers to run the analysis with. Default is [\"H1\", \"L1\"]
-    local_data_path: str
+    interferometer_list: `list`
+        List of interferometers to run the analysis with. Default is [\"H1\", \"L1\"].
+    local_data_path: `str`
         Path(s) to local data, if the local data option is chosen. Default is empty.
-    notch_list_path: str
+    notch_list_path: `str`
         Path to the notch list file. Default is empty.
-    N_average_segments_welch_psd: int
+    N_average_segments_welch_psd: `int`
         Number of segments to average over when calculating the psd with Welch method. Default is 2.
-    window_fft_dict: dict
+    window_fft_dict: `dict`
         Dictionary containing name and parameters describing which window to use when producing fftgrams for psds and csds. Default is \"hann\".
-    calibration_epsilon: float
+    calibration_epsilon: `float`
         Calibation coefficient. Default is 0.
-    overlap_factor: float
[…]
