+++ pygwb/simulator.py
@@ -1,3 +1,41 @@
+"""
+This module provides an easy-to-use tool to simulate data, both a stochastic gravitational-wave 
+background, as well as individual compact binary coalescence waveforms.
+    
+Examples
+--------
+    
+To illustrate how to create a ``simulator`` object, we start by 
+importing the relevant packages:
+    
+>>> import numpy as np
+>>> from pygwb.simulator import Simulator
+>>> import bilby.gw.detector as bilbydet
+    
+For concreteness, we will simulate a signal for the LIGO Hanford-Livingston
+baseline. We create the two detectors as follows:
+    
+>>> H1 = bilbydet.get_empty_interferometer('H1')
+>>> L1 = bilbydet.get_empty_interferometer('L1')
+    
+As an example, we consider a flat signal PSD:
+    
+>>> freqs = np.linspace(0, 1000, 10000)
+>>> signal = gwpy.frequencyseries.FrequencySeries(1e-42*np.ones(10000), frequencies=freqs)
+    
+The simulator object is then created as follows:
+    
+>>> simulator_object = Simulator(interferometers = [H1,L1], N_segments = 1, duration = 64, 
+    sampling_frequency = 1024, intensity_GW = signal)
+        
+The actual data is then simulated by calling the ``generate_data`` method:
+    
+>>> data = simulator_object.generate_data()
+    
+For more information, we refer the reader to the ``simulator`` tutorials.
+    
+"""
+
 import bilby
 import gwpy
 import numpy as np
@@ -11,44 +49,6 @@ from pygwb.util import interpolate_frequency_series
 
 
 class Simulator(object):
-    """
-    This module provides an easy-to-use tool to simulate data, both a stochastic gravitational-wave 
-    background, as well as individual compact binary coalescence waveforms.
-    
-    Examples
-    --------
-    
-    To illustrate how to create a ``simulator`` object, we start by 
-    importing the relevant packages:
-    
-    >>> import numpy as np
-    >>> from pygwb.simulator import Simulator
-    >>> import bilby.gw.detector as bilbydet
-    
-    For concreteness, we will simulate a signal for the LIGO Hanford-Livingston
-    baseline. We create the two detectors as follows:
-    
-    >>> H1 = bilbydet.get_empty_interferometer('H1')
-    >>> L1 = bilbydet.get_empty_interferometer('L1')
-    
-    As an example, we consider a flat signal PSD:
-    
-    >>> freqs = np.linspace(0, 1000, 10000)
-    >>> signal = gwpy.frequencyseries.FrequencySeries(1e-42*np.ones(10000), frequencies=freqs)
-    
-    The simulator object is then created as follows:
-    
-    >>> simulator_object = Simulator(interferometers = [H1,L1], N_segments = 1, duration = 64, 
-        sampling_frequency = 1024, intensity_GW = signal)
-        
-    The actual data is then simulated by calling the ``generate_data`` method:
-    
-    >>> data = simulator_object.generate_data()
-    
-    For more information, we refer the reader to the ``simulator`` tutorials.
-    
-    """
-    
     def __init__(
         self,
         interferometers,
@@ -69,27 +69,27 @@ class Simulator(object):
         Parameters
         ==========
         
-        interferometers: list
+        interferometers: ``list``
             List of bilby interferometer objects
-        intensity_GW: gwpy.frequencyseries.FrequencySeries, optional
+        intensity_GW: ``gwpy.frequencyseries.FrequencySeries``, optional
             A gwpy.frequencyseries.FrequencySeries containing the desired strain power spectrum
             which needs to be simulated. Defaults to None (then ``injection_dict`` should be passed).
-        injection_dict: dict, optional
+        injection_dict: ``dict``, optional
[…]
