### 2023-06-06 decc610
6 =.
