### 2023-06-06 #5454 @a-renzini
@plaplant - ok thanks! that all makes sense to me, and re. the mirror, yes, that is why we submitted from GitHub. We will discuss internally whether a name change would be useful (we may end up migrating to GitHub, but the timescale of the migration is I think beyond the scope of this review...)

@Sbozzolo and @cmbiwer -- apologies for the delay on the checklist, we have been populating the API and tutorials (see e.g. the new simulator page https://a-renzini.github.io/pygwb/api/pygwb.simulator.html by Max and Kevin, whom I can't seem to be able to tag at the moment) and I will take on the task of re-doing the install guide and contribution guide next week. I have been traveling last week/this week and I will now be offline until next week, fyi.
