### 2023-06-07 99f3766
Example detector.
