### 2023-06-07 963d863
Example for detector.
