### 2023-06-07 b2757ff
Adding examples to notch, post and preprocessing and spectral.
