### 2023-06-09 cdccdc4
update coh histogram, update calculation prediction, fix automatic histogram plotting, add 'significance' to outlier file and split outliers by notched/not notched
