### 2023-06-12 1f51911
Changing binning in hist sigma squared
