+++ pygwb/omega_spectra.py
@@ -93,17 +93,18 @@ class OmegaSpectrogram(Spectrogram):
 
     @classmethod
     def read(cls, source, *args, **kwargs):
-        """Read data into a Spectrogram. Same usage as read method of gwpy.Spectrogram.
+        """Read data into a Spectrogram. Same usage as read method of gwpy.spectrogram.Spectrogram.
 
         Parameters
         ==========
         source: :code:`str`
             Source file path.
+
         Returns
         ======
         Data: ``gwpy.spectrogram.Spectrogram``
-            The read in spectrogram from ``source``.
-        
+            The read in spectrogram from the source.
+
         See also
         --------
         gwpy.spectrogram.Spectrogram.read : the same implementation as this function.
@@ -112,13 +113,13 @@ class OmegaSpectrogram(Spectrogram):
         return io_registry.read(cls, source, *args, **kwargs)
 
     def write(self, target, *args, **kwargs):
-        """Write this Spectrogram to a file. Same usage as write method of gwpy.Spectrogram.
+        """Write this Spectrogram to a file. Same usage as write method of gwpy.spectrogram.Spectrogram.
 
         Parameters
         ==========
         target: str
             Target file path
-        
+
         See also
         --------
         gwpy.spectrogram.Spectrogram.write : the same implementation as this function.
