+++ pygwb/statistical_checks.py
@@ -54,6 +54,7 @@ class StatisticalChecks(object):
         baseline_name,
         param_file,
         frequency_mask = None,
+        notch_list = None,
         gates_ifo1 = None,
         gates_ifo2 = None,
         file_tag = None,
@@ -94,6 +95,8 @@ class StatisticalChecks(object):
             String with path to the file containing the parameters that were used for the analysis run.
         frequency_mask: array
             Boolean mask applied to the specrtra in broad-band analyses. 
+        notch_list: StochNotchList object
+            A StochNotchList object containg all the individual notches applied to the specrtra in broad-band analyses. 
         gates_ifo1/gates_ifo2: list
             List of gates applied to interferometer 1/2.
         file_tag: str
@@ -127,9 +130,12 @@ class StatisticalChecks(object):
         else:
             self.frequency_mask = True
 
+        self.notch_list = notch_list
+
         self.coherence_spectrum = coherence_spectrum
         fftlength = int(1.0 / (self.frequencies[1] - self.frequencies[0]))
         self.n_segs = coherence_n_segs * int(np.floor(self.params.segment_duration/(fftlength*(1.-self.params.overlap_factor_welch)))-1) #fftlength/2.
+        self.n_segs_statement = r"The number of segements is" + f" {self.n_segs}."
 
         self.sigma_spectrum = sigma_spectrum
         self.point_estimate_spectrum = point_estimate_spectrum
@@ -590,7 +596,7 @@ class StatisticalChecks(object):
         plt.close()
 
 
-    def plot_hist_coherence(self):
+    def plot_hist_coherence(self,total_bins = None):
         r"""
         Generates and saves a histogram of the coherence distribution. The plot shows the data after the delta-sigma cut (bad GPS times) was applied. This function does not require any input parameters, as it accesses the data through the attributes of the class.
         Furthermore, it also saves a text file which contains the frequencies at which outliers of the coherence distribution were identified, i.e. spectral artefacts.
@@ -598,7 +604,7 @@ class StatisticalChecks(object):
         if self.coherence_spectrum is None or self.coherence_spectrum.size==1:
             return
 
-        coherence = selfi.coherence_spectrum
+        coherence = self.coherence_spectrum
         coherence_clipped = np.ones(len(coherence))
         clip_val = 100* 1/self.n_segs
         for i in range(len(coherence_clipped)):
@@ -607,20 +613,24 @@ class StatisticalChecks(object):
             else:
                 coherence_clipped[i] = coherence[i]
         frequencies = self.frequencies
-        total_bins = 1000
+        if total_bins is None
+            total_bins = 250
         bins =  np.linspace(0, max(coherence), total_bins)
-        bins_clipped =  np.linspace(0, max(coherence), total_bins)
+        bins_clipped =  np.linspace(0, max(coherence_clipped), total_bins)
         n_frequencies = len(frequencies)
+        delta_coherence = bins[1]-bins[0]
         delta_coherence_clipped = bins_clipped[1]-bins_clipped[0]
         resolution = frequencies[1] - frequencies[0]
         fftlength = int(1.0 / resolution)
         
-        
+        coherence_notched = coherence*self.frequency_mask
+        coherence_notched_clipped = coherence_clipped*self.frequency_mask
+
         coherence_highres = np.arange(0,1,1e-6)
         predicted_highres = (self.n_segs-1) * (1- coherence_highres)**(self.n_segs-2)
         threshold = coherence_highres[np.where(predicted_highres <= 1/(n_frequencies*delta_coherence_clipped))[0][0]]
 
[…]
