+++ pygwb/baseline.py
@@ -1,3 +1,38 @@
+"""
+The ``Baseline`` module is by design the core of the pygwb stochastic analysis. Its main role is to manage the cross-
+correlation between ``Interferometer`` data products, combine these into a single cross-spectrum, which represents the
+point estimate of the analysis, and calculate the associated error.
+
+The ``Baseline`` object relies on the ``pygwb.spectral`` module to calculate cross-correlations between the data
+streams. Similarly, it relies on the ``pygwb.postprocessing`` module to obtain the point estimate and its variance.
+Calculating these, as well as performing parameter estimation on the GWB spectrum, requires the two-detector
+overlap reduction function (ORF). The ORF is calculated using the ``pygwb.orfs`` module at ``Baseline`` object 
+initialization, then stored as an attribute.
+
+Examples
+--------
+
+To show how a ``Baseline`` object can be instantiated, we start by importing the relevant
+packages:
+
+>>> import numpy as np
+>>> from pygwb.detector import Interferometer
+>>> from pygwb.baseline import Baseline
+
+For concreteness, we work with the LIGO Hanford and Livingston detectors, which we 
+instantiate through:
+
+>>> H1 = Interferometer.get_empty_interferometer("H1")
+>>> L1 = Interferometer.get_empty_interferometer("L1")
+
+The standard initialization of a ``Baseline`` object then simply requires a pair of 
+``Interferometer`` objects:
+
+>>> H1L1_baseline = baseline.Baseline("H1-L1", H1, L1)
+
+"""
+
+
 import json
 import pickle
 import warnings
@@ -21,10 +56,6 @@ from .spectral import cross_spectral_density
 
 
 class Baseline(object):
-    """
-    Baseline object for stochastic analyses.
-    """
-
     def __init__(
         self,
         name,
@@ -44,29 +75,30 @@ class Baseline(object):
         Instantiate a Baseline.
 
         Parameters
-        ----------
-        name: str
+        ==========
+
+        name: ``str``
             Name for the baseline, e.g H1H2
         interferometer_1/2: bilby Interferometer object
             The two detectors spanning the baseline
-        duration: float, optional
+        duration: ``float``, optional
             The duration in seconds of each data segment in the interferometers.
             None by default, in which case duration is inherited from the interferometers.
-        frequencies: array_like, optional
+        frequencies: ``array_like``, optional
             The frequency array for the Baseline and
             interferometers
-        calibration_epsilon: float, optional
+        calibration_epsilon: ``float``, optional
             Calibration uncertainty for this baseline
-        notch_list_path: str, optional
+        notch_list_path: ``str``, optional
             File path of the baseline notch list
-        overlap_factor: float, optional
+        overlap_factor: ``float``, optional
             Factor by which to overlap the segments in the psd and csd estimation.
             Default is 1/2, if set to 0 no overlap is performed.
-        zeropad_csd: bool, optional
+        zeropad_csd: ``bool``, optional
             If True, applies zeropadding in the csd estimation. True by default.
-        window_fftgram_dict: dictionary, optional
+        window_fftgram_dict: ``dictionary``, optional
             Dictionary containing name and parameters describing which window to use when producing fftgrams for psds and csds. Default is \"hann\".
-        N_average_segments_welch_psd: int, optional
+        N_average_segments_welch_psd: ``int``, optional
             Number of segments used for PSD averaging (from both sides of the segment of interest)
             N_avg_segs should be even and >= 2.
         """
@@ -176,8 +208,11 @@ class Baseline(object):
 
         Parameters
         ==========
-        notch_list_path: str
-            Path to notch list to apply to frequency array.
+        notch_list_path: ``str``, optional
+            Path to notch list to apply to frequency array. If not 
+            provided, no notching will be applied at this point.
+        apply_notches: ``bool``, optional
[…]
