+++ pygwb/pe.py
@@ -1,3 +1,8 @@
+"""
+
+
+"""
+
 from abc import abstractmethod
 
 import bilby
