+++ pygwb/orfs.py
@@ -1,3 +1,41 @@
+"""
+The relative position and orientation of two detectors is taken into account by the 
+overlap reduction function (ORF) in gravitational-wave background searches. The ``orfs``
+module combines the different component methods to compute the ORF for a given detector
+baseline, and polarization. By default, general relativity (GR) polarization is assumed, i.e., tensor. However, 
+the ``orfs`` module also supports non-GR polarizations (scalar and vector).
+
+Examples
+--------
+    
+To illustrate how to compute the ORF, we start by 
+importing the relevant packages:
+    
+>>> import numpy as np
+>>> from pygwb.orfs import *
+>>> import matplotlib.pyplot as plt
+    
+For concreteness, we consider the LIGO Hanford-Livingston baseline, and compute
+the ORF for this baseline. We define empty detectors:
+    
+>>> H1 = bilbydet.get_empty_interferometer('H1')
+>>> L1 = bilbydet.get_empty_interferometer('L1')
+    
+We  now compute the ORF for a set of frequencies by using the relevant information 
+contained in the interferometer objects defined above:
+    
+>>> freqs = np.arange(10.25, 256.25, 0.25)
+>>> orf = calc_orf(freqs, H1.vertex, L1.vertex, H1.x, L1.x, H1.y, L1.y, polarization = "tensor")
+    
+Note that the ``calc_orf`` method combines the various other methods of the module. The resulting ORF 
+looks as follows:
+    
+>>> plt.plot(freqs, orf)
+    
+Note that, in practice, these methods are not called by the user, but are
+called by the ``baseline`` module directly.
+"""
+
 import numpy as np
 from scipy.special import spherical_jn
 
@@ -12,18 +50,23 @@ def Tplus(alpha, beta):
     Parameters
     ==========
     
-    alpha: array_like
+    alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
-    beta: float
+    beta: ``float``
         Angle between detectors from center of the Earth
         
     Returns
     =======
     
-    Tplus: array_like
+    Tplus: ``array_like``
         Tplus parameter, as defined in Eq. (34) of https://arxiv.org/pdf/0903.0528.pdf.
+
+    See also
+    --------
+
+    scipy.special.spherical_jn: Used for computation of the spherical Bessel function.
     
     """
     return (
@@ -55,18 +98,23 @@ def Tminus(alpha, beta):
     Parameters
     ==========
     
-    alpha: array_like
+    alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
-    beta: float
+    beta: ``float``
         Angle between detectors from center of the Earth
     
     Returns
     =======
     
-    Tminus: array_like
+    Tminus: ``array_like``
         Tminus parameter, as defined in Eq. (35) of https://arxiv.org/pdf/0903.0528.pdf.
+
+    See also
+    --------
+
+    scipy.special.spherical_jn: Used for computation of the spherical Bessel function.
     
     """
     return (
@@ -84,18 +132,23 @@ def Vplus(alpha, beta):
     Parameters
     ==========
     
-    alpha: array_like
+    alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
-    beta: float
+    beta: ``float``
         Angle between detectors from center of the Earth
     
     Returns
     =======
     
-    Vplus: array_like
+    Vplus: ``array_like``
         Vplus parameter, as defined in Eq. (37) of https://arxiv.org/pdf/0903.0528.pdf.
+
[…]
