+++ pygwb/constants.py
@@ -1,3 +1,24 @@
+"""
+The ``constants`` module contains several constants that are used in stochastic gravitational-wave 
+background searches, such as the speed of light and the Hubble parameter. This module relies on 
+``astropy`` to obtain the value of these constants. 
+
+Examples
+--------
+
+To import one of the constants from this module inside another code, one would call the following
+line:
+
+>>> from pygwb.constants import H0
+
+See also
+--------
+
+astropy.constants: Used to obtain the speed of light.
+
+astropy.cosmology: Used to obtain the cosmology as given by Planck18 results.
+"""
+
 from astropy.constants import c
 from astropy.cosmology import Planck18 as cosmo
 
