+++ pygwb/omega_spectra.py
@@ -1,10 +1,9 @@
-"""
-This module contains two self-implemented classes that are responsible for dealing with the spectra and spectrograms in the pygwb analysis. 
+"""This module contains two self-implemented classes that are responsible for dealing with the spectra and spectrograms in the pygwb analysis. 
 These objects can be reweighted with different h0 and power law indices.
 
-During the creation of the pygwb code, it was decided that the capabilities of the ``gwpy.spectrogram.Spectrogram``
+During the creation of the pygwb code, it was decided that the capabilities of the :code: `gwpy.spectrogram.Spectrogram`
 were not sufficient, even though they are already very expanded, and thus we made some subclasses inheriting 
-from the ``Spectrogram`` class. 
+from the :code:`Spectrogram` class. 
 
 Their main purpose was to expand the functionality of the spectrogram to include the ability 
 to read and save to a pickle, but more crucially to reweight the data inside the spectrogram.
@@ -13,12 +12,12 @@ quickly some new code and/or run PE. Therefore, we opted to let the spectrogram
 be reweighted for a certain power law index of the background and also for a certain value
 of h0, the Hubble constant at present time. 
 
-We also introduced a new spectrum class based on the ``gwpy.frequencyseries.FrequencySeries`` class.
+We also introduced a new spectrum class based on the :code:`gwpy.frequencyseries.FrequencySeries` class.
 This new subclass needed the same additions as the spectrograms.
 
 We also coded one function which is actually the function that performs the reweighting 
-both for the ``OmegaSpectrum`` and ``OmegaSpectrogram``. That function is 
-``reweight_spectral_object``.
+both for the :code:`OmegaSpectrum` and :code:`OmegaSpectrogram`. That function is 
+:code:`reweight_spectral_object`.
 """
 
 import pickle
@@ -40,7 +39,6 @@ class OmegaSpectrogram(Spectrogram):
     
     See also
     --------
-
     gwpy.spectrogram.Spectrogram : the parent class of this implementation.
     """
 
@@ -95,30 +93,26 @@ class OmegaSpectrogram(Spectrogram):
 
     @classmethod
     def read(cls, source, *args, **kwargs):
-        """
-        Read data into a Spectrogram. Same usage as read method of gwpy.spectrogram.Spectrogram.
+        """Read data into a Spectrogram. Same usage as read method of gwpy.spectrogram.Spectrogram.
 
         Parameters
         ==========
-
-        source: str
+        source: :code:`str`
             Source file path.
 
         Returns
-        =======
-
+        ======
         Data: ``gwpy.spectrogram.Spectrogram``
             The read in spectrogram from the source.
+        
         """
         return io_registry.read(cls, source, *args, **kwargs)
 
     def write(self, target, *args, **kwargs):
-        """
-        Write this Spectrogram to a file. Same usage as write method of gwpy.spectrogram.Spectrogram.
+        """Write this Spectrogram to a file. Same usage as write method of gwpy.spectrogram.Spectrogram.
 
         Parameters
         ==========
-
         target: str
             Target file path
         """
@@ -131,13 +125,11 @@ class OmegaSpectrogram(Spectrogram):
 
         Parameters
         ==========
-
-        filename: ``str``
+        filename: `str`
             Filename (inclusive of path) to load the pickled spectrogram from.
             
         Returns
-        =======
-
+        ======
         Spectrogram
             The spectrogram you wanted to read from ``filename``.
         """
@@ -150,7 +142,6 @@ class OmegaSpectrogram(Spectrogram):
 
         Parameters
         ==========
-
         filename: str
[…]
