### 2023-06-13 0fe91ff
Update parameters_detector_test.ini
