+++ test/test_data/parameters_detector_test.ini
@@ -18,7 +18,6 @@ number_cropped_seconds: 2
 window_downsampling: hamming
 ftype: fir
 [gating]
-path_gate_data: /home/max.lalleman/public_html/Code/New_Clone_pyGWB/pygwb/test/test_data
 gate_data: False
 gate_whiten: True
 gate_tzero: 1
@@ -51,5 +50,3 @@ delta_sigma_cut: 0.2
 [output]
 save_data_type: json
 [local_data]
-H1: /home/arianna.renzini/PROJECTS/SMDC_2021/power_law_plus_peak_fork/MDC_data_plpp_d_60_a_1p8_DO5/output/H-H1_STRAIN-0-20480.gwf
-L1: /home/arianna.renzini/PROJECTS/SMDC_2021/power_law_plus_peak_fork/MDC_data_plpp_d_60_a_1p8_DO5/output/L-L1_STRAIN-0-20480.gwf
