+++ pygwb_pipe/parameters.ini
@@ -29,10 +29,8 @@ window_downsampling: hamming
 ftype: fir
 [gating]
 ; gating parameters
-; Path to file containing gates to load
-path_gate_data: ../test/test_data/point_estimate_sigma_1247644138-1247645038.npz
 ; Running self-gating or not
-gate_data: True
+gate_data: False
 ; Whitening the data before gating
 gate_whiten: True
 ; Tzero for gating: half-width time duration (seconds) in which the timeseries is set to zero
