### 2023-06-13 18e1677
editing stats: adding stochmon time conventions
