+++ pygwb/statistical_checks.py
@@ -9,6 +9,8 @@ import matplotlib.pyplot as plt
 import matplotlib.transforms as mt
 from loguru import logger
 
+from astropy.time import Time
+
 matplotlib.rcParams['figure.figsize'] = (8,6)
 matplotlib.rcParams['axes.grid'] = True
 matplotlib.rcParams['grid.linestyle'] = ':'
@@ -54,11 +56,11 @@ class StatisticalChecks(object):
         baseline_name,
         param_file,
         frequency_mask = None,
-        notch_list = None,
         gates_ifo1 = None,
         gates_ifo2 = None,
         file_tag = None,
-        legend_fontsize = 16
+        legend_fontsize = 16,
+        convention = 'pygwb'
     ):
         """
         The statistical checks class performs various tests by plotting different quantities and saving this plots. This allows the user to check for consistency with expected results. Concretely, the following tests and plots can be generated: running point estimate, running sigma, (cumulative) point estimate integrand, real and imaginary part of point estimate integrand, FFT of the point estimate integrand, (cumulative) sensitivity, evolution of omega and sigma as a function of time, omega and sigma distribution, KS test, and a linear trend analysis of omega in time. Furthermore, part of these plots compares the values of these quantities before and after the delta sigma cut. Each of these plots can be made by calling the relevant class method (e.g. `plot_running_point_estimate()`).
@@ -95,8 +97,6 @@ class StatisticalChecks(object):
             String with path to the file containing the parameters that were used for the analysis run.
         frequency_mask: array
             Boolean mask applied to the specrtra in broad-band analyses. 
-        notch_list: StochNotchList object
-            A StochNotchList object containing all the individual notches applied to the specrtra in broad-band analyses. 
         gates_ifo1/gates_ifo2: list
             List of gates applied to interferometer 1/2.
         file_tag: str
@@ -129,8 +129,6 @@ class StatisticalChecks(object):
             self.frequency_mask = frequency_mask
         else:
             self.frequency_mask = True
-
-        self.notch_list = notch_list
 
         self.coherence_spectrum = coherence_spectrum
         fftlength = int(1.0 / (self.frequencies[1] - self.frequencies[0]))
@@ -173,7 +171,11 @@ class StatisticalChecks(object):
             self.running_sigmas,
         ) = self.compute_running_quantities()
 
-        self.time_tag = f"{int(self.sliding_times_all[0])}"+"$-$"+f"{int(self.params.tf)}"
+        t0 = Time(self.sliding_times_all[0], format='gps')
+        t0 = Time(t0, format='iso', scale='utc', precision=0)
+        tf = Time(self.params.tf, format='gps')
+        tf = Time(tf, format='iso', scale='utc', precision=0)
+        self.time_tag = f"{t0}"+" $-$ "+f"{tf}"
 
         if file_tag:
             self.file_tag = file_tag
@@ -184,6 +186,14 @@ class StatisticalChecks(object):
         self.axes_labelsize = legend_fontsize + 2
         self.title_fontsize = legend_fontsize + 4
         self.annotate_fontsize = legend_fontsize - 4
+
+        ## convention: stochmon
+        if convention == 'stochmon':
[…]
