### 2023-06-13 aab6b13
changed clipping threshold for coherence plots
