### 2023-06-14 77235f3
Fix logical error when saving (un)notched coherence outliers to txt file.
