+++ pygwb/constants.py
@@ -16,6 +16,9 @@ See also
 
 astropy.constants: Used to obtain the speed of light.
 
+See also
+--------
+
 astropy.cosmology: Used to obtain the cosmology as given by Planck18 results.
 """
 
