+++ pygwb/constants.py
@@ -14,9 +14,9 @@ line:
 See also
 --------
 
-astropy.constants: Used to obtain the speed of light.
+astropy.constants : Used to obtain the speed of light.
 
-astropy.cosmology: Used to obtain the cosmology as given by Planck18 results.
+astropy.cosmology : Used to obtain the cosmology as given by Planck18 results.
 
 """
 
