### 2023-06-15 9e1ecdb
Fix normalization
