### 2023-06-16 396437f
fixed coherence calculation
