+++ pygwb/statistical_checks.py
@@ -1,3 +1,14 @@
+"""
+The statistical checks class performs various tests by plotting different quantities and saving this plots. 
+This allows the user to check for consistency with expected results. Concretely, the following tests and plots
+can be generated: running point estimate, running sigma, (cumulative) point estimate integrand, real and imaginary 
+part of point estimate integrand, FFT of the point estimate integrand, (cumulative) sensitivity, evolution of omega 
+and sigma as a function of time, omega and sigma distribution, KS test, and a linear trend analysis of omega in time. 
+Furthermore, part of these plots compares the values of these quantities before and after the delta sigma cut. Each of 
+these plots can be made by calling the relevant class method (e.g. `plot_running_point_estimate()`).
+
+"""
+
 from os import listdir
 from os.path import isfile, join
 from pathlib import Path
@@ -48,27 +59,28 @@ class StatisticalChecks(object):
         legend_fontsize = 16
     ):
         """
-        The statistical checks class performs various tests by plotting different quantities and saving this plots. This allows the user to check for consistency with expected results. Concretely, the following tests and plots can be generated: running point estimate, running sigma, (cumulative) point estimate integrand, real and imaginary part of point estimate integrand, FFT of the point estimate integrand, (cumulative) sensitivity, evolution of omega and sigma as a function of time, omega and sigma distribution, KS test, and a linear trend analysis of omega in time. Furthermore, part of these plots compares the values of these quantities before and after the delta sigma cut. Each of these plots can be made by calling the relevant class method (e.g. `plot_running_point_estimate()`).
+        Instantiate a StatisticalChecks object.
 
         Parameters
         ==========
-        sliding_times_all: array
+
+        sliding_times_all: array_like
             Array of GPS times before the bad GPS times from the delta sigma cut are applied.
-        sliding_omega_all: array
+        sliding_omega_all: array_like
             Array of sliding omegas before the bad GPS times from the delta sigma cut are applied.
-        sliding_sigmas_all: array
+        sliding_sigmas_all: array_like
             Array of sliding sigmas before the bad GPS times from the delta sigma cut are applied.
-        naive_sigmas_all: array
+        naive_sigmas_all: array_like
             Array of naive sigmas before the bad GPS times from the delta sigma cut are applied.
-        point_estimate_spectrum: array
+        point_estimate_spectrum: array_like
             Array containing the point estimate spectrum. Each entry in this array corresponds to the point estimate spectrum evaluated at the corresponding frequency in the freqs array.
-        sigma_spectrum: array
+        sigma_spectrum: array_like
             Array containing the sigma spectrum. Each entry in this array corresponds to the sigma spectrum evaluated at the corresponding frequency in the freqs array.
-        freqs: array
+        freqs: array_like
             Array containing the frequencies.
-        badGPStimes: array
+        badGPStimes: array_like
             Array of bad GPS times, i.e. times that do not pass the delta sigma cut.
-        delta_sigmas: array
+        delta_sigmas: array_like
             Array containing the value of delta sigma for all times in sliding_times_all.
         plot_dir: str
             String with the path to which the output of the statistical checks (various plots) will be saved.
@@ -77,9 +89,6 @@ class StatisticalChecks(object):
         param_file: str
             String with path to the file containing the parameters that were used for the analysis run.
 
-        Returns
-        =======
-        Initializes an instance of the statistical checks class.
         """
         self.params = Parameters()
         self.params.update_from_file(param_file)
@@ -134,23 +143,26 @@ class StatisticalChecks(object):
 
[…]
