### 2023-06-16 77afa17
adding note about coherence limits
