+++ pygwb/statistical_checks.py
@@ -133,6 +133,7 @@ class StatisticalChecks(object):
         fftlength = int(1.0 / (self.frequencies[1] - self.frequencies[0]))
         self.n_segs = coherence_n_segs*(1.-self.params.overlap_factor) * int(np.floor(self.params.segment_duration/(fftlength*(1.-self.params.overlap_factor_welch)))-1)
         if self.params.coarse_grain_csd:
+            # Note: this breaks down when self.params.segment_duration/fftlength < 3
             self.n_segs = coherence_n_segs*(1.-self.params.overlap_factor) * int(np.floor(self.params.segment_duration/(fftlength)))
         self.n_segs_statement = r"The number of segments is" + f" {self.n_segs}."
 
