### 2023-06-19 d70a12b
deprecating do_overlap flag: it was not used and created bugs
