+++ pygwb/postprocessing.py
@@ -16,7 +16,6 @@ def postprocess_Y_sigma(
     new_sample_rate,
     frequency_mask=True,
     badtimes_mask=None,
-    do_overlap=True,
     window_fftgram_dict={"window_fftgram": "hann"},
     window_fftgram_dict_welch={"window_fftgram": "hann"},
     overlap_factor=0.5,
@@ -42,8 +41,6 @@ def postprocess_Y_sigma(
         Boolean mask to apply to frequencies for the calculation. Defaults to True which includes all frequencies in the analysis.
     badtimes_mask: array-like, optional
         Boolean mask to apply to GPStimes in the calculation. Defaults to None such that all times are included.
-    do_overlap: bool, optional
-        Indicates whether to proceed with the odd/even postprocessing calculation which assumes a 50\% overlap between consecutive segments. Default is True.
     window_fftgram_dict: dictionary, optional
         Dictionary with window characteristics used in PSD estimation. Default is `(window_fftgram_dict={"window_fftgram": "hann"}`
     overlap_factor: float, optional
@@ -69,7 +66,7 @@ def postprocess_Y_sigma(
         Y = Y_fs[sli]
         var = var_fs[sli]
 
-        if len(Y) == 1 or not do_overlap:
+        if len(Y) == 1:
             Y_fs_sliced.append(Y[0])
             var_fs_sliced.append(var[0])
         else:
