+++ JOSS_submission/paper.md
@@ -1,163 +1,0 @@
----
-title: '`pygwb`: a Python-based library for gravitational-wave background searches'
-tags:
-  - Python
-  - astronomy®
-  - physics
-  - gravitational waves
-authors:
-  - name: Arianna I. Renzini
-    orcid: 0000-0002-4589-3987
-    corresponding: true
-    email: arenzini@caltech.edu
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
-  - name: Alba Romero-Rodriguez
-    orcid: 0000-0003-2275-4164
-    affiliation: "3"
-  - name: Colm Talbot
-    orcid: 0000-0003-2053-5582
-    affiliation: "4"
-  - name: Max Lalleman
-    orcid: 0000-0002-2254-010X
-    affiliation: "5"
-  - name: Shivaraj Kandhasamy
-    orcid: 0000-0002-4825-6764
-    affiliation: "6"
-  - name: Kevin Turbang
-    orcid: 0000-0002-9296-8603
-    affiliation: "3, 5"
-  - name: Sylvia Biscoveanu
-    orcid: 0000-0001-7616-7366
-    affiliation: "4, 7"
-  - name: Katarina Martinovic
-    orcid: 0000-0001-5299-7744
-    affiliation: "8"
-  - name: Patrick Meyers
-    orcid: 0000-0002-2689-0190
-    affiliation: "9"
-  - name: Leo Tsukada
-    affiliation: "10, 11"
-  - name: Kamiel Janssens
-    orcid: 0000-0001-8760-4429
-    affiliation: "5, 12"
-  - name: Derek Davis
-    orcid: 0000-0001-5620-6751
-    affiliation: "1, 2"
-  - name: Andrew Matas 
-    affiliation: "13"
-  - name: Philip Charlton
-    orcid: 0000-0002-4263-2706
-    affiliation: "14"
-  - name: Guo-chin Liu
-    affiliation: "15"
-  - name: Irina Dvorkin
-    orcid: 0000-0002-2353-9194
-    affiliation: "16"
-    
-affiliations:
- - name: LIGO Laboratory,  California  Institute  of  Technology,  Pasadena,  California  91125,  USA
-   index: 1
- - name: Department of Physics, California Institute of Technology, Pasadena, California 91125, USA
-   index: 2
- - name: Theoretische Natuurkunde, Vrije Universiteit Brussel, Pleinlaan 2, B-1050 Brussels, Belgium
-   index: 3
- - name: Kavli Institute for Astrophysics and Space Research, Massachusetts Institute of Technology, 77 Massachusetts Ave, Cambridge, MA 02139, USA
-   index: 4
- - name: Universiteit Antwerpen, Prinsstraat 13, 2000 Antwerpen, België
-   index: 5
- - name: Inter-University Centre for Astronomy and Astrophysics, Pune 411007, India
-   index: 6
- - name: LIGO Laboratory, Massachusetts Institute of Technology, 185 Albany St, Cambridge, MA 02139, USA
-   index: 7
- - name: Theoretical Particle Physics and Cosmology Group, Physics Department, King’s College London, University of London, Strand, London WC2R 2LS, United Kingdom
-   index: 8
- - name: Theoretical Astrophysics Group, California Institute of Technology, Pasadena, CA 91125, USA
-   index: 9
- - name: Department of Physics, The Pennsylvania State University, University Park, Pennsylvania 16802, USA
-   index: 10
- - name: Institute for Gravitation and the Cosmos, The Pennsylvania State University, University Park, Pennsylvania 16802, USA
-   index: 11
- - name: Université Côte d’Azur, Observatoire Côte d’Azur, ARTEMIS, Nice, France
-   index: 12
[…]
