+++ JOSS_submission/paper.bib
@@ -1,224 +1,0 @@
-@article{gwpy,
-    title = "{GWpy: A Python package for gravitational-wave astrophysics}",
-   author = {{Macleod}, D.~M. and {Areeda}, J.~S. and {Coughlin}, S.~B. and {Massinger}, T.~J. and {Urban}, A.~L.},
-  journal = {SoftwareX},
-   volume = 13,
-    pages = 100657,
-     year = 2021,
-     issn = {2352-7110},
-      doi = {10.1016/j.softx.2021.100657},
-      url = {https://www.sciencedirect.com/science/article/pii/S2352711021000029},
-}
-@article{Ashton_2019,
-	doi = {10.3847/1538-4365/ab06fc},
-
-	url = {https://doi.org/10.3847%2F1538-4365%2Fab06fc},
-  
-	year = 2019,
-	month = {apr},
-  
-	publisher = {American Astronomical Society},
-  
-	volume = {241},
-  
-	number = {2},
-  
-	pages = {27},
-  
-	author = {Gregory Ashton and Moritz Hübner and Paul D. Lasky and Colm Talbot and Kendall Ackley and Sylvia Biscoveanu and Qi Chu and Atul Divakarla and Paul J. Easter and Boris Goncharov and Francisco Hernandez Vivanco and Jan Harms and Marcus E. Lower and Grant D. Meadors and Denyz Melchor and Ethan Payne and Matthew D. Pitkin and Jade Powell and Nikhil Sarin and Rory J. E. Smith and Eric Thrane},
-  
-	title = {Bilby: A User-friendly Bayesian Inference Library for Gravitational-wave Astronomy},
-  
-	journal = {The Astrophysical Journal Supplement Series}
-}
-@article{Speagle_2020,
-	doi = {10.1093/mnras/staa278},
-  
-	url = {https://doi.org/10.1093%2Fmnras%2Fstaa278},
-  
-	year = 2020,
-	month = {feb},
-  
-	publisher = {Oxford University Press ({OUP})},
-  
-	volume = {493},
-  
-	number = {3},
-  
-	pages = {3132--3158},
-  
-	author = {Joshua S Speagle},
-  
-	title = {dynesty: a dynamic nested sampling package for estimating Bayesian posteriors and evidences},
-  
-	journal = {Monthly Notices of the Royal Astronomical Society}
-}
-@article{Collaboration2022TheAP,
-  title={The Astropy Project: Sustaining and Growing a Community-oriented Open-source Project and the Latest Major Release (v5.0) of the Core Package},
[…]
