+++ pygwb/pe.py
@@ -54,9 +54,9 @@ class GWBModel(bilby.Likelihood):
         super(GWBModel, self).__init__()
         # list of baselines
         if baselines is None:
-            raise ValueError("list of baselines must be supplied")
+            raise ValueError("List of baselines must be supplied!")
         if model_name is None:
-            raise ValueError("model_name must be supplied")
+            raise ValueError("model_name must be supplied!")
         # if single baseline supplied, that's fine
         # just make it a list
         if isinstance(baselines, Baseline):
@@ -82,33 +82,33 @@ class GWBModel(bilby.Likelihood):
 
     @abstractmethod
     def parameters(self):
-        """Parameters to be inferred from the data. Should return a dictionary"""
+        """Parameters to be inferred from the data. Should return a dictionary."""
         pass
 
     @abstractmethod
     def model_function(self):
-        """Function for evaluating model"""
+        """Function for evaluating model."""
         pass
 
     def log_likelihood_IJ(self, baseline, noise=False):
         """
-        Function for evaluating log likelihood of IJ baseline pair
+        Function for evaluating log likelihood of IJ baseline pair.
         
         Parameters
-        ==========
+        =======
         
-        baseline: pygwb.Baseline
+        baseline: ``pygwb.Baseline``
             Baseline for which to run parameter estimation on.
         
-        noise: bool, optional
+        noise: ``bool``, optional
             Parameter to indicate whether the likelihood should be evaluated assuming
             the signal model, or assuming only noise is present in the data.
         
         Returns
         =======
         
-        logL_IJ: float
-            Log likelihood value for the IJ baseline pair
+        logL_IJ: ``float``
+            Log likelihood value for the IJ baseline pair.
         """
         if noise:
             Y_model_f = 0
@@ -154,7 +154,7 @@ class GWBModel(bilby.Likelihood):
 
     def log_likelihood(self):
         """
-        Function for evaluating log likelihood of detector network
+        Function for evaluating log likelihood of detector network.
         """
         ll = 0
         for baseline in self.baselines:
@@ -213,15 +213,15 @@ class PowerLawModel(GWBModel):
     def __init__(self, **kwargs):
         """
         Parameters
-        ==========
+        =======
         
-        fref : float
-            Reference frequency for defining the model (:math:`f_{\text{ref}}`)
-        omega_ref : float
-            Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
-        alpha : float
-            Spectral index of the power law (:math:`\alpha`)
-        frequencies : numpy.ndarray
+        fref : ``float``
+            Reference frequency for defining the model (:math:`f_{\\text{ref}}`)
+        omega_ref : ``float``
+            Amplitude of signal at :math:`f_{\\text{ref}}` (:math:`\Omega_{\\text{ref}}`)
+        alpha : ``float``
+            Spectral index of the power law (:math:`\\alpha`)
+        frequencies : ``numpy.ndarray``
             Array of frequencies at which to evaluate the model
         
         See also
@@ -303,17 +303,17 @@ class BrokenPowerLawModel(GWBModel):
     def __init__(self, **kwargs):
         """
         Parameters
-        ==========
+        =======
         
-        omega_ref : float
-            Amplitude of signal at fref (:math:`\Omega_{\text{ref}}`)
-        alpha_1 : float
[…]
