+++ pygwb/orfs.py
@@ -45,17 +45,17 @@ from .constants import speed_of_light
 def Tplus(alpha, beta):
     """
     Function used in the computation of the tensor ORF, as given by 
-    Eq. (34) of https://arxiv.org/pdf/0903.0528.pdf
+    Eq. (34) of https://arxiv.org/pdf/0903.0528.pdf.
     
     Parameters
-    ==========
+    =======
     
     alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
     beta: ``float``
-        Angle between detectors from center of the Earth
+        Angle between detectors from center of the Earth.
         
     Returns
     =======
@@ -93,17 +93,17 @@ def Tplus(alpha, beta):
 def Tminus(alpha, beta):
     """
     Function used in the computation of the tensor ORF, as given by 
-    Eq. (35) of https://arxiv.org/pdf/0903.0528.pdf
+    Eq. (35) of https://arxiv.org/pdf/0903.0528.pdf.
     
     Parameters
-    ==========
+    =======
     
     alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
     beta: ``float``
-        Angle between detectors from center of the Earth
+        Angle between detectors from center of the Earth.
     
     Returns
     =======
@@ -127,17 +127,17 @@ def Tminus(alpha, beta):
 def Vplus(alpha, beta):
     """
     Function used in the computation of the vector ORF, as given by 
-    Eq. (37) of https://arxiv.org/pdf/0903.0528.pdf
+    Eq. (37) of https://arxiv.org/pdf/0903.0528.pdf.
     
     Parameters
-    ==========
+    =======
     
     alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
     beta: ``float``
-        Angle between detectors from center of the Earth
+        Angle between detectors from center of the Earth.
     
     Returns
     =======
@@ -175,10 +175,10 @@ def Vplus(alpha, beta):
 def Vminus(alpha, beta):
     """
     Function used in the computation of the vector ORF, as given by 
-    Eq. (38) of https://arxiv.org/pdf/0903.0528.pdf
+    Eq. (38) of https://arxiv.org/pdf/0903.0528.pdf.
     
     Parameters
-    ==========
+    =======
     
     alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
@@ -209,17 +209,17 @@ def Vminus(alpha, beta):
 def Splus(alpha, beta):
     """
     Function used in the computation of the scalar ORF, as given by 
-    Eq. (40) of https://arxiv.org/pdf/0903.0528.pdf
+    Eq. (40) of https://arxiv.org/pdf/0903.0528.pdf.
     
     Parameters
-    ==========
+    =======
     
     alpha: ``array_like``
         Given below Eq. (32) of https://arxiv.org/pdf/0903.0528.pdf. 
         Has same shape as frequencies used to compute the ORF.
     
     beta: ``float``
-        Angle between detectors from center of the Earth
+        Angle between detectors from center of the Earth.
     
     Returns
     =======
@@ -257,7 +257,7 @@ def Splus(alpha, beta):
 def Sminus(alpha, beta):
     """
     Function used in the computation of the scalar ORF, as given by 
-    Eq. (41) of https://arxiv.org/pdf/0903.0528.pdf
[…]
