+++ pygwb/delta_sigma_cut.py
@@ -46,28 +46,29 @@ def dsc_cut(
     >>> print(dsigma)
     
     Parameters
-    ==========
-    naive_sigma: array_like
-        Naive sigma
+    =======
 
-    slide_sigma: array_like
-        Sliding sigma
+    naive_sigma: ``array_like``
+        Array containing the naive sigmas.
 
-    dsc: float, optional
+    slide_sigma: ``array_like``
+        Array containing the sliding sigmas.
+
+    dsc: ``float``, optional
         Threshold used for the delta sigma cut. Default is 0.2.
 
-    bf_ss: float, optional
+    bf_ss: ``float``, optional
         Bias factor for sliding sigmas. Default is 1.
 
-    bf_ns: float, optional
+    bf_ns: ``float``, optional
         Bias factor for naive sigmas. Default is 1.
 
     Returns
     =======
-    dsigma >= dsc: bool
+    dsigma >= dsc: ``bool``
         Mask containing bools indicating whether the segment's delta sigma exceeds the threshold value or not. True indicates that the 
         corresponding GPS times were bad, whereas False denotes good GPS times.
-    dsigma: array_like
+    dsigma: ``array_like``
         Values of the difference between sliding sigma and naive sigma, i.e., the actual value of the delta sigma per segment.
     
     """
@@ -99,58 +100,60 @@ def run_dsc(
     Function that runs the delta sigma cut
 
     Parameters
-    ==========
-    dsc: float
+    =======
+
+    dsc: ``float``
         Threshold used for the delta sigma cut.
 
-    segment_duration: int
+    segment_duration: ``int``
         Duration of each segment.
 
-    psd1_naive, psd2_naive: array_like
+    psd1_naive, psd2_naive: ``array_like``
         An FFTgram of the PSD computed naively, as in in the particular bin J for detector 1 and 2.
 
-    psd1_slide, psd2_slide: array_like
+    psd1_slide, psd2_slide: ``array_like``
         An FFTgram of the PSD computed by considering the noise in adjacent bins to the bin J, i.e. J-1, J+1 for
         detectors 1 and 2.
 
-    alphas: array_like
+    alphas: ``array_like``
         The spectral indices to use. The bad GPS times from all alphas are combined at the end of this code.
 
-    sample_rate: int
+    sample_rate: ``int``
         Sampling rate (Hz)
 
-    notch_list_path: str
+    notch_list_path: ``str``
         Path to the file containing the frequency notches to apply.
         
-    orf: array_like
+    orf: ``array_like``
         The overlap reduction function as a function of frequency that quantifies the overlap of a detector baseline,
         which depends on the detector locations, relative orientations, etc.
 
-    fref: int
+    fref: ``int``
         Reference frequency (Hz).
 
-    window_fftgram_dict: dictionary, optional
+    window_fftgram_dict: ``dictionary``, optional
         Dictionary with window characteristics used in the computation of the sigmas, given the PSD. Default is 
         `(window_fftgram_dict={"window_fftgram": "hann"}`.
 
-    frequency_mask: array_like, optional
+    frequency_mask: ``array_like``, optional
         Frequency mask to apply when computing the sigmas. Default is `True`.
     
-    overlap_factor: float, optional
+    overlap_factor: ``float``, optional
         Overlap factor to use when computing the sigmas, given the PSD. Default is 0.5.
         
-    N_average_segments_psd: int, optional
+    N_average_segments_psd: ``int``, optional
         Number of segments to use during Welch averaging. Used in the computation of the bias factors. Default is 2.
 
-    return_naive_and_averaged_sigmas: bool
+    return_naive_and_averaged_sigmas: ``bool``
         Option to return the naive and sliding sigmas. Default is `False`.
 
[…]

[…]
