+++ pygwb/baseline.py
@@ -83,33 +83,34 @@ class Baseline(object):
 
         name: ``str``
             Name for the baseline, e.g H1H2
-        interferometer_1/2: bilby Interferometer object
-            The two detectors spanning the baseline
+        interferometer_1/2: ``bilby.Interferometer`` object
+            The two detectors spanning the baseline.
         duration: ``float``, optional
             The duration in seconds of each data segment in the interferometers.
             None by default, in which case duration is inherited from the interferometers.
         frequencies: ``array_like``, optional
             The frequency array for the Baseline and
-            interferometers
+            interferometers.
         calibration_epsilon: ``float``, optional
-            Calibration uncertainty for this baseline
+            Calibration uncertainty for this baseline.
         notch_list_path: ``str``, optional
-            File path of the baseline notch list
-        coarse_grain_psd: bool
+            File path of the baseline notch list.
+        coarse_grain_psd: ``bool``, optional
             Whether to apply coarse graining to obtain PSD spectra. Default is False.
-        coarse_grain_csd: bool
+        coarse_grain_csd: ``bool``, optional
             Whether to apply coarse graining to obtain CSD spectra. Default is True.
-        overlap_factor_welch: float, optional
-            Overlap factor to use when if using Welch's method to estimate spectra (NOT coarsegraining). For \"hann\" window use 0.5 overlap_factor and for \"boxcar"\ window use 0 overlap_factor. Default is 0.5 (50% overlap), which is optimal when using Welch's method with a \"hann\" window.
- 
-        overlap_factor: float, optional
+        overlap_factor_welch: ``float``, optional
+            Overlap factor to use when if using Welch's method to estimate spectra (NOT coarsegraining). 
+            For \"hann\" window use 0.5 overlap_factor and for \"boxcar"\ window use 0 overlap_factor. Default is 0.5 (50% overlap), 
+            which is optimal when using Welch's method with a \"hann\" window.
+        overlap_factor: ``float``, optional
             Factor by which to overlap the segments in the psd and csd estimation.
             Default is 1/2, if set to 0 no overlap is performed.
-        window_fftgram_dict: dictionary, optional
+        window_fftgram_dict: ``dict``, optional
             Dictionary containing name and parameters describing which window to use when producing fftgrams for csds (and psds if these are coarse-grained). Default is \"hann\".
-        window_fftgram_dict_welch: dictionary, optional
+        window_fftgram_dict_welch: ``dict``, optional
             Dictionary containing name and parameters describing which window to use when producing fftgrams with Welch's method. Default is \"hann\".
-        N_average_segments_psd: int, optional
+        N_average_segments_psd: ``int``, optional
             Number of segments used for PSD averaging (from both sides of the segment of interest)
             N_avg_segs should be even and >= 2.
         """
@@ -252,6 +253,7 @@ class Baseline(object):
         notch_list_path: ``str``, optional
             Path to notch list to apply to frequency array. If not 
             provided, no notching will be applied at this point.
+
         apply_notches: ``bool``, optional
             If True, apply frequency notches. Default is True.
 
@@ -300,7 +302,15 @@ class Baseline(object):
     @duration.setter
     def duration(self, dur):
         """
-        Sets the duration for the Baseline and interferometers
+        Sets the duration for the Baseline and interferometers.
+
+        Parameters
+        =======
+        dur: ``float``
+            The duration to set for the Baseline and interferometers.
+
+        Notes
+        -----
 
         If `duration` is passed, check that it matches the `duration`
         in the interferometers, if present.
@@ -312,11 +322,6 @@ class Baseline(object):
         interferometer from that.
[…]
