+++ docs/index.rst
@@ -64,6 +64,7 @@ API:
    generate_stochastic_background_in_network
    inject_simulated_data_in_network
    simulate_CBC_GWB
+   test_orfs
    make_notchlist
    run_statistical_checks
    run_pe
