+++ .gitlab-ci.yml
@@ -127,7 +127,6 @@ documentation:
     - cp ../tutorials/*.ipynb ./
     - cp ../tutorials/*.dat ./
     - cp -r ../tutorials/input ./
-    - rm integrationTest*.ipynb
     - make clean
     - make html
     - cd -
