+++ pygwb/statistical_checks.py
@@ -1,6 +1,3 @@
-import json
-import warnings
-
 """
 The statistical checks class performs various tests by plotting different quantities and saving this plots. 
 This allows the user to check for consistency with expected results. Concretely, the following tests and plots
@@ -9,8 +6,10 @@ part of point estimate integrand, FFT of the point estimate integrand, (cumulati
 and sigma as a function of time, omega and sigma distribution, KS test, and a linear trend analysis of omega in time. 
 Furthermore, part of these plots compares the values of these quantities before and after the delta sigma cut. Each of 
 these plots can be made by calling the relevant class method (e.g. `plot_running_point_estimate()`).
-
 """
+
+import json
+import warnings
 
 from os import listdir
 from os.path import isfile, join
@@ -77,45 +76,64 @@ class StatisticalChecks(object):
         Instantiate a StatisticalChecks object.
 
         Parameters
-        ==========
+        =======
 
-        sliding_times_all: array_like
+        sliding_times_all: ``array_like``
             Array of GPS times before the bad GPS times from the delta sigma cut are applied.
-        sliding_omega_all: array_like
-            Array of sliding omegas before the bad GPS times from the delta sigma cut are applied.
-        sliding_sigmas_all: array_like
-            Array of sliding sigmas before the bad GPS times from the delta sigma cut are applied.
-        naive_sigmas_all: array_like
-            Array of naive sigmas before the bad GPS times from the delta sigma cut are applied.
-        coherence_spectrum: array
-            Array containing a coherence spectrum. Each entry in this array corresponds to the 2-detector coherence spectrum evaluated at the corresponding frequency in the frequencies array.
-        coherence_n_segs: int
-            Number of segments used for coherence calculation.
-        point_estimate_spectrum: array
-            Array containing the point estimate spectrum. Each entry in this array corresponds to the point estimate spectrum evaluated at the corresponding frequency in the frequencies array.
-        sigma_spectrum: array
-            Array containing the sigma spectrum. Each entry in this array corresponds to the sigma spectrum evaluated at the corresponding frequency in the frequencies array.
-        frequencies: array
-            Array containing the frequencies.
-        badGPStimes: array_like
-            Array of bad GPS times, i.e. times that do not pass the delta sigma cut.
-        delta_sigmas: array_like
-            Array containing the value of delta sigma for all times in sliding_times_all.
-        plot_dir: str
-            String with the path to which the output of the statistical checks (various plots) will be saved.
-        baseline_name: str
-            Name of the baseline under consideration.
-        param_file: str
-            String with path to the file containing the parameters that were used for the analysis run.
-        frequency_mask: array
-            Boolean mask applied to the specrtra in broad-band analyses. 
-        gates_ifo1/gates_ifo2: list
-            List of gates applied to interferometer 1/2.
-        file_tag: str
-            Tag to be used in file naming convention.
-        legend_fontsize: int
-            Font size for plot legends. Default is 16. All other fonts are scaled to this font.
 
+        sliding_omega_all: ``array_like``
+            Array of sliding omegas before the bad GPS times from the delta sigma cut are applied.
+
+        sliding_sigmas_all: ``array_like``
+            Array of sliding sigmas before the bad GPS times from the delta sigma cut are applied.
+
+        naive_sigmas_all: ``array_like``
+            Array of naive sigmas before the bad GPS times from the delta sigma cut are applied.
+
+        coherence_spectrum: ``array_like``
+            Array containing a coherence spectrum. Each entry in this array corresponds to the 2-detector coherence spectrum evaluated at the corresponding frequency in the frequencies array.
+        
[…]
