### 2023-07-06 e57b4b5
Adding simulator intro page to contents of docs
