+++ docs/simulator.rst
@@ -1,0 +1,242 @@
+============
+Simulating data with ``pygwb.simulator``
+============
+
+**1. Simulating a stochastic gravitational-wave background**
+=========
+
+**1.1 Injecting a power spectrum in random LIGO noise**
+---------
+
+For concreteness, we consider a broken power-law signal PSD to be injected in random LIGO Gaussian noise.  
+The ``simulator`` module takes a ``gwpy.FrequencySeries`` as input for the signal PSD to be injected. 
+We start by building a custom input signal by defining an ``IntensityGW`` function, which outputs the 
+desired signal PSD in the form of a ``gwpy.FrequencySeries``.
+
+.. code-block:: python
+
+    frequencies_x = np.linspace(0, 1000, 10000)
+
+    alpha1 = 6
+    alpha2 = 0
+    fref = 10
+    omegaRef = 5.e-5
+
+    def IntensityGW(freqs, omegaRef, alpha1, fref, alpha2 = 2/3):
+        ''' GW Intensity function from broken power law in OmegaGW
+    
+        Parameters
+        ----------
+
+        freqs: np.array
+            frequency array
+        fref: 
+            reference frequency
+        omegaRef: 
+            Value of OmegaGW at reference frequency
+        alpha1:
+            first spectral index
+        alpha2:
+            second spectral index
+        
+        Return
+        ------
+        FrequencySeries
+        '''
+        from pygwb.constants import H0
+        H_theor = (3 * H0.si.value ** 2) / (10 * np.pi ** 2)
+        
+        fknee = fref
+        
+        power = np.zeros_like(freqs)
+        
+        power[freqs<fknee] = H_theor * omegaRef * (freqs[freqs<fknee]) ** (alpha1 -3) * fref**(-alpha1)
+        power[freqs>fknee] = H_theor * omegaRef * (freqs[freqs>fknee]) ** (alpha2 - 3) * fref**(-alpha2)
+        power[freqs==fknee] = H_theor * omegaRef * (fknee) ** (alpha2 -3) * fref**(-alpha2)
+        
+        power[0] = power[1]
+        
+        return gwpy.frequencyseries.FrequencySeries(power, frequencies=freqs)
+
+    Intensity_GW_inject = IntensityGW(frequencies_x, omegaRef = omegaRef, alpha1 = alpha1, fref = fref)
+
+Note that the above signal PSD was chosen for illustrative purposes. However, in practice, any 
+signal PSD can be chosen to be injected, and one does not need to restrict themselves to a broken power-law.
+
+One also needs to specify the parameters that will serve as input to the ``simulator``. Concretely, we specify 
+the duration of each simulated segment, the number of segments, and the sampling frequency.
+   
+.. code-block:: python
+
+    duration = 60 # duration of each segment of data (s)
+    N_segs = 10  # number of data segments to generate
+    sampling_frequency = 1024 # Hz
+
+The detectors for which data with the above signal PSD need to be simulated, have to be passed 
+to the ``simulator`` module. By relying on the ``detector`` module, we instantiate various detectors below.  
+In addition, we note that these detectors are ``Interferometer`` objects, but are based on ``bilby`` detectors, 
+which have default noise PSDs saved in them, in the ``power_spectral_density`` attribute of the ``bilby`` detector. 
+Below, we load in this noise PSD and make sure the duration and sampling frequency of the detector is set to the desired value of 
+these parameters.
+
+.. code-block:: python
+
+    H1 = Interferometer.get_empty_interferometer("H1")
+    L1 = Interferometer.get_empty_interferometer("L1")
+
+    ifo_list = [H1, L1, V1, K, G]
+
+    for ifo in ifo_list:
+        ifo.duration = duration
+        ifo.sampling_frequency = sampling_frequency
[…]
