### 2023-07-17 a-renzini/pygwb#1 @a-renzini
Hello @Sbozzolo and @cmbiwer - apologies on the delay. I'm making a list here of the things we _have addressed_ which you can take a look at, as we continue working on documentation+. I remember I had a few things to improve in the unit tests as well, I'm working on it!

READY (ish)
- About page
- Installation page
- All the API pages

We're working on restructuring the docs, essentially turning some of the tutorials into actual pages and leaving some remaining tutorials as "extras". The idea is that the API pages present each module/object, but we will add pages that describe how these are then used for analysis, and clarify the objectives/deliverables of the package, as well as how to obtain them. We will probably add about one page per script present in the `pygwb_pipe` folder (these are the scripts used in our day to day analyses)
