### 2023-07-18 e092d3b
fixing bug in implementation of notching in statistical checks coherence histogram plots
