+++ pygwb/detector.py
@@ -17,45 +17,38 @@ from .spectral import before_after_average, power_spectral_density
 class Interferometer(bilby.gw.detector.Interferometer):
 
     """
-    Subclass of bilby's Interferometer class which is charged with handling, storing and saving all relevant interferometer data.
+    Subclass of bilby's Interferometer class which is charged with handling, storing and
+    saving all relevant interferometer data.
     It handles all data analysis parts relating to the individual detectors in a baseline of a network.
     An example would be loading in data from a certain channel and computing the psd of the detector.
     
     Examples
     --------
     
-    In this example we will load in data from the publicly available GWOSC servers.
-    We will gate this data, compute the PSD of the detector and the average PSD of the detector object.
-    This gives a brief overview of the most critical capabilities of the pygwb detector module.
+    In this example, we will load in data from the publicly available GWOSC servers using the detector module.
+    We will gate the data, compute the PSD and the average PSD of the detector object.
+    This example gives a brief overview of the most critical capabilities of the pygwb detector module.
     We start by importing the Interferometer class from pygwb.
     
     >>> from pygwb.detector import Interferometer
 
     To load in some data, first we make an empty detector object.
-    Based on the name of the object, in our case "H1", the module will make an Interferometer object
-    that has no data, hence empty, but with all correct parameters, like the location vertex of the detector
-    it represents. The name can be any one of the detectors supported in ``bilby.gw.detector``, 
+    Based on the name of the object, the module will make an Interferometer object that has no data.
+    The name can be any one of the detectors supported in ``bilby.gw.detector``,
     the parent class of our Interferometer class.
     
     >>> ifo_1 = Interferometer.get_empty_interferometer("H1")
     
     Then, we load in data using one of the provided ``set_timeseries_from`` functions.
-    To make life easy, we will look at public data where we only need a channel name.
-    We take a start time t0 and an end time tf. We want to use public data so we 
-    set data_type to be public. We are working with H1 detector and we would like to have normal strain
-    data, hence we use the channel "H1:GWOSC-4KHZ_R1_STRAIN". All the other parameters are taken to be
-    default values which are used in the most general pygwb pipeline analysis, see pygwb paper.
-    
-    >>> t0 = 1247644138
-    >>> tf = 1247648138
-    >>> data_type = "public"
-    >>> channel = "H1:GWOSC-4KHZ_R1_STRAIN"
-    >>> 
+    We take a start time t0 and an end time tf. We want to use public data, so we
+    set data_type to public. We use the channel "H1:GWOSC-4KHZ_R1_STRAIN". All the other parameters are taken to be
+    default values.
+
     >>> ifo_1.set_timeseries_from_channel_name(
-        channel,
-        t0=t0,
-        tf=tf,
-        data_type=data_type,
+        "H1:GWOSC-4KHZ_R1_STRAIN",
+        t0=1247644138,
+        tf=1247648138,
+        data_type="public",
         local_data_path = "",
         new_sample_rate=4096,
         input_sample_rate=4096,
@@ -67,26 +60,16 @@ class Interferometer(bilby.gw.detector.Interferometer):
         time_shift=0,
     )
     
-    Now, imagine we would like to gate the detector data.
-    In gating, windows are put on top of the timeseries data to remove
-    any extra loud glitches that would influence the analysis.
-    In that case, we can call
-    
-    >>> gate_tzero = 1.0
-    >>> gate_tpad = 0.5
-    >>> gate_threshold = 50.0
-    >>> cluster_window = 0.5
-    >>> gate_whiten = True
+    Noww, we gate the detector data. In that case, we can call
+
     >>> ifo_1.gate_data_apply(
-        gate_tzero=gate_tzero,
-        gate_tpad=gate_tpad,
-        gate_threshold=gate_threshold,
[…]
