+++ pygwb/preprocessing.py
@@ -1,39 +1,36 @@
-"""Preprocessing module collects all the functions that handle the preprocessing (what's in a name) of the data used in the analysis.
+"""Preprocessing module collects all the functions that handle the preprocessing of the data used in the analysis.
 
-The functions of the preprocessing module can handle anything related with preparing the data for the pygwb analysis run.
-Functions are present that read data from frame files, locally or publicly. Other functions are able to resample the data, apply a high-pass filter to it or apply a timeshift to the data.
-These functionalities comes together in the triplet of ``preprocessing_data`` functions which read in data and resample and/or high-pass it immediately.
-The triplet can work from a given ``gwpy.timeseries.TimeSeries``, a normal array or using a gravitational wave channel that will read in that channel from the provided local or public frame files. 
+The functions of the preprocessing module handle anything related with preparing the data for the pygwb analysis run.
+It contains the ability to read data from frame files, locally or publicly.
+Other functions are able to resample the data, apply a high-pass filter to data or apply a timeshift.
+These functionalities come together in the triplet of ``preprocessing_data`` functions
+which reads in data and resamples and/or high-passes it immediately.
+The triplet can work for a ``gwpy.timeseries.TimeSeries``, a normal array or using a gravitational wave channel
+that will read data from that channel using the provided local or public frame files.
 
-The last function can gate the data based on the gate function in gwpy, ``gwpy.timeseries.TimeSeries.gate``.
+The last function of the module can gate data based on the gate function in gwpy, ``gwpy.timeseries.TimeSeries.gate``.
 
 Examples
 --------
 
-As an example, we will read in some data from a certain channel and then resample, high-pass and apply gating to the data.
+As an example, we read in some data from a certain channel and then resample, high-pass and apply gating to the data.
 First, we have to import the module.
 
 >>> import pygwb.preprocessing as ppp
 
 Then, we read in some data utilising the ``read_data`` function of preprocessing.
 We focus on reading in public data out of convenience from the "H1" detector.
-We want to read strain data hence the channel name.
+We want to read strain data hence the choice for the channel name.
 
 >>> IFO = "H1"
->>> data_type = "public"
->>> channel = "H1:GWOSC-16KHZ_R1_STRAIN"
->>> t0 = 1247644138
->>> tf = 1247648138
->>> local_data_path = ""
->>> input_sample_rate = 16384
 >>> data_timeseries = ppp.read_data(
         IFO,
-        data_type,
-        channel,
-        t0,
-        tf,
-        local_data_path,
-        input_sample_rate,
+        "public",                   # data_type
+        "H1:GWOSC-16KHZ_R1_STRAIN", # channel
+        1247644138,                 # t0
+        1247648138,                 # tf
+        "",                         # local_data_path
+        16384                       # input_sample_rate
     )
 >>> print(data_timeseries.sample_rate)
 16384.0 Hz
@@ -41,23 +38,17 @@ We want to read strain data hence the channel name.
 We print the sample rate just to show it is indeed 16 kHz at the moment.
 Now, we will preprocess the data, meaning it will be resampled and a high-pass
 filter will be applied to the data. We want to resample to 4 kHz.
-All the other values for parameters are default values as ever.
 
 >>> new_sample_rate = 4096
->>> cutoff_frequency = 11
->>> number_cropped_seconds = 2
[…]
