+++ pygwb/parameters.py
@@ -1,3 +1,12 @@
+"""Parameters module contains the Parameters class which is used to store and handle all parameters of pygwb.
+
+This module contains only one class, the Parameters class and its accompanying helper class ParametersHelp.
+An object of the Parameters class is utilised as the storage object of all parameters in the pygwb analysis.
+It contains all the required ingredients for performing the whole isotropic stochastic LVK search.
+
+The ParametersHelp class is not meant to be interacted with as a user.
+
+"""
 import argparse
 import configparser
 import enum
@@ -12,91 +21,98 @@ import json5
 @dataclass
 class Parameters:
     """
-    Parameters class: a dataclass which contains all parameters required for initialising a pygwb Interferometer, a pygwb Baseline, and run pygwb_pipe.
+    Parameters class: a dataclass which contains all parameters required for initialising a pygwb Interferometer,
+    a pygwb Baseline, and run pygwb_pipe.
 
     Attributes
-    --------
-    t0 : `float`
+    =======
+    t0 : ``float``
         Initial time.
-    tf: `float`
+    tf: ``float``
         Final time.
-    data_type: `str`
+    data_type: ``str``
         Type of data to access/download; options are private, public, local. Default is public.
-    channel: `str`
+    channel: ``str``
         Channel name; needs to match an existing channel. Default is \"GWOSC-16KHZ_R1_STRAIN\".
-    frametype: `str`
-        Frame type; Optional, desired channel needs to be found in listed frametype. Only used when data_type=private. Default is empty.
-    new_sample_rate: `int`
+    frametype: ``str``
+        Frame type; Optional, desired channel needs to be found in listed frametype. Only used when data_type=private.
+        Default is empty.
+    new_sample_rate: ``int``
         Sample rate to use when downsampling the data (Hz). Default is 4096 Hz.
-    input_sample_rate: `int`
+    input_sample_rate: ``int``
         Sample rate of the read data (Hz). Default is 16384 Hz.
-    cutoff_frequency: `int`
+    cutoff_frequency: ``int``
         Lower frequency cutoff; applied in filtering in preprocessing (Hz). Default is 11 Hz.
-    segment_duration: `int`
+    segment_duration: ``int``
         Duration of the individual segments to analyse (seconds). Default is 192 seconds.
-    number_cropped_seconds: `int`
+    number_cropped_seconds: ``int``
         Number of seconds to crop at the start and end of the analysed data (seconds). Default is 2 seconds.
-    window_downsampling: `str`
+    window_downsampling: ``str``
         Type of window to use in preprocessing. Default is \"hamming\".
-    ftype: `str`
+    ftype: ``str``
         Type of filter to use in downsampling. Default is \"fir\".
-    frequency_resolution: `float`
+    frequency_resolution: ``float``
         Frequency resolution of the final output spectrum (Hz). Default is 1\/32 Hz.
-    polarization: `str`
-        Polarisation type for the overlap reduction function calculation; options are scalar, vector, tensor. Default is tensor.
-    alpha: `float`
+    polarization: ``str``
+        Polarization type for the overlap reduction function calculation; options are scalar, vector, tensor.
+        Default is tensor.
+    alpha: ``float``
         Spectral index to filter the data for. Default is 0.
-    fref: `int`
+    fref: ``int``
         Reference frequency to filter the data at (Hz). Default is 25 Hz.
-    flow: `int`
+    flow: ``int``
         Lower frequency to include in the analysis (Hz). Default is 20 Hz.
-    fhigh: `int`
+    fhigh: ``int``
         Higher frequency to include in the analysis (Hz). Default is 1726 Hz.
-    interferometer_list: `list`
+    interferometer_list: ``list``
[…]
