### 2023-08-03 f49170d
Editing combine script such that it can iterate over combined files.
