### 2023-08-04 f257e3a
adding CBC background calculation function
