+++ .gitlab-ci.yml
@@ -111,6 +111,22 @@ python-3.10:
   <<: *test-python
   image: python:3.10
 
+workflow:
+  stage: test
+  image: python3.10
+  before_script:
+    - wget -qO - https://research.cs.wisc.edu/htcondor/ubuntu/HTCondor-Release.gpg.key | sudo apt-key add -
+    - echo "deb http://research.cs.wisc.edu/htcondor/ubuntu/8.9/focal focal contrib" | sudo tee -a /etc/apt/sources.list
+    - echo "deb-src http://research.cs.wisc.edu/htcondor/ubuntu/8.9/focal focal contrib" | sudo tee -a /etc/apt/sources.list
+    - sudo apt-get -o Acquire::Retries=3 update
+    - sudo apt-get -o Acquire::Retries=3 install minihtcondor
+    - sudo systemctl start condor
+    - sudo systemctl enable condor 
+    - pip install --upgrade pip setuptools
+  script:
+    - pip install .
+    - condor_status
+
 documentation:
   stage: documentation
   image: python:3.8
