### 2023-08-09 1ce08a1
Fix image typo
