### 2023-08-09 04f4b76
Switch condor install method
