### 2023-08-09 8933576
Add simple public workflow ini
