+++ pygwb_pipe/simple_workflow_config_public.ini
@@ -1,0 +1,52 @@
+[general]
+accounting_group = ligo.dev.o4.detchar.explore.test
+; times around GW170817
+t0 = 1187008682
+tf = 1187008982
+ifos = H1 L1
+
+[executables]
+pygwb_pipe = pygwb_pipe
+pygwb_combine = pygwb_combine
+pygwb_stats = pygwb_stats
+pygwb_html = pygwb_html
+
+[pygwb_pipe]
+interferometer_list = ${general:ifos}
+data_type = public
+channel = GWOSC-16KHZ_R1_STRAIN
+time_shift = 0
+new_sample_rate = 256
+input_sample_rate = 16384
+cutoff_frequency = 11
+segment_duration = 16
+number_cropped_seconds = 2
+window_downsampling = hamming
+ftype = fir
+window_fftgram = hann
+frequency_resolution = 0.03125
+n_average_segments_welch_psd = 2
+coarse_grain = 0
+overlap_factor = 0.5
+zeropad_csd = True
+polarization = tensor
+alpha = 0.0
+fref = 25
+flow = 20
+fhigh = 100
+;notch_list_path = /home/arianna.renzini/PROJECTS/pygwb/test/test_data/Official_O3_HL_notchlist.txt
+calibration_epsilon = 0
+alphas_delta_sigma_cut = -5 0 3
+delta_sigma_cut = 0.2
+return_naive_and_averaged_sigmas = False
+save_data_type = npz
+
+[pygwb_combine]
+alpha = ${pygwb_pipe:alpha}
+fref = ${pygwb_pipe:fref}
+h0 = 0.7
+
+[pygwb_stats]
+
+[pygwb_html]
+
