+++ .gitlab-ci.yml
@@ -116,13 +116,7 @@ workflow:
   image: python:3.10
   before_script:
     - pip install --upgrade pip setuptools
-    - curl -fsSL https://get.htcondor.org | /bin/bash -s -- --download
-    - tar -x -f condor.tar.gz
-    - mv condor-*stripped condor
-    - cd condor
-    - ./bin/make-personal-from-tarball
-    - . ~/condor/condor.sh
-    - condor_master
+    - curl -fsSL https://get.htcondor.org | /bin/bash -s -- --no-dry-run
   script:
     - pip install .
     - condor_status
