### 2023-08-09 513cbdb
update simple workflow ini
