+++ pygwb_pipe/simple_workflow_config_public.ini
@@ -25,8 +25,9 @@ window_downsampling = hamming
 ftype = fir
 window_fftgram = hann
 frequency_resolution = 0.03125
-n_average_segments_welch_psd = 2
-coarse_grain = 0
+n_average_segments_psd = 2
+coarse_grain_psd = True
+coarse_grain_csd = True
 overlap_factor = 0.5
 zeropad_csd = True
 polarization = tensor
