### 2023-08-09 23ae3f7
run workflow ci locally instead of condor
