### 2023-08-09 70d76b4
additional updates to simple config
