+++ pygwb_pipe/simple_workflow_config_public.ini
@@ -24,7 +24,7 @@ number_cropped_seconds = 2
 window_downsampling = hamming
 ftype = fir
 window_fftgram = hann
-frequency_resolution = 0.03125
+frequency_resolution = 1.
 n_average_segments_psd = 2
 coarse_grain_psd = True
 coarse_grain_csd = True
@@ -41,6 +41,7 @@ alphas_delta_sigma_cut = -5 0 3
 delta_sigma_cut = 0.2
 return_naive_and_averaged_sigmas = False
 save_data_type = npz
+calc_coh = True
 
 [pygwb_combine]
 alpha = ${pygwb_pipe:alpha}
