### 2023-08-09 e6827ab
Remove extra use of np.int
