### 2023-08-09 df74760
Added function to calculate the number of injections.
