### 2023-08-09 983b254
Added function to calculate the number of injections.
