+++ pygwb_pipe/workflow_config.ini
@@ -48,7 +48,6 @@ l1 = /home/derek.davis/public_html/detchar/preO4/stochmon_update/test_pygwb_work
 [pygwb_combine]
 alpha = ${pygwb_pipe:alpha}
 fref = ${pygwb_pipe:fref}
-h0 = 0.7
 
 [pygwb_stats]
 
