+++ .gitlab-ci.yml
@@ -119,7 +119,7 @@ workflow:
   script:
     - pip install .
     - mkdir workflow_result
-    - pygwb_create_isotropic_workflow --configfile ./pygwb_pipe/workflow_config.ini --basedir ./workflow_result/ --run_locally
+    - pygwb_create_isotropic_workflow --configfile ./test/test_data/config_workflow_test.ini --basedir ./workflow_result/ --run_locally
 
 documentation:
   stage: documentation
