+++ test/test_data/config_workflow_test.ini
@@ -1,11 +1,9 @@
 [general]
-accounting_group = ligo.dev.o4.sgwb.isotropic.stochastic
+accounting_group = ligo.dev.o4.detchar.explore.test
+; times around GW170817
+t0 = 1187008682
+tf = 1187008982
 ifos = H1 L1
-plot_segment_results = False
-max_job_dur = 5000
-min_job_dur = 900
-t0 = 1368921618
-tf = 1374883218
 
 [executables]
 pygwb_pipe = pygwb_pipe
@@ -13,48 +11,42 @@ pygwb_combine = pygwb_combine
 pygwb_stats = pygwb_stats
 pygwb_html = pygwb_html
 
-[data_quality]
-science_segment = DMT-ANALYSIS_READY
-veto_definer = /home/arianna.renzini/public_html/ER15_pygwb_run/old_setup/H1L1-HOFT_C00_O4_CBC.xml
-
 [pygwb_pipe]
 interferometer_list = ${general:ifos}
-data_type = private
-channel = GDS-CALIB_STRAIN_CLEAN
-frametype = H1:H1_HOFT_C00,L1:L1_HOFT_C00
-new_sample_rate = 4096
+data_type = public
+channel = GWOSC-16KHZ_R1_STRAIN
+time_shift = 0
+new_sample_rate = 256
 input_sample_rate = 16384
-frequency_resolution = 0.03125
 cutoff_frequency = 11
-segment_duration = 192
+segment_duration = 16
 number_cropped_seconds = 2
 window_downsampling = hamming
 ftype = fir
 window_fftgram = hann
+frequency_resolution = 1.
 n_average_segments_psd = 2
-coarse_grain_psd = False
+coarse_grain_psd = True
 coarse_grain_csd = True
-overlap_factor_welch = 0.5
 overlap_factor = 0.5
+zeropad_csd = True
 polarization = tensor
 alpha = 0.0
 fref = 25
 flow = 20
-fhigh = 1726
-notch_list_path = /home/shivaraj.kandhasamy/stochastic/O4/detchar/O4_notch_lines.txt
+fhigh = 100
+;notch_list_path = /home/arianna.renzini/PROJECTS/pygwb/test/test_data/Official_O3_HL_notchlist.txt
 calibration_epsilon = 0
 alphas_delta_sigma_cut = -5 0 3
 delta_sigma_cut = 0.2
 return_naive_and_averaged_sigmas = False
 save_data_type = npz
 calc_coh = True
-gate_data = True
-apply_dsc = True
 
 [pygwb_combine]
 alpha = ${pygwb_pipe:alpha}
 fref = ${pygwb_pipe:fref}
-combine_coherence = True
+h0 = 0.7
 
 [pygwb_stats]
 
