### 2023-08-11 ad4be29
Add combine-combine step to isotropic workflow
