+++ pygwb/workflow.py
@@ -26,6 +26,10 @@ ACCOUNTING_GROUP_USER = os.getenv(
     getuser(),
 )
 
+def _split(orig_list, N):
+    k, m = divmod(len(orig_list), N)
+    return list(orig_list[i*k+min(i, m):(i+1)*k+min(i+1, m)] for i in range(N))
+
 def makedir(dir_path):
     if not os.path.exists(dir_path):
         os.makedirs(dir_path)
@@ -435,12 +439,15 @@ class IsotropicWorkflow(Workflow):
     
     # FIX THE JOB LIST
     def create_pygwb_combine_job(self, t0=None, tf=None, parents=[], input_file=None, output_path=None,
-        data_path=None, coherence_path=None, param_file=None,
-        alpha=0., fref=30, h0=0.7):
+        data_path=None, coherence_path=None, dsc_path=None, param_file=None,
+        alpha=0., fref=30, h0=0.7, file_tag_extra=None):
     #pygwb_combine --data_path output/ --param_file output/parameters_final.ini --alpha 0 --fref 30 --h0 0.7 --out_path ./output/
         dur = int(tf-t0)
         file_tag = f'{int(t0)}-{dur}'
         name = f'pygwb_combine_{int(t0)}_{dur}'
+        if file_tag_extra:
+            file_tag = f'{file_tag}-{file_tag_extra}'
+            name = f'{name}_{file_tag_extra}'
         args = _collect_job_arguments(self.config, 'pygwb_combine')
         if isinstance(data_path, str):
             data_path = [data_path]
@@ -448,7 +455,7 @@ class IsotropicWorkflow(Workflow):
             coherence_path = [coherence_path]
     
         input_file = data_path + coherence_path + [param_file]
-    
+
         data_path = ' '.join(data_path)
         coherence_path = ' '.join(coherence_path)
     
@@ -459,6 +466,17 @@ class IsotropicWorkflow(Workflow):
             '--coherence_path', coherence_path,
             '--file_tag', file_tag
             ]
+
+        if dsc_path:
+            # only provided for the "combine combine" jobs
+            if isinstance(dsc_path, str):
+                dsc_path = [dsc_path]
+            input_file = input_file + dsc_path
+            dsc_path = ' '.join(dsc_path)
+            args = args + [
+                '--delta_sigma_path', dsc_path,
+                ]
+
         output_file = [os.path.join(output_path, f'point_estimate_sigma_spectra_alpha_{alpha:.1f}_fref_{int(fref)}_{file_tag}.npz'),
                        os.path.join(output_path, f'coherence_spectrum_{file_tag}.npz'),
                        os.path.join(output_path, f'delta_sigma_cut_{file_tag}.npz')]
@@ -513,6 +531,43 @@ class IsotropicWorkflow(Workflow):
         for parent in parents:
             job.add_parent(parent)
         return job
+
+    def create_pygwb_multi_stage_combine(self, t0=None, tf=None, parents=[], input_file=None, output_path=None,
+        data_path=None, coherence_path=None, dsc_path=None, param_file=None,
+        alpha=0., fref=30, h0=0.7, combine_factor=1):
+
+        # set up individual combines
+        assert len(data_path) == len(coherence_path)
+        data_path_lists = _split(data_path, int(combine_factor))
+        coherence_path_lists = _split(coherence_path, int(combine_factor))
+
+        combined_data_path = []
+        combined_coherence_path = []
+        combined_dsc_path = []
+        combined_jobs = []
+
+        for i in range(int(combine_factor)):
+            extra_tag = f'PART{int(i)}'
+            combine_job, combine_output = self.create_pygwb_combine_job(
+                t0=self.t0, tf=self.tf,
+                parents=parents, output_path=output_path,
[…]
