### 2023-08-17 a-renzini/pygwb#1 @kevinturbang
Hello @Sbozzolo and @cmbiwer,

We have made significant progress in updating and refining our code and the documentation. Concretely,

- All module API pages have been updated, ensuring that all class methods contain a description, and that the style used across module API pages is consistent throughout. Each of the API “landing pages" contains a brief description of the module itself, often with a few lines of code to illustrate the basic functionalities of the module.
- The about, contribution and installation pages have been updated.
- A few additional pages were created as well, and will be added moving forward. This to highlight some of the key scripts and functionalities of the pygwb package. The addition of those pages goes together with the removal of some tutorials. The idea behind these changes is to group all general information in these pages, whereas tutorials should contain more details or anything that deviates from the default usage.

At this point of the review process, it would be most useful if you could  provide feedback on the work that has been done so far, while we finalize the documentation. This will allow us to incorporate any proposed changes as we are making these final modifications to the documentation.
