### 2023-08-17 a-renzini/pygwb#1 @Sbozzolo
Hi @kevinturbang.

Thank you for your hard work. I'll review the new documentation over the weekend and provide my feedback.
