### 2023-08-18 a-renzini/pygwb#1 @a-renzini
@Sbozzolo no - we are now working directly in the `gitlab` as suggested: https://pygwb.docs.ligo.org/pygwb/ (I have updated the README in the repo to reflect this)
