### 2023-08-21 a-renzini/pygwb#1 @Sbozzolo
Thanks, @a-renzini. I went through the documentation and here are some comments.

- The diagram in the about page feels too technical detail for an "about" page. I'd suggest moving it somewhere else because it is not useful to readers that are not already familiar with pyGWB.
- I don't see where you specify what versions of python are supported (or recommended).
- There is no clear progression on how users should approach the documentation. Right now, if you follow the "next" buttons in the docs you get: About, Installation, Contributing guide, Using pygwb_pipe: a quickstart manual, ... It would be nice to have a general quickstart. It's unclear how I should use the package. If pygwb_pipe is a central piece, it should be explained.
- Following on the previous point, given that I am unfamiliar with pygwb_pipe, I do not understand what it means to be a pipeline, or what a dag file is, or why I should care about this. Similar to pygwb_pe.
- The tutorials look good. I would recommend backlinking all the various functions/objects you name in the notebooks, so that one can click on the link and learn more about the object/function from the API manual.
- I'd recommend running the tutorials in a sandbox while you generate your documentation instead of uploading an evaluated notebook. Right now, the tutorials come with warnings generated on Keving Turbang's machine. If you run in a sandbox, you can do some additional quality control on your docs/package, and be closer to the environment a new user might have. (Also, it is not great to see unexplanined warnings in tutorials)
- There's a typo "choise -> choice" in the inject a power spectrum tutorial, "cto -> to" in inject SGWB. If you care about it, there's also  a mix of British (e.g., initialise) and American (e.g., visualize) spelling of words. There are also extra new lines.
- Some acronyms are never defined. Most people in the field know what PSD or H1 are, but you can make your package even more accessible to outsiders (or new students!) by defining all the acronyms at least once.
- What is a notch list/notches?
- I'd also recommend ensuring that your plots are readable (e.g., with a large font size, with different line styles for visually impaired people)

I haven't gone through the APIs carefully yet, but they seem fairly comprehensive.

To sum up my current assessment of your documentation: you wrote good pages/API/tutorials, but there is no underlying thread that joins them. As a new user, I still feel lost and I couldn't explain what pyGWB does exactly. I would recommend holding your users's hands a little bit more (e.g., providing a learning progression). I can give you examples of packages with good documentation if you want some inspiration on how to structure yours. One that might be close is [LEGWORK](https://legwork.readthedocs.io/en/latest/index.html).
