+++ docs/citing.rst
@@ -1,0 +1,29 @@
+============
+Citing pygwb
+============
+
+When using `pygwb` in your scientific work, we would like you to cite the `pygwb` package accordingly. `pygwb` was published in ApJ, for which we provide a bibtex entry below. 
+
+  .. code-block:: bash
+
+    @article{Renzini_2023,
+    doi = {10.3847/1538-4357/acd775},
+    
+    url = {https://doi.org/10.3847%2F1538-4357%2Facd775},
+    
+    year = 2023,
+    month = {jul},
+    
+    publisher = {American Astronomical Society},
+    
+    volume = {952},
+    
+    number = {1},
+    
+    pages = {25},
+    
+    author = {Arianna I. Renzini and Alba Romero-Rodr{\'{\i}}guez and Colm Talbot and Max Lalleman and Shivaraj Kandhasamy and Kevin Turbang and Sylvia Biscoveanu and Katarina Martinovic and Patrick Meyers and Leo Tsukada and Kamiel Janssens and Derek Davis and Andrew Matas and Philip Charlton and Guo-Chin Liu and Irina Dvorkin and Sharan Banagiri and Sukanta Bose and Thomas Callister and Federico De Lillo and Luca D\'Onofrio and Fabio Garufi and Gregg Harry and Jessica Lawrence and Vuk Mandic and Adrian Macquet and Ioannis Michaloliakos and Sanjit Mitra and Kiet Pham and Rosa Poggiani and Tania Regimbau and Joseph D. Romano and Nick van Remortel and Haowen Zhong},
+    
+    title = {pygwb: A Python-based Library for Gravitational-wave Background Searches},
+    
+    journal = {The Astrophysical Journal}}
