### 2023-08-21 caf42d8
Removing schema
