+++ docs/api.rst
@@ -1,0 +1,41 @@
+============
+API
+============
+
+The ``pygwb``` package consists of several modules, each with a distinct functionality. A summary of all ``pygwb`` modules and its main external dependencies can be seen in the schema below.
+
+
+.. image:: pygwb_modules.png
+   :width: 800
+
+This API section contains information about each of the modules, as well as a description of all methods therein. The modules can be accessed through the table below, or through the table of contents on the left hand side of the page.
+
+.. currentmodule:: pygwb
+
+.. autosummary::
+   :toctree: api
+   :template: custom-module-template.rst
+   :caption: API:
+   :recursive:
+   
+   detector
+   baseline
+   network
+
+   preprocessing
+   spectral
+   coherence
+   postprocessing
+   omega_spectra
+   pe
+   statistical_checks
+
+   simulator
+
+   delta_sigma_cut
+   notch
+   
+   constants
+   orfs   
+   parameters
+   util
