+++ docs/demos.rst
@@ -1,0 +1,17 @@
+============
+Demos
+============
+      
+.. currentmodule:: pygwb
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Demos:
+
+   generate_stochastic_background_in_network
+   inject_simulated_data_in_network
+   simulate_CBC_GWB
+   test_orfs
+   make_notchlist
+   run_statistical_checks
+   run_pe
