+++ docs/tutorials.rst
@@ -1,0 +1,9 @@
+============
+Tutorials
+============
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Tutorials:
+      pipeline
+      pe
