### 2023-08-21 82c0b20
updated for outlier.txt procedure
