### 2023-08-21 2c9c243
updated outlier function
