+++ docs/index.rst
@@ -13,6 +13,15 @@ pygwb documentation
 
 `pygwb` also includes a set of pre-packaged analysis scripts which may be used to analyse data and perform large-scale searches on a high-performance computing cluster efficiently.
 
+.. raw:: html
+
+   <a href="about.html"><button style="background-color:#307FC1;border-color:#307FC1;color:white;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">More about pygwb</button></a>
+   <a href="installation.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Installing pygwb</button></a>
+   <a href="contributing.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Contributing to pygwb</button></a>
+   <a href="api.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Module API</button></a>
+   <a href="tutorials.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Tutorials</button></a>
+   <a href="demos.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 100.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Demos</button></a>
+
 .. toctree::
    :maxdepth: 1
    :caption: Contents:
