### 2023-08-22 242dc94
Adding logo and reshuffling index page for logic
