+++ docs/demos.rst
@@ -1,13 +1,28 @@
 ============
 Demos
 ============
+
+In addition to the `tutorials <tutorials.html>`_, we give a few more demos to illustrate the capabilities of the ``pygwb``
+package. This differ from the tutorials in the sense that they are more specific and show how to customize some of the ``pygwb``
+functionalities to your own needs.
+
+.. raw:: html
+
+   <a href="Demo1.html"><button style="background-color:#307FC1;border-color:#307FC1;color:white;width: 200.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Demo 1</button></a>
+   <a href="Demo2.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Demo 2</button></a>
+   <a href="Demo3.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 200.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Demo 3</button></a>
+   
+   <br />
+
+.. tip::
+   If you are feeling overwhelmed with these demos, make sure to have a look to the `tutorials <tutorials.html>`_ before.
       
 .. currentmodule:: pygwb
 
 .. toctree::
    :maxdepth: 1
-   :caption: Demos:
-
+   :hidden:
+   
    generate_stochastic_background_in_network
    inject_simulated_data_in_network
    simulate_CBC_GWB
