### 2023-08-22 38bdb96
Fix bug in check_required() so that final result jobs are always rerun
