+++ pygwb/coherence.py
@@ -1,3 +1,15 @@
+"""
+The ``coherence`` is a very small module and only contains one method, which is used to calculate the coherence spectrum given two power spectrograms and a cross-spectrogram.
+
+Examples
+--------
+    
+Given two power specvtrograms and a cross-spectrogram, the coherence can be computed by
+    
+>>> from pygwb.coherence import calculate_coherence
+>>> calculate_coherence(psd_1, psd_2, csd)
+"""
+
 import numpy as np
 
 
@@ -6,13 +18,18 @@ def calculate_coherence(psd_1, psd_2, csd):
     Calculate a coherence spectrum given two power spectrograms and a cross-spectrogram.
 
     Parameters
-    ==========
-    psd_1_spectrum: np.array
+    =======
+    psd_1_spectrum: ``np.array``
         PSD spectrum of first detector.
-    psd_2_spectrum: np.array
+    psd_2_spectrum: ``np.array``
         PSD spectrum of second detector.
-    csd_spectrum: np.array
+    csd_spectrum: ``np.array``
         Spectrum of the cross-spectral density of the two detectors.
+
+    Returns
+    =======
+    coherence: ``np.array``
+        Coherence spectrum
     '''
     #I don't think we need these terms anymore:
     #fftlength = int(1.0/psd_1_spectrogram.df.value)
