+++ docs/pe.rst
@@ -1,56 +1,92 @@
-============
+=======================================
 Parameter estimation using ``pygwb_pe``
-============
+=======================================
+
+The ``pygwb_pe`` script performs parameter estimation (PE) as an integral part of the
+analysis, which naturally follows the computation of the optimal estimate of the gravitational-wave background (GWB) performed in ``pygwb``.
+Parameter estimation consists of inferring parameters from the data, using Bayesian inference. Concretely, one uses the likelihood
+
+.. math::
+
+    p(\hat{C}^{IJ}(f_k) | \mathbf{\Theta}) \propto\exp\left[  -\frac{1}{2} \sum_{IJ}^N \sum_k \left(\frac{\hat{C}^{IJ}(f_k) - \Omega_{\rm M}(f_k|\mathbf{\Theta})}{\sigma^2_{IJ}(f_k)}\right)^2  \right],
+
+where :math:`\Omega_{\rm M}(f_k|\mathbf{\Theta})` is the model being fit to data, and :math:`\mathbf{\Theta}` are the model's parameters.
+This allows to infer, or put constraints on, some GWB model parameters :math:`\mathbf{\Theta}`, given the optimal GWB estimator :math:`\hat{C}^{IJ}(f_k)` computed
+by ``pygwb``.
+
+The ``pygwb`` package comes with a script, ``pygwb_pe``, which wraps the ``pygwb.pe`` module. In this tutorial, we show how to
+use this script and escribe the basic functionalities. However, we show some additional features of the model in one of the demos, 
+to illustrate the customizability of the module.
+
+.. tip::
+
+    Make sure to have a look at the ``pygwb.pe`` `module page <api/pygwb.pe.html>`_ before going through the tutorial for additional information.
+
 
 **1. Script parameters**
-=========
+========================
 
-You can see the script's run options by running
+The input paramaters of the ``pygwb_pe`` script  can be accessed by running
 
 .. code-block:: shell
 
    pygwb_pe --help
    
-This will show all the parameters of the script as follows: 
+This will display the parameters of the script as follows: 
 
 .. code-block:: shell
 
-    usage: pygwb_pe [-h] --path_to_file PATH_TO_FILE [--ifos IFOS] [--model MODEL] [--model_prior_file MODEL_PRIOR_FILE] [--non_prior_args NON_PRIOR_ARGS] [--output_dir OUTPUT_DIR]
-                [--injection_parameters INJECTION_PARAMETERS] [--quantiles QUANTILES] [--f0 F0] [--fhigh FHIGH] [--df DF]
+    pygwb_pe [-h] --input_file INPUT_FILE [--ifos IFOS] [--model MODEL][--model_prior_file MODEL_PRIOR_FILE]
+    [--non_prior_args NON_PRIOR_ARGS] [--outdir OUTDIR][--apply_notching APPLY_NOTCHING]
+    [--notch_list_path NOTCH_LIST_PATH][--injection_parameters INJECTION_PARAMETERS]
+    [--quantiles QUANTILES][--calibration_epsilon CALIBRATION_EPSILON]
 
     optional arguments:
-      -h, --help            show this help message and exit
-      --path_to_file PATH_TO_FILE
-                            Path to data file (or pickled baseline) to use for analysis.
-      --ifos IFOS           List of names of two interferometers for which you want to run PE, default is H1 and L1. Not needed when running from a pickled baseline.
-      --model MODEL         Model for which to run the PE. Default is "Power-law".
-      --model_prior_file MODEL_PRIOR_FILE
-                            The required parameters for the model and their priors. Default value is power-law parameters omega_ref with LogUniform bilby prior from 1e-11 to 1e-8 and alpha with Uniform bilby
-                            prior from -4 to 4.
-      --non_prior_args NON_PRIOR_ARGS
[…]
