+++ docs/index.rst
@@ -1,5 +1,6 @@
 pygwb documentation
-==============================================
+===================
+
 .. image:: pygwb_logo_plasma.png
    :width: 250
    :align: center

[…]
