+++ pygwb/preprocessing.py
@@ -1,14 +1,14 @@
-"""Preprocessing module collects all the functions that handle the preprocessing of the data used in the analysis.
-
-The functions of the preprocessing module handle anything related with preparing the data for the pygwb analysis run.
-It contains the ability to read data from frame files, locally or publicly.
-Other functions are able to resample the data, apply a high-pass filter to data or apply a timeshift.
+"""The ``preprocessing`` module combines all the functions that handle the preprocessing of the data used in the analysis.
+This is anything related to the preparation of the data for the pygwb analysis run.
+It can read data from frame files, locally or publicly.
+Other functionalities include resampling the data, applying a high-pass filter to data or applying a timeshift.
 These functionalities come together in the triplet of ``preprocessing_data`` functions
 which reads in data and resamples and/or high-passes it immediately.
-The triplet can work for a ``gwpy.timeseries.TimeSeries``, a normal array or using a gravitational wave channel
+The triplet can work for a ``gwpy.timeseries.TimeSeries``, a normal array or using a gravitational-wave channel
 that will read data from that channel using the provided local or public frame files.
 
-The last function of the module can gate data based on the gate function in gwpy, ``gwpy.timeseries.TimeSeries.gate``.
+A last functionality of the module is to gate data based on the gating function in ``gwpy``, ``gwpy.timeseries.TimeSeries.gate``.
+More information can be found `here <https://gwpy.github.io/docs/stable/api/gwpy.timeseries.TimeSeries/#gwpy.timeseries.TimeSeries.gate>`_.
 
 Examples
 --------
@@ -18,9 +18,9 @@ First, we have to import the module.
 
 >>> import pygwb.preprocessing as ppp
 
-Then, we read in some data utilising the ``read_data`` function of preprocessing.
-We focus on reading in public data out of convenience from the "H1" detector.
-We want to read strain data hence the choice for the channel name.
+Then, we read in some data using the ``read_data`` method.
+For concreteness, we read in public data from the LIGO Hanford "H1" detector. This can be done as shown below. 
+The "public" tag indicates we are obtaining public data from the `GWOSC <https://gwosc.org/>`_.
 
 >>> IFO = "H1"
 >>> data_timeseries = ppp.read_data(
@@ -35,9 +35,9 @@ We want to read strain data hence the choice for the channel name.
 >>> print(data_timeseries.sample_rate)
 16384.0 Hz
 
-We print the sample rate just to show it is indeed 16 kHz at the moment.
-Now, we will preprocess the data, meaning it will be resampled and a high-pass
-filter will be applied to the data. We want to resample to 4 kHz.
+The sample rate is shown for illustrative purposes. Now, we will preprocess the data, 
+meaning it will be resampled and a high-pass
+filter will be applied to the data. As an example, the data is resampled to 4 kHz.
 
 >>> new_sample_rate = 4096
 >>> preprocessed_timeseries = ppp.preprocessing_data_gwpy_timeseries(
@@ -50,14 +50,12 @@ filter will be applied to the data. We want to resample to 4 kHz.
         "fir",     # ftype
         0          # timeshift
     )
-    
-Let's do a quick check if the sample rate was changed.
-
 >>> print(preprocessed_timeseries.sample_rate)
 4096.0 Hz
 
-It was successfully changed. The last important part of preprocessing is gating the data.
-In that case, using again default values for parameters, you can run
+One can see that the sample rate was indeed modified. 
+Another important part of preprocessing is gating the data.
+In that case, using again default values for parameters, one can run teh following lines:
 
 >>> gated_timeseries, deadtime = ppp.self_gate_data(
[…]
