+++ pygwb/detector.py
@@ -1,5 +1,5 @@
 """
-The ``detector`` module is a subclass of bilby's Interferometer class which is charged with handling, storing and
+The ``detector`` module is a subclass of bilby's Interferometer class (more details `here <https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.interferometer.Interferometer.html#bilby.gw.detector.interferometer.Interferometer>`_) which is charged with handling, storing and
 saving all relevant interferometer data. 
 It handles all data analysis parts relating to the individual detectors in a baseline of a network.
 For example, it loads the data from a certain channel and computes the psd of the detector.
@@ -7,23 +7,23 @@ For example, it loads the data from a certain channel and computes the psd of th
 Examples
 --------
     
-In this example, we will load in data from the publicly available GWOSC servers using the detector module.
-We will gate the data, compute the PSD and the average PSD of the detector object.
-This example gives a brief overview of the most critical capabilities of the pygwb detector module.
-We start by importing the Interferometer class from pygwb.
+In this example, we will load in data from the publicly available `GWOSC <https://gwosc.org/>`_ servers using the ``detector`` module.
+We will gate the data, compute the PSD and the average PSD of the ``detector`` object.
+This example gives a brief overview of the most critical capabilities of the ``pygwb.detector`` module.
+We start by importing the Interferometer class from ``pygwb``.
     
 >>> from pygwb.detector import Interferometer
 
-To load in some data, first we make an empty detector object.
-Based on the name of the object, the module will make an Interferometer object that has no data.
+To load in the data, an empty ``detector`` object is first created, for which,
+based on the name of the object, the module will make an Interferometer object without data.
 The name can be any one of the detectors supported in ``bilby.gw.detector``,
-the parent class of our Interferometer class.
+the parent class of our Interferometer class (more details `here <https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.interferometer.Interferometer.html#bilby.gw.detector.interferometer.Interferometer>`_).
     
 >>> ifo_1 = Interferometer.get_empty_interferometer("H1")
     
-Then, we load in data using one of the provided ``set_timeseries_from`` functions.
-We take a start time t0 and an end time tf. We want to use public data, so we
-set data_type to public. We use the channel "H1:GWOSC-4KHZ_R1_STRAIN". All the other parameters are taken to be
+Then, we load the data using the ``set_timeseries_from`` methods, and pass a start and end time, ``t0`` and ``tf``, respectively.
+We are obtaining data from the GWOSC servers, i.e., public data, and indicate this by marking the data_type tag to public. We use 
+the channel "H1:GWOSC-4KHZ_R1_STRAIN", for concreteness. All the other parameters are taken to be
 default values.
 
 >>> ifo_1.set_timeseries_from_channel_name(
@@ -42,7 +42,8 @@ default values.
     time_shift=0,
     )
     
-Now, we gate the detector data. In that case, we can call
+To illustrate one of the features of the module, we gate the detector data. This procedure applies a window to stretches of data to get rid of
+glitches and other artefacts. More information on the gating procedure can be found in `this technical note <https://dcc.ligo.org/public/0172/P2000546/002/gating-mdc.pdf>`_.  
 
 >>> ifo_1.gate_data_apply(
     gate_tzero=1.0,
@@ -52,9 +53,9 @@ Now, we gate the detector data. In that case, we can call
     gate_whiten=True,
     )
 
-Next, we will compute the PSD spectrogram of the detector. A spectrogram
-shows the PSD both per time and per frequency. We will use the common frequency
-resolution of pygwb analysis.
+The module also allows to compute the power spectral density (PSD) spectrogram of the detector. A spectrogram
[…]

[…]
