+++ pygwb/postprocessing.py
@@ -1,33 +1,29 @@
-"""The postprocessing modules groups all functions which are useful in the end stages of the analysis.
-In particular when combining spectrograms into spectra and spectra into one overall point estimate for the background.
+"""The postprocessing module combines all functions which are useful in the end stages of the analysis.
+In particular, when combining spectrograms into spectra, and spectra into one overall point estimate for the gravitational-wave background.
 
-When computing an estimator for the gravitational-wave background in the isotropic search,
-we go through different steps of the analysis. Postprocessing is the last stage of this analysis,
-after the computation of the PSDs and CSDs.
-
-This module contains functions such as ``postprocess_Y_sigma`` which combines point estimate and sigma spectrograms
+This module contains functions such as ``postprocess_Y_sigma``, which combines point estimate and sigma spectrograms
 into spectra using a method where odd and even segments are treated differently if the data are overlapping.
-To account for the overlap, it uses another function of the postprocessing module, ``odd_even_segment_postprocessing``.
+To account for the overlap, it uses another function of the postprocessing module, ``odd_even_segment_postprocessing``. 
+Additional information about this procedure can be found `here <https://arxiv.org/pdf/2303.15696.pdf>`_.
 
-These spectrograms have to be computed from the CSDs and PSDs.
-The postprocessing module also contains the function responsible for this calculation,
+These spectrograms have to be computed from the cross spectral density (CSD) and the power spectral density (PSD).
+The postprocessing module also contains the method which takes care of the above, namely
 ``calculate_point_estimate_sigma_spectra``.
 Starting from a set of CSD and PSD spectrograms, one can compute the point estimate and sigma spectrograms,
 objects that contain both frequency and segment data information.
-Then the spectrograms are combined into spectra with the functions mentioned above.
-These spectra then have to be combined into one single point estimate and its variance.
-To achieve that goal, another function, ``calc_Y_sigma_from_Yf_sigmaf``, is utilised.
+Then, the spectrograms are combined into spectra with the functions aforementioned.
+These spectra then have to be combined into one single point estimate and its variance, acoomplished by calling ``calc_Y_sigma_from_Yf_sigmaf``.
 
 Examples
 --------
 
 Starting from averaged PSDs and the CSD of a baseline,
-we can compute the overall point estimate for the gravitational-wave background of the isotropic analysis.
+we can compute the overall point estimate for the gravitational-wave background.
 
-We start our example with already computed CSDs and PSDs, see other modules.
-Before we can start the actual computation, we also need the overlap reduction function, see ``pygwb.orf``.
-Then, we can compute the :math:`\Omega`\ :sub:`GWB`\
-and its variance spectrograms.
+Assuming we already computed the CSDs and PSDs, see for example in :doc:`pygwb.spectral`, 
+we also need the overlap reduction function, see :doc:`pygwb.orfs`.
+With these at hand, we can compute the point estimate spectrogram
+and its variance.
 
 >>> Y_spectrogram, var_spectrogram = calculate_point_estimate_sigma_spectra(
         CSD_baseline.frequencies.value,
@@ -43,12 +39,9 @@ and its variance spectrograms.
         alpha=0.0,
     )
 
-Where we have used the CSD from a certain baseline of interferometers, their respective average PSDs and the orf.
-
-We have calculated the spectrograms, which are the point estimate as a function of frequency and time (for each segment).
-
[…]
