### 2023-08-25 7d86aab
Writing API fixes
