+++ docs/pe.rst
@@ -15,13 +15,15 @@ This allows to infer, or put constraints on, some GWB model parameters :math:`\m
 by ``pygwb``.
 
 The ``pygwb`` package comes with a script, ``pygwb_pe``, which wraps the ``pygwb.pe`` module. In this tutorial, we show how to
-use this script and escribe the basic functionalities. However, we show some additional features of the model in one of the demos, 
+use this script and describe the basic functionalities. In addition, we show some advanced features of the module in one of the `demos <run_pe.html>`_, 
 to illustrate the customizability of the module.
 
 .. tip::
 
-    Make sure to have a look at the ``pygwb.pe`` `module page <api/pygwb.pe.html>`_ before going through the tutorial for additional information.
+    Make sure to also have a look at the ``pygwb.pe`` `module page <api/pygwb.pe.html>`_ before going through the tutorial for additional information.
 
+.. note::
+    The ``pygwb.pe`` module is based on the ``bilby`` package. For more information, we refer the reader to the documentation of the ``bilby`` package `here <https://lscsoft.docs.ligo.org/bilby/index.html>`_.
 
 **1. Script parameters**
 ========================
@@ -75,23 +77,19 @@ This will display the parameters of the script as follows:
         --calibration_epsilon CALIBRATION_EPSILON
                         Calibration uncertainty. Default is 0.
 
-The only required parameter is ``path_to_file``, which refers to an output file from a ``pygwb`` analysis run. The script can be run simply with
+**2. Running the script**
+=========================
+
+The only required argument is the ``input_file``, which should point to the output file of a ``pygwb`` run containing the point estimate spectrum and its variance. 
+The script can then simply be run with
 
 .. code-block:: shell
 
     pygwb_pe --input_file {path_to_pygwb_output_file}
     
-This produces the usual output files of a ``bilby`` parameter estimation run and is saved in the ``output_dir``, which by default is ``./PE_Output``. 
-Note that upon successful completion of the parameter estimation the ``output_dir`` should contain a ``result.json`` file and a ``bilby`` corner plot 
-(in ``png`` format) summarizing the results of the PE run.
-
-**2. Customizing a run**
-========================
-
-Although the ``pygwb_pe`` can be run with the default setup as outlined above, 
-most users will want to customize some of the parameters for the run. For example, one may choose to run using a different model and different priors. 
-The model may be specified using the ``--model argument``, and must be a ``pygwb``--supported model.
-These inculde
+The above command would run ``pygwb_pipe`` with all script parameter values set to their default values. However, the various script parameters of ``pygwb_pipe``, 
+as shown above, allow for a certain level of customization. In particular, the ``pygwb_pe`` script accommodates all the models present in the ``pygwb.pe`` module through the 
+``--model`` argument (more information on available models `here <api/pygwb.pe.html>`_). These include:
 
 .. code-block:: python
 
@@ -103,12 +101,9 @@ These inculde
         "Parity-Violation": pe.PVPowerLawModel
         "Parity-Violation-2": pe.PVPowerLawModel2
 
-We refer the user to the ``pe`` documentation for a full overview of available models. 
-The priors may be specified by passing a ``prior.json`` file through the ``--model_prior_file`` argument. 
-
-This is illustrated with an example below for power-law model priors. A Log-uniform prior for Omega from 
-1e-13 to 1e-5 and a Gaussian prior for alpha with mean of 2/3 and sigma 1.5 are taken. Such a json file 
-can be made using the following code:
[…]
