+++ docs/pipeline.rst
@@ -2,18 +2,24 @@
 Using ``pygwb_pipe``: a quickstart manual
 =========================================
 
+The various modules of the ``pygwb`` package can be combined into a pipeline, as done in the ``pygwb_pipe`` script. This script 
+takes the data as input and outputs an estimator of the point estimate and variance of the gravitational-wave background (GWB) for these
+data. More information on how the various modules interact and are combined into a pipeline can be found in the `pygwb paper <https://arxiv.org/pdf/2303.15696.pdf>`_.
 
-**1. Testing the pipeline**
-===========================
+.. note::
+  The proposed ``pygwb_pipe`` pipeline is only one of the many ways to assemble the ``pygwb`` modules, and users should
+  feel free to create their own pipeline, that addresses their needs.
 
-You can view the pipeline run options by executing
+**1. Script parameters**
+========================
 
+The parameters of the ``pygwb_pipe`` script can be visualized by running the following command:
 
 .. code-block:: shell
 
    pygwb_pipe --help
 
-This will list all arguments of the pipeline:
+This will display the following set of parameters, which can be passed to the pipeline:
 
 .. code-block:: shell
 
@@ -115,34 +121,53 @@ This will list all arguments of the pipeline:
                         option to return naive and sliding sigmas from delta sigma
                         cut. Default value: False
 
-To test the pipeline, simply run a command like
+As can be seen, all of the parameters above come with a brief description, which should help the user identify its functionality. In particular,
+we note that the above parameters are the ones present in the ``pygwb.parameters`` module. For more information, one can have a look at the 
+`pygwb paper <https://arxiv.org/pdf/2303.15696.pdf>`_, where more details are provided.
+
+.. tip::
+  Feeling overwhelmed with the amount of parameters? Make sure to have a look to the ``pygwb.parameters`` `documentation <api/pygwb.parameters.html>`_.
+
+**2. Running the script**
+========================
+
+Although all of the parameters shown above can be passed to the script, we start by running ``pygwb_pipe`` without passing any optional parameters directly to the script.
+The only required argument is a path to a parameter file, which contains the parameter values
+to use for the analysis. As an example, one can run the script with the ``parameters.ini`` file provided in the ``pygwb_pipe`` directory of the 
+`repository <https://github.com/a-renzini/pygwb/blob/master/pygwb_pipe/parameters.ini>`_. To test the pipeline, run the command:
 
 .. code-block:: shell
 
-   pygwb_pipe --param_file {path_to_param_file} --apply_dsc False
+  pygwb_pipe --param_file {path_to_param_file}
 
-When running on the file ``pygwb_pipe/parameters.ini`` in the repo, one should get as final result
+The output should be of the command above should be:
 
 .. code-block:: c
 
-   2023-02-21 14:43:40.817 | SUCCESS  | __main__:main:160 - Ran stochastic search over times 1247644138-1247645038                                           
-   2023-02-24 16:35:25.625 | SUCCESS  | __main__:main:163 - POINT ESTIMATE: -6.496991e-06
-   2023-02-24 16:35:25.625 | SUCCESS  | __main__:main:164 - SIGMA: 2.688777e-06
+  2023-02-21 14:43:40.817 | SUCCESS  | __main__:main:160 - Ran stochastic search over times 1247644138-1247645038                                           
+  2023-02-24 16:35:25.625 | SUCCESS  | __main__:main:163 - POINT ESTIMATE: -6.496991e-06
+  2023-02-24 16:35:25.625 | SUCCESS  | __main__:main:164 - SIGMA: 2.688777e-06
 
-Note that this automatically includes the default notching. If an error related to the notch file appears, it may be necessary to add the correct path explicitly in the ``.ini`` file used.
[…]
