### 2023-08-29 a833fa6
Finalizing pygwb stats tutorial page
