+++ docs/api.rst
@@ -2,13 +2,15 @@
 API
 ============
 
-The ``pygwb`` package consists of several modules, each with a distinct functionality. A summary of all ``pygwb`` modules and its main external dependencies can be seen in the schema below.
+The ``pygwb`` package consists of several modules, each with a distinct functionality. A summary of all ``pygwb`` modules, their interaction with each other, and the main external dependencies can be seen in the schema below.
+This API section contains information about each of the modules, as well as a description of all methods therein. The module documentation can be accessed through the table below for each of the modules, or through the table of contents on the left hand side of the page.
 
+.. tip::
+
+   Additional information about each of the modules can be found in the `pygwb paper <https://arxiv.org/pdf/2303.15696.pdf>`_.
 
 .. image:: pygwb_modules.png
    :width: 800
-
-This API section contains information about each of the modules, as well as a description of all methods therein. The modules can be accessed through the table below, or through the table of contents on the left hand side of the page.
 
 .. currentmodule:: pygwb
 
