+++ docs/simulator.rst
@@ -2,6 +2,10 @@
 Simulating data with ``pygwb.simulator``
 ========================================
 
+In this tutorial, we show how the ``pygwb.simulator`` module can be used to simulate a gravitational-wave background. We consider both the case
+where a power-law spectrum is injected, and where individual compact binaries are injected. For more information about the module, we refer
+the reader to the ``pygwb.simulator`` `documentation page <api/pygwb.simulator.html>`_.
+
 **1. Simulating a stochastic gravitational-wave background**
 ============================================================
 
@@ -63,6 +67,10 @@ desired signal PSD in the form of a ``gwpy.FrequencySeries``.
 Note that the above signal PSD was chosen for illustrative purposes. However, in practice, any 
 signal PSD can be chosen to be injected, and one does not need to restrict themselves to a broken power-law.
 
+.. seealso::
+
+    More information about ``gwpy.frequencyseries.FrequencySeries`` can be found `here <https://gwpy.github.io/docs/stable/api/gwpy.frequencyseries.FrequencySeries/>`_.
+
 One also needs to specify the parameters that will serve as input to the ``simulator``. Concretely, we specify 
 the duration of each simulated segment, the number of segments, and the sampling frequency.
    
@@ -71,6 +79,10 @@ the duration of each simulated segment, the number of segments, and the sampling
     duration = 60 # duration of each segment of data (s)
     N_segs = 10  # number of data segments to generate
     sampling_frequency = 1024 # Hz
+
+.. tip::
+
+    Not sure about what the above parameters do? Make sure to check out the `documentation <api/pygwb.simulator.html>`_ of the ``simulator`` module.
 
 The detectors for which data with the above signal PSD need to be simulated, have to be passed 
 to the ``simulator`` module. By relying on the ``detector`` module, we instantiate various detectors below.  
@@ -93,14 +105,26 @@ these parameters.
     
      net_HL = Network('HL', ifo_list)
 
+.. seealso::
+
+    Additional informational information about the ``Interferometer`` object can be found `here <api/pygwb.detector.Interferometer.html>`_. For more information, we also refer the reader to the ``bilby``
+    `documentation <https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.html>`_.
+
+
 We are now ready to simulate the data, consisting of a signal and Gaussian noise, colored by the noise PSD saved in each of the detectors. 
-We rely on the ``network`` module to simulate the data by calling the ``set_interferometer_data_from_simulator`` method (which uses the ``simulator`` module).
+We rely on the ``network`` module to simulate the data by calling the ``set_interferometer_data_from_simulator()`` method (which uses the ``simulator`` module).
+More information on the method can be found `here <api/pygwb.network.Network.html#pygwb.network.Network.set_interferometer_data_from_simulator>`_.
 
 .. code-block:: python
 
      net_HL.set_interferometer_data_from_simulator(N_segments=N_segs, GWB_intensity=Intensity_GW_inject, sampling_frequency=sampling_frequency)
 
-Note that one may save the data by calling ``save_interferometer_data_to_file`` and specifying the file format as an argument. This wraps ``gwpy.TimeSeries.write()``.
+
+.. note::
+
+    One may save the data by calling ``pygwb.network.save_interferometer_data_to_file()`` (see `here <api/pygwb.network.Network.html#pygwb.network.Network.save_interferometer_data_to_file>`_) 
+    and specifying the file format as an argument. This wraps the ``gwpy.TimeSeries.write()`` method (more details can be found 
+    `here <https://gwpy.github.io/docs/stable/api/gwpy.timeseries.TimeSeries/#gwpy.timeseries.TimeSeries.write>`_).
 
[…]
