+++ docs/multiple_jobs.rst
@@ -2,18 +2,29 @@
 Running multiple ``pygwb_pipe`` jobs
 ====================================
 
-**1. Writing and submitting a `dag` file**
-=========
+In practice, one will probably want to run ``pygwb`` on long stretches of data. This is achieved most easily by splitting
+the large data set in smaller chunks of data. These can then be analyzed individually, and combined after the analysis to form
+one overall result for the whole data set. To this end, ``pygwb`` comes with two scripts: ``pygwb_dag`` and ``pygwb_combine``. 
+The former allows the user to run ``pygwb_pipe`` (see `here <pipeline.html>`_) simultaneously on shorter stretches of data, 
+whereas the latter allows to combine the output of the individual runs into an overall result for the whole data set.
 
-We are now ready to condorise the pipeline and run a batch of jobs, just like the one run in point 4.
+**The pygwb_dag script**
+========================
 
-* *writing the* ``dag`` *file*
+**1. Script parameteres**
+-------------------------
 
-To prepare a dag file one  can use the ``pygwb_dag`` script:
+To be able to run multiple ``pygwb_pipe`` jobs simultaneously, ``pygwb`` relies on `Condor <https://htcondor.readthedocs.io/en/latest/>`_.
+This requires a ``dag`` file, which contains information about all the jobs, i.e., running ``pygwb_pipe`` on different stretches of data.
+In ``pygwb``, this file can be created by using the ``pygwb_dag`` script. To visualize the expected argument of the script, one can call:
 
 .. code-block:: shell
 
    pygwb_dag --help
+
+This will display the required parameters, together with a small description:
+
+.. code-block:: shell
 
   --subfile SUBFILE     Submission file.
   --jobfile JOBFILE     Job file with start and end times and duration for each job.
@@ -35,27 +46,91 @@ To prepare a dag file one  can use the ``pygwb_dag`` script:
   --calc_pt_est CALC_PT_EST
                         Calculate omega point estimate and sigma from data.
 
-This script passes on relevant arguments to ``pygwb_pipe``, such as the parameter file and the ``apply_dsc`` Flag, etc.
-Note that the condor submission file is not included in the package. Its compilation will depend on the specific cluster/setup used, and is left up to the user.
+An important argument of the script, is the path to the job file, passed through ``--jobfile``. The job file is a simple ``.txt`` file and contains the different jobs, or in other words,
+the different stretches of data to run the analysis on. For concretenes, consider the case where one would want to run ``pygwb`` on 12000 seconds of data, but split this in smaller jobs.
+The job file could then look as follows:
 
-* *submitting the job*
+.. code-block:: shell
 
-The ``dag`` file is now created in the ``output`` folder. To submit the job, navigate to that folder and run
+  1 0 4000  4000
+  1 4000  9000  5000
+  1 9000 12000 3000
+
+The first column does not play a role, the second and third colum indicate the start and end time of the job, respectively, whereas the last column shows the duration of the job, i.e., the 
+difference between end and start time. The job file therefore allows the script to *know* on which stretches of data to run. In case one wants to run on a subset of the jobs in the 
+job file, one can pass an additional start and end time to the script through the ``--t0`` and ``--tf`` arguments.
+
+The ``--parentdir`` allows to pass the full path to the run directory, and the ``--param_file`` should point to the parameter file to be used by ``pygwb_pipe``.
+
+.. seealso::
+  For more information about ``pygwb_pipe`` and the usage of a parameter file, we refer the user to the tutorial `here <pipeline.html>`_.
+
[…]
