+++ docs/workflow.rst
@@ -1,0 +1,5 @@
+======================
+The ``pygwb`` workflow
+======================
+
+BLABLA
