+++ docs/workflow.rst
@@ -2,4 +2,126 @@
 The ``pygwb`` workflow
 ======================
 
-BLABLA
+In practice, one will run the ``pygwb_pipe`` script on multiple jobs through the ``pygwb_dag`` script, and combine the output with 
+``pygwb_combine``. After this, ``pygwb_stats`` can be run to obtain the statistical checks diagnostic plots. To facilitate the usage
+of the different script and the user-friendliness, ``pygwb`` comes with an additional script that combines the scipts above into a worklfow,
+executing one after the other.
+
+.. warning::
+
+    The ``pygwb`` workflow is tailored to the LIGO-Virgo-KAGRA data stream. Therefore, it is possible that not all
+    data outside the LIGO-Virgo-KAGRA format are supported by the current version of the workflow.
+
+.. tip::
+    
+    Still want to run ``pygwb`` on your data without relying on the workflow? Use the `tutorial <pipeline.html>`_ about ``pygwb_pipe`` to 
+    get started.
+
+**1. Script parameters**
+========================
+
+To show the optional arguments of the script, one can run:
+
+.. code-block:: shell
+
+    pygwb_create_isotropic_workflow --help
+
+This will display the following arguments of the script, with a brief description of the parameters:
+
+.. code-block:: shell
+
+    -h, --help            show this help message and exit
+    -v, --verbose         increase verbose output
+    --debug               increase verbose output to debug
+    -V, --version         show program's version number and exit
+    --basedir BASEDIR     Build directory for condor logging
+    --configfile CONFIGFILE
+                        config file with workflow parameters
+    --config-overrides CONFIG_OVERRIDES [CONFIG_OVERRIDES ...]
+                        config overrides listed in the form SECTION:KEY:VALUE
+    --submit              Submit workflow automatically
+    --run_locally         Run job on local universe
+
+
+**2. Running the script**
+=========================
+
+To run the workflow, one executes the command:
+
+.. code-block:: shell
+
+    pygwb_create_isotropic_workflow --debug --basedir ./ --configfile ./my_config_file.ini --submit
+
+The ``--debug`` argument increases the amount of verbose printed out by the script, whereas the ``--submit`` tells the script to automatically submit the different jobs to Condor.
+
+.. tip::
+
+    Need a reminder about multiple jobs being submitted to the cluster? Check out the tutorial on multiple jobs `here <multiple_jobs.html>`_.
+
+
+In addition, a configuration file needs to be passed to the script through the ``--configfile`` argument. An example of a possible configuration file for the workflow is 
+provided in the ``pygwb`` repo `here <https://github.com/a-renzini/pygwb/blob/master/pygwb_pipe/workflow_config.ini>`_. The configuration file is divided into different
+categories, which we discuss step by step below.
+
+A first part of the file contains general information to initialize the worksflow, such as the interferometers for which to retrieve data, the duration of the jobs, as well as
+the start and end time for which to analyze data.
+
+.. code-block:: shell
+
+    [general]
+    accounting_group = ligo.dev.o4.sgwb.isotropic.stochastic
+    ifos = H1 L1
+    plot_segment_results = False
+    max_job_dur = 5000
+    min_job_dur = 900
+    t0 = 1368921618
+    tf = 1374883218
+
+As mentioned earlier, the workflow combines different script of the ``pygwb`` package. These are passed through the ``[executables]`` section of the configuration file:
+
+.. code-block:: shell
+
+    [executables]
+    pygwb_pipe = pygwb_pipe
+    pygwb_combine = pygwb_combine
+    pygwb_stats = pygwb_stats
+    pygwb_html = pygwb_html
+
[…]
