+++ docs/demos.rst
@@ -8,7 +8,7 @@ functionalities to your own needs.
 
 .. raw:: html
 
-   <a href="run_pe.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 330.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">PE with multiple models</button></a>
+   <a href="run_pe.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 330.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Customize parameter estimation</button></a>
    <a href="make_notchlist.html"><button style="background-color:#307FC1;color:white;border-color:#307FC1;width: 330.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Make a frequency notch list</button></a>
    <a href="test_orfs.html"><button style="background-color:#307FC1;border-color:#307FC1;color:white;width: 330.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Compute the overlap reduction function</button></a>
    <a href="run_statistical_checks.html"><button style="background-color:#307FC1;border-color:#307FC1;color:white;width: 330.0px; height: 50.0px;border-radius: 8px;margin-bottom: 10px;margin-left: 10px">Interpreting statistical checks</button></a>
@@ -16,6 +16,7 @@ functionalities to your own needs.
    <br />
 
 .. tip::
+   
    If you are feeling overwhelmed with these demos, make sure to have a look to the `tutorials <tutorials.html>`_ before.
       
 .. currentmodule:: pygwb
