### 2023-09-08 9c3c2b4
Fixing typos in documentation
