### 2023-09-08 c37dca1
Finishing stat checks tutorial
