+++ pygwb/parameters.py
@@ -10,74 +10,74 @@ It is a container for all the required ingredients to perform a stochastic gravi
 Examples
 --------
 
-When using a packaged script, parameters passed to the script directly will be stored in a ```Parameters``` object. 
-To pass these, the standard ```argparse``` usage applies; for example, when using ```pygwb_pipe```:
+When using a packaged script, parameters passed to the script directly will be stored in a ``Parameters`` object. 
+To pass these, the standard ``argparse`` usage applies; for example, when using ``pygwb_pipe``.
 
 >>> pygwb_pipe --parameter_to_pass [my-parameter-value]
 
-A parameter class may also be instantiated from a parameter file using the appropriate method, ```Parameters.update_from_file```.
-A parameter file must be structured specifically to be read by the ```Parameters``` class. The following structure should be followed:
+A parameter class may also be instantiated from a parameter file using the appropriate method, ``Parameters.update_from_file``.
+A parameter file must be structured specifically to be read by the ``Parameters`` class. The following structure should be followed::
 
->>> [data_specs]
->>> interferometer_list: ["IFO1", "IFO2"]
->>> t0:
->>> tf:
->>> data_type:
->>> channel:
->>> frametype:
->>> time_shift:
->>> [preprocessing]
->>> new_sample_rate: 
->>> input_sample_rate: 
->>> cutoff_frequency: 
->>> segment_duration: 
->>> number_cropped_seconds: 
->>> window_downsampling: 
->>> ftype: 
->>> [gating]
->>> gate_data: 
->>> gate_whiten: 
->>> gate_tzero: 
->>> gate_tpad: 
->>> gate_threshold: 
->>> cluster_window: 
->>> [window_fft_specs]
->>> window_fftgram: 
->>> alpha:
->>> [window_fft_welch_specs]
->>> window_fftgram: 
->>> alpha:
->>> [density_estimation]
->>> frequency_resolution: 
->>> N_average_segments_psd: 
->>> coarse_grain_psd: 
->>> coarse_grain_csd: 
->>> overlap_factor_welch: 
->>> overlap_factor: 
->>> [postprocessing]
->>> polarization:  
->>> alpha: 
->>> fref: 
->>> flow: 
->>> fhigh: 
->>> [data_quality]
->>> notch_list_path: 
->>> calibration_epsilon: 
->>> alphas_delta_sigma_cut:
->>> delta_sigma_cut: 
->>> return_naive_and_averaged_sigmas: 
->>> [output]
->>> save_data_type:
->>> [local_data]
->>> local_data_path:
+    [data_specs]
+    interferometer_list: ["IFO1", "IFO2"]
+    t0:
+    tf:
+    data_type:
+    channel:
+    frametype:
+    time_shift:
+    [preprocessing]
+    new_sample_rate: 
+    input_sample_rate: 
+    cutoff_frequency: 
+    segment_duration: 
+    number_cropped_seconds: 
+    window_downsampling: 
+    ftype: 
+    [gating]
+    gate_data: 
+    gate_whiten: 
+    gate_tzero: 
+    gate_tpad: 
+    gate_threshold: 
+    cluster_window: 
+    [window_fft_specs]
+    window_fftgram: 
+    alpha:
+    [window_fft_welch_specs]
+    window_fftgram: 
+    alpha:
+    [density_estimation]
+    frequency_resolution: 
+    N_average_segments_psd: 
+    coarse_grain_psd: 
+    coarse_grain_csd: 
+    overlap_factor_welch: 
+    overlap_factor: 
+    [postprocessing]
+    polarization:  
+    alpha: 
+    fref: 
+    flow: 
+    fhigh: 
+    [data_quality]
+    notch_list_path: 
+    calibration_epsilon: 
[…]
