### 2023-09-14 a-renzini/pygwb#1 @kevinturbang
Dear @Sbozzolo,

First of all, thank you very much for all the useful feedback. This has been very valuable to get the documentation to its current status. We have implemented the comments above, and tried to add a more logical structure within the documentation itself. Could you please have a look and let us know if any further improvements are necessary? Thank you!
