### 2023-09-16 a-renzini/pygwb#1 @Sbozzolo
Dear PyGWB team,

Thank you for your efforts. The user documentation is looking great (I like the "tips" to direct the reader to other pages).

I only have two remaining comments, one trivial, and the other less so:
- I think it is uncommon to have external links in the main menu (the one of the side, e.g., "pyGWB paper"), so I was surprised when I clicked on some items and left the documentation. There's nothing wrong about it, but I think it's always good to follow the principle of least surprise.
- It would be great if you could describe in more detail the input/outputs of the various scripts, and maybe provide some examples and comments. You are already kind of doing that, but you fall short from actually showing and discussing the output. Instead of telling the user "load the file and print the keys", I'd recommend being very explicit. Users that might be interested in your package would want to know what it produces. Moreover, plots are a great way for visual learners to get a sense of what `pyGWB` is about.
