### 2023-09-16 a-renzini/pygwb#1 @Sbozzolo
Also, the `CONTRIBUTING.MD` file and the contributing page in the docs are not perfectly in agreement.

By the way, can non-LIGO users contribute using LIGO's gitlab?
