+++ setup.py
@@ -24,7 +24,7 @@ long_description = get_long_description()
 
 setup(
     name="pygwb",
-    description="Lighweight python stochastic GWB analysis pipeline",
+    description="Lighweight python stochastic gravitational-wave background analysis pipeline",
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://git.ligo.org/pygwb/pygwb",
@@ -53,6 +53,7 @@ setup(
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
