+++ pygwb/spectral.py
@@ -71,9 +71,7 @@ The CSD is calculated by invoking the following lines of codes.
     
 For any additional information on the parameters that enter the methods described here,
 we refer the reader to the more detailed API documentation of the module.
-
 """
-
 import copy
 import warnings
 
@@ -84,7 +82,6 @@ from gwpy.timeseries import TimeSeries
 from scipy.signal import get_window, spectrogram
 
 from pygwb.util import get_window_tuple, parse_window_dict
-
 
 def fftgram(
     time_series_data: TimeSeries,
@@ -130,7 +127,6 @@ def fftgram(
     gwpy.frequencyseries.FrequencySeries
         More information `here <https://gwpy.github.io/docs/latest/api/gwpy.frequencyseries.FrequencySeries/#gwpy.frequencyseries.FrequencySeries>`_.
     """
-
     sample_rate = int(1 / time_series_data.dt.value)
 
     # get the window function
@@ -165,7 +161,6 @@ def fftgram(
 
     return data_fftgram
 
-
 def pwelch_psd(
     psdgram: Spectrogram, segment_duration: int, overlap_factor: float = 0.5
 ):
@@ -193,7 +188,6 @@ def pwelch_psd(
     gwpy.spectrogram.Spectrogram
         More information `here <https://gwpy.github.io/docs/stable/spectrogram/>`_.
     """
-
     averaging_factor = round(segment_duration * psdgram.dy.value)
 
     if averaging_factor < 1:
@@ -232,7 +226,6 @@ def pwelch_psd(
 
     return avg_psdgram
 
-
 def before_after_average(psdgram: Spectrogram, segment_duration: int, N_avg_segs: int):
     """
     Average the requested number of PSDs from segments adjacent to the segment 
@@ -258,7 +251,6 @@ def before_after_average(psdgram: Spectrogram, segment_duration: int, N_avg_segs
     gwpy.spectrogram.Spectrogram
         More information `here <https://gwpy.github.io/docs/stable/spectrogram/>`_.
     """
-
     if N_avg_segs < 2:
         raise ValueError("N_avg_segs should be >=2")
 
@@ -307,7 +299,6 @@ def before_after_average(psdgram: Spectrogram, segment_duration: int, N_avg_segs
     )
     return avg_psdgram
 
-
 def coarse_grain(data, coarsening_factor):
     """
     Coarse grain a frequency series by an integer factor.
@@ -354,7 +345,6 @@ def coarse_grain(data, coarsening_factor):
         coarsened += (left_edges - right_edges) / 2 / coarsening_factor
     return coarsened
 
-
 def coarse_grain_exact(data, coarsening_factor):
     """
     Coarse grain an array using any coarsening factor.
@@ -395,7 +385,6 @@ def coarse_grain_exact(data, coarsening_factor):
 
     return output
 
-
 def coarse_grain_naive(data, coarsening_factor):
     """
     Naive implementation of a coarse graining factor that ignores edge effects.
@@ -420,7 +409,6 @@ def coarse_grain_naive(data, coarsening_factor):
         data = data[:-n_remove]
     coarsened = np.mean(data.reshape(-1, coarsening_factor), axis=-1)
     return coarsened
-
 
 def coarse_grain_spectrogram(
     spectrogram,
@@ -507,9 +495,7 @@ def coarse_grain_spectrogram(
         epoch=coarse_times[0].value,
         dt=coarse_times[1].value - coarse_times[0].value,
     )
-
     return output
-
 
 def cross_spectral_density(
     time_series_data1: TimeSeries,
@@ -573,7 +559,6 @@ def cross_spectral_density(
     gwpy.spectrogram.Spectrogram
[…]
