+++ pygwb/parameters.py
@@ -115,6 +115,7 @@ from typing import List
 
 import json5
 
+
 @dataclass
 class Parameters:
     """
