+++ pygwb/util.py
@@ -11,6 +11,7 @@ from scipy.signal import get_window
 
 from pygwb.constants import H0
 
+
 def parse_window_dict(window_dict):
     """
     Parse the window dictionary properly for scipy compatibility.
