### 2023-09-28 f1791bf
Removing comparisons with booleans.
