### 2023-09-29 e542be4
Adding contribution guidelines for non-LVK
