+++ docs/pipeline.rst
@@ -201,7 +201,7 @@ Additionally, the power-spectral densities (PSDs) and cross-spectral densities (
   Not sure about what is exactly in a file? Load in the file and print out all its `keys` as shown 
   `here <https://stackoverflow.com/questions/49219436/how-to-show-all-the-element-names-in-a-npz-file-without-having-to-load-the-compl>`_.
   
-Printing these keys will show you:
+Printing these keys displays the following:
 
 .. code-block:: shell
 
@@ -212,32 +212,30 @@ Printing these keys will show you:
    'csd_times', 'avg_csd_times', 'psd_times', 'avg_psd_times',
    'coherence', 'psd_1_coh', 'psd_2_coh', 'csd_coh', 'n_segs_coh']
   
-All these keys are saved in the ``.npz`` file mentioned above. Some of those might be empty when not  Their corresponding data can be read using:
+The above keys of the ``.npz`` have corresponding data associated to them, which can be read using:
 
 .. code-block:: shell
 
   variable = npzfile['{key}']
 
-These keys can provide the frequencies used in the analysis, both for naive estimates (``'freqs'``) and averaged estimates (``'avg_freqs'``) of the spectral densities. Those can be read using the corresponding keys. For example the csd is read using the key ``'csd'`` and the average csd can be found with the key ``'avg_csd'``. Same applies for the PSDs of the interferometers.
+More specifically, the frequencies for naive estimates can be accessed through the ``'freqs'`` key, whereas the ones for averaged 
+estimates of the spectral densitities can be accessed through the ``'avg_freqs'`` key. Additionally, the CSD can be read using the 
+``'csd'`` key and the average CSD can be found with the key ``'avg_csd'``. Analogously, one can load the PSDs of the interferometers.
+One can also read the times associated to these spectral densities by using the keys ``'{insert_spectral_density}_times'``. If the ``--calc_coh`` 
+argument was set to ``True`` during the analysis, the coherence information will also be stored in this file under the ``'coherence'`` key
+together with the PSDs, CSD and amount of segments used to compute coherence. 
 
-You can also read the correct times of these spectral densities by using the keys ``'{insert_spectral_density}_times'``. If the ``--calc_coh`` argument was put to ``True``, the coherence information will also be stored in this file under the key ``'coherence'`` together with the PSDs, CSD and amount of segments used to compute coherence. 
+.. note::
+  
+  Depending on the parameters used to run ``pygwb_pipe``, some keys above might not have a avalue associated to them.
 
-A second file contains the actual point estimate spectrum, variance spectrum, point estimate and variance. This information is accessible in:
+A second file contains the actual point estimate spectrum, variance spectrum, point estimate and variance. These can be found in:
 
 .. code-block:: shell
 
   point_estimate_sigma_{start_time_of_job}_{job_duration}.npz
 
-Furthermore, if the script was run with ``--pickle_out True``, a ``pickle`` file will be present in the output directory, containing a pickled
-version of the baseline. This contains all the information present in the other two ``npz`` files, but allows the user to create a baseline object
-from this ``pickle`` file. More information about how to create a baseline from such a file can be found `here <api/pygwb.baseline.Baseline.html#pygwb.baseline.Baseline.load_from_pickle>`_.
-
-.. warning::
-
-  Saving ``pickle`` files can take up a lot of memory. Furthermore, loading in a baseline from ``pickle`` file can take quite some time. Working 
-  with ``npz`` files is therefore recommended, when possible.
-
-The second file can be read the same way as the first one. It has the following keys available:
+This file can be read in similarly to the previous file, and has the following keys:
 
 .. code-block:: shell
 
[…]
