+++ docs/workflow.rst
@@ -121,19 +121,32 @@ Parameters related to the last part of the workflow are passed through the follo
 **3. Output of the script**
 ===========================
 
-As mentioned in the introduction of the tutorial, the workflow combines the different ``pygwb`` scripts. Therefore, the output of the workflow will be similar to that of the
-individual scripts. We refrain from going over the different outputs again, but refer the user to the dedicated tutorials for more information (e.g. `pygwb_pipe <pipeline.html#output-of-the-script>`_, 
-`pygwb_combine <multiple_jobs.html#id8>`_, `pygwb_stats <stat_checks.html#output-of-the-script>`_). However, we do note the additional feature of the workflow, which displays all 
-results of the run in an ``html`` page.
+As mentioned in the introduction of this tutorial, the workflow combines the different ``pygwb`` scripts. Therefore, the output of the
+workflow will be similar to that of the
+individual scripts. We refrain from going over the different outputs again, but refer the user to the dedicated tutorials for more 
+information (e.g. `pygwb_pipe <pipeline.html#output-of-the-script>`_, 
+`pygwb_combine <multiple_jobs.html#id8>`_, `pygwb_stats <stat_checks.html#output-of-the-script>`_). 
 
-The workflow creates a bunch of directories for the output of the workflow and the submitting of the ``.dag`` files on a cluster. In the designated directory, given by ``--basedir BASEDIR`` parameter. In that directory, the following files and directories are shown
+Futhermore, we note the additional feature of the workflow which displays all 
+results of the run in an ``html`` page. The workflow script creates several directories for the output of the workflow 
+and generates the files that will be used for the submission of the ``dag`` file on a cluster. 
+In the designated directory, given by ``--basedir BASEDIR`` parameter, the following files and directories can be found:
 
 .. code-block:: shell
 
    about  condor  config.ini  index.html  output  pygwb_cache.txt
    
-The ``about`` directory has information of the analysis run you put in the workflow and is mainly for ``.html`` purposes. The ``condor`` directory stores all relevant files needed for the ``.dag`` submission to the cluster. Every individual job will have an ``.output``, ``.submit``, ``.error`` and a ``.log`` file starting with ``pygwb_pipe_t0_length``.
-The dag ensemble has more components. The dag file itself will be called ``pygwb_dagman.submit``. Other files can help finding errors when running dag files. See the overal condor documentation `here <https://htcondor.org/documentation/htcondor.html>`_ for more information. You can also find the submit files for the combine, stats and html script in these folders, starting with the name of the script.
+The ``about`` directory contains information about the analysis run, as well as the input passed to the workflow and is mainly used to set up the ``html`` pages. 
+The ``condor`` directory stores all relevant files needed for the ``dag`` submission to the cluster. Every individual job 
+will have an ``output``, ``submit``, ``error`` and a ``log`` file starting with ``pygwb_pipe_{t0}_{length}``.
+The dag file itself is labeled ``pygwb_dagman.submit``, and can be submitted through ``condor_submit_dag pygwb_dagman.submit``. The ``error`` and ``log`` files can help finding 
+errors when running the ``dag`` file. See the Condor documentation `here <https://htcondor.org/documentation/htcondor.html>`_ 
+for more information. The submit files for the ``pygwb_combine``, ``pygwb_stats`` and ``pygwb_html`` scripts are also present in this folder.
 
-The file ``config.ini`` is a copy of your configuration file given to the workflow script via ``--configfile CONFIGFILE``. The ``output`` directory contains a subdirectory for every individual job. In such subdirectories, the output files mentioned `here <pipeline.html#output-of-the-script>`_ can be found.
[…]
