### 2023-10-05 8ef81b9
Changing some small lines mentioned in review issue 6.
