+++ pygwb/statistical_checks.py
@@ -48,7 +48,7 @@ from pygwb.parameters import Parameters
 from pygwb.util import StatKS, calc_bias
 
 
-class StatisticalChecks(object):
+class StatisticalChecks:
     def __init__(
         self,
         sliding_times_all,
