+++ pygwb/delta_sigma_cut.py
@@ -266,7 +266,7 @@ def run_dsc(
     dsigmas_dict["times"] = times
     dsigmas_dict["values"] = dsigmas
 
-    if return_naive_and_averaged_sigmas == True:
+    if return_naive_and_averaged_sigmas:
         dsigmas_dict["slide_sigmas"] = slide_sigmas * bf_ss
         dsigmas_dict["naive_sigmas"] = naive_sigmas * bf_ns
 
