+++ pygwb_pipe/Official_O3_HL_notchlist.txt
@@ -1,0 +1,100 @@
+12.421876             ,  12.453124             ,  Pulsar injection                                                                                     
+16.578126             ,  17.609374             ,  H1 calibration line                                                                                  
+17.078126             ,  18.109374             ,  H1 calibration line                                                                                  
+20.046876             ,  20.203124             ,  ASC dither line                                                                                      
+20.203126             ,  20.234374             ,  H: EBAY seirack magnetometers - L: Input Mode Cleaner pitch mode                                     
+20.234376             ,  20.265624             ,  H: EBAY seirack magnetometers - L:Input Mode Cleaner pitch mode                                      
+20.328126             ,  20.359374             ,  H: BLND_GS13Z - L:Input Mode Cleaner yaw mode                                                        
+20.703126             ,  20.859374             ,  ASC dither line                                                                                      
+21.828126             ,  21.984374             ,  ASC dither line                                                                                      
+22.265626             ,  22.421874             ,  ASC dither line                                                                                      
+24.921876             ,  25.078124             ,  Unknown comb; probably digital; not all comb teeth may be visible                                    
+26.328126             ,  26.359374             ,  Pulsar injection                                                                                     
+27.390626             ,  27.546874             ,  Triple suspension bounce mode possibly SRM                                                           
+27.640626             ,  27.796874             ,  SR3 bounce mode                                                                                      
+29.921876             ,  30.078124             ,  Unknown comb; probably digital; not all comb teeth may be visible                                    
+29.921876             ,  30.078124             ,  1Hz comb in H and 10Hz comb in L                                                                     
+31.421876             ,  31.453124             ,  Pulsar injection                                                                                     
+31.421876             ,  31.578124             ,  Unknown 0.5Hz comb; probably digital; not all comb teeth may be visible                              
+31.421876             ,  31.578124             ,  1Hz comb with 0.5Hz offset in H and L                                                                
+31.890626             ,  32.109374             ,  environmental identification: link with H corner station accelerometers - regularly intermittent     
+32.421876             ,  32.578124             ,  Unknown 0.5Hz comb; probably digital; not all comb teeth may be visible                              
+32.421876             ,  32.578124             ,  1Hz comb with 0.5Hz offset in H and L                                                                
+33.171876             ,  33.203124             ,  Calibration line nonlinearity                                                                        
+33.421876             ,  33.578124             ,  Unknown 0.5Hz comb; probably digital; not all comb teeth may be visible                              
+33.421876             ,  33.578124             ,  1Hz comb with 0.5Hz offset in H and L                                                                
+33.921876             ,  34.078124             ,  Unknown 0.5Hz comb; probably digital; not all comb teeth may be visible                              
+35.640626             ,  35.796874             ,  Environmental disturbance also seen in O1/O2                                                         
+37.609376             ,  37.828124             ,  Pulsar injection                                                                                     
+39.921876             ,  40.078124             ,  Unknown comb; probably digital; not all comb teeth may be visible                                    
+39.921876             ,  40.078124             ,  1Hz comb in H and 10Hz comb in L                                                                     
+40.828126             ,  40.984374             ,  SR2/MC2 roll mode                                                                                    
+40.859376             ,  41.015624             ,  PR2 roll mode                                                                                        
+46.984376             ,  49.015624             ,  H non-stationary, non-linear noise - scat. light -fixed at end of Sep 2019                           
+49.921876             ,  50.078124             ,  environmental identification: In L coherence with many channels including magnetic, seismic and suspension isolation. For LLO harmonic of 10 Hz comb and LHO harmonic of 1Hz comb according to CW studies
+52.796876             ,  52.828124             ,  Pulsar injection                                                                                     
[…]
