+++ pygwb_pipe/parameters.ini
@@ -64,7 +64,7 @@ flow: 20
 ; Upper frequency bound used for analysis in Hz
 fhigh: 1726
 [data_quality]
-notch_list_path: /home/arianna.renzini/PROJECTS/pygwb_main/test/test_data/Official_O3_HL_notchlist.txt
+notch_list_path: 
 ; calibration bias factor
 calibration_epsilon: 0
 ; alphas for the dsc
