### 2023-10-27 2101bbd
Fixes to correct threshold calculation and some cosmetic changes to plots
