+++ pygwb/statistical_checks.py
@@ -6,7 +6,7 @@ part of point estimate integrand, FFT of the point estimate integrand, (cumulati
 and sigma as a function of time, omega and sigma distribution, KS test, and a linear trend analysis of omega in time. 
 Furthermore, part of these plots compares the values of these quantities before and after the delta sigma cut.
 
-For additional information on how to run the statistical checks, and interpret them, we refer the user to the dedicated
+For additional information on how to run the statistical checks, and interpret them, we refer the user to the dedicatedplot_
 tutorials and demos, as well as the `pygwb paper <https://arxiv.org/pdf/2303.15696.pdf>`_.
 
 """
@@ -21,6 +21,7 @@ import matplotlib.pyplot as plt
 import matplotlib.transforms as mt
 from astropy.time import Time
 from loguru import logger
+from pygwb.util import get_window_tuple, effective_welch_averages
 
 matplotlib.rcParams['figure.figsize'] = (8,6)
 matplotlib.rcParams['axes.grid'] = True
@@ -67,6 +68,7 @@ class StatisticalChecks(object):
         baseline_name,
         param_file,
         frequency_mask = None,
+        coherence_FAR = 1,
         gates_ifo1 = None,
         gates_ifo2 = None,
         file_tag = None,
@@ -109,6 +111,8 @@ class StatisticalChecks(object):
             String with path to the file containing the parameters that were used for the analysis run.
         frequency_mask: ``array_like``
             Boolean mask applied to the specrtra in broad-band analyses. 
+        coherence_FAR: ``int``
+            Target false alarm rate for number of frequency bins in the coherence spectrum exceeding the coherence threshold.
         gates_ifo1/gates_ifo2: ``list``
             List of gates applied to interferometer 1/2.
         file_tag: ``str``
@@ -133,6 +137,7 @@ class StatisticalChecks(object):
         self.gates_ifo1 = gates_ifo1
         self.gates_ifo2 = gates_ifo2
 
+        # Possibly we should exclude the frequencies outside of the pipeline range?
         self.frequencies = frequencies
         if frequency_mask is not None:
             self.frequency_mask = frequency_mask
@@ -140,11 +145,30 @@ class StatisticalChecks(object):
             self.frequency_mask = True
 
         self.coherence_spectrum = coherence_spectrum
-        fftlength = int(1.0 / (self.frequencies[1] - self.frequencies[0]))
+        self.coherence_FAR = coherence_FAR
         if coherence_n_segs is not None:
-            self.n_segs = coherence_n_segs*(1.-self.params.overlap_factor) * int(np.floor(self.params.segment_duration/(fftlength*(1.-self.params.overlap_factor_welch)))-1)
+            fftlength = int(1.0 / (self.frequencies[1] - self.frequencies[0]))
+            nFFT = fftlength*self.params.new_sample_rate
+            nSamples = self.params.segment_duration*self.params.new_sample_rate
+            window_tuple = get_window_tuple(self.params.window_fft_dict_welch)
+            # Effective number of segments in a coherence from a single segment
+            N_eff = effective_welch_averages(nSamples, nFFT, window_tuple, self.params.overlap_factor_welch)
+
+            # Total number of effective segments (something may not right here... need to carefully check how many
+            # independent segments there should be)
+            # Also - not sure about how segments are combined to give the long-term coherence, it looks like
[…]
