### 2023-10-31 8c189a8
Added coherence FAR parameter to pygwb_stats
