+++ pygwb/statistical_checks.py
@@ -68,7 +68,7 @@ class StatisticalChecks(object):
         baseline_name,
         param_file,
         frequency_mask = None,
-        coherence_FAR = 1,
+        coherence_FAR = 1.0,
         gates_ifo1 = None,
         gates_ifo2 = None,
         file_tag = None,
@@ -111,7 +111,7 @@ class StatisticalChecks(object):
             String with path to the file containing the parameters that were used for the analysis run.
         frequency_mask: ``array_like``
             Boolean mask applied to the specrtra in broad-band analyses. 
-        coherence_FAR: ``int``
+        coherence_FAR: ``float``
             Target false alarm rate for number of frequency bins in the coherence spectrum exceeding the coherence threshold.
         gates_ifo1/gates_ifo2: ``list``
             List of gates applied to interferometer 1/2.
@@ -1617,7 +1617,7 @@ def sortingFunction(item):
 
 
 def run_statistical_checks_from_file(
-        combine_file_path, dsc_file_path, plot_dir, param_file, coherence_FAR = 1, legend_fontsize=16, coherence_file_path = None,
+        combine_file_path, dsc_file_path, plot_dir, param_file, coherence_FAR = 1.0, legend_fontsize=16, coherence_file_path = None,
         file_tag = None, convention='pygwb'
 ):
     """
@@ -1637,6 +1637,8 @@ def run_statistical_checks_from_file(
         Full path where the plots generated by the statistical checks module should be saved.
     param_file: ``str``
         Full path to the parameter file that was used for the analysis.
+    coherence_FAR: ``float``
+        Coherence false alarm rate
     legend_fontsize: ``int``, optional
         Fontsize used in the plots generated by the module. Defaults to 16.
 
