### 2023-11-03 6d90bd2
Added threshold annotation to coherence spectrum plot; slight change to how number of effective segments is calculated
