### 2023-11-04 7ee1bf2
Take real part of IFFT integrand to prevent warning message about plotting complex
