### 2023-11-06 f939bd3
Fixed incorrect calculation of coherence segments
