### 2023-11-07 aa58d4f
Changed variable name to coherence_far; changed last column of coherence outliers to be the expected count of bins above that level
