### 2023-11-13 891cecd
updating color palette in stat checks
