+++ pygwb/statistical_checks.py
@@ -32,11 +32,9 @@ from matplotlib import rc
 
 rc('font', **{'family': 'serif', 'serif': ['Computer Modern']})
 # FIXME - changed to get workflow running
-rc('text', usetex=False)
+rc('text', usetex=True)
 
 import seaborn as sns
-
-sea = sns.color_palette("tab10")
 
 import numpy as np
 from scipy import integrate, stats
@@ -70,7 +68,8 @@ class StatisticalChecks(object):
         gates_ifo2 = None,
         file_tag = None,
         legend_fontsize = 16,
-        convention = 'pygwb'
+        convention = 'pygwb',
+        seaborn_palette = 'tab10',
     ):
         """
         Instantiate a StatisticalChecks object.
@@ -205,6 +204,9 @@ class StatisticalChecks(object):
             self.xaxis = f"Hours since {t0}"
         else:
             self.xaxis = f"Days since {t0}"
+
+        self.sea = sns.color_palette(seaborn_palette)
+
 
     def get_data_after_dsc(self):
         """
@@ -379,14 +381,14 @@ class StatisticalChecks(object):
             self.days_cut,
             self.running_pt_estimate + 1.65 * self.running_sigmas,
             ".",
-            color=sea[2],
+            color=self.sea[2],
             markersize=2,
         )
         plt.plot(
             self.days_cut,
             self.running_pt_estimate - 1.65 * self.running_sigmas,
             ".",
-            color=sea[0],
+            color=self.sea[0],
             markersize=2,
         )
         plt.grid(True)
@@ -420,7 +422,7 @@ class StatisticalChecks(object):
             return
         fig = plt.figure(figsize=(10, 8))
         plt.plot(
-            self.days_cut, self.running_sigmas, '.', markersize=2, color=sea[0], label=self.baseline_name
+            self.days_cut, self.running_sigmas, '.', markersize=2, color=self.sea[0], label=self.baseline_name
         )
         plt.grid(True)
         plt.yscale("log")
@@ -447,7 +449,7 @@ class StatisticalChecks(object):
             return
 
         fig = plt.figure(figsize=(10, 8))
-        plt.plot(t_array, omega_array, color=sea[0], label=self.baseline_name)
+        plt.plot(t_array, omega_array, color=self.sea[0], label=self.baseline_name)
         plt.grid(True)
         plt.xlim(t_array[0], t_array[-1])
         plt.xlabel("Lag (s)", size=self.axes_labelsize)
@@ -471,7 +473,7 @@ class StatisticalChecks(object):
         axs.semilogy(
             self.frequencies,
             abs(self.point_estimate_spectrum / self.sigma_spectrum),
-            color=sea[0],
+            color=self.sea[0],
         )
         trans = mt.blended_transform_factory(axs.transData, axs.transAxes) 
         axs.vlines(self.frequencies[~self.frequency_mask], ymin=0, ymax=1, linewidth=1, linestyle=':', color='black', alpha=0.5, transform=trans)
@@ -499,7 +501,7 @@ class StatisticalChecks(object):
         )
         cum_pt_estimate = cum_pt_estimate / cum_pt_estimate[-1]
         fig, axs = plt.subplots(figsize=(10, 8))
-        axs.plot(self.frequencies[:-1], cum_pt_estimate, color=sea[0])
+        axs.plot(self.frequencies[:-1], cum_pt_estimate, color=self.sea[0])
         axs.set_xlabel("Frequency (Hz)", size=self.axes_labelsize)
         axs.set_ylabel(r"Cumulative $|{\rm SNR}(f)|$", size=self.axes_labelsize)
         axs.set_xscale("log")
[…]
