### 2023-11-13 6267720
removing latex compilation :(
