### 2023-11-14 bf4ba7d
Final update to calculation of effective segments; added some comments
