### 2023-11-15 cddd9f9
fixing workflow; changing tests to reduce memo
