+++ test/baseline_test.py
@@ -28,7 +28,7 @@ class TestBaseline(unittest.TestCase):
         self.interferometer_2.name = "H2"
         self.interferometer_3 = bilby.gw.detector.get_empty_interferometer("L1")
         self.duration = 4.0
-        self.sampling_frequency = 2048.0
+        self.sampling_frequency = 512.0
         self.frequencies = np.arange(100)
 
     def tearDown(self):
@@ -68,27 +68,27 @@ class TestBaseline(unittest.TestCase):
     def test_interferometer_duration_mismatch(self):
         ifo1 = copy.deepcopy(self.interferometer_1)
         ifo1.duration = 8.0
-        ifo1.sampling_frequency = 2048.0
+        ifo1.sampling_frequency = 512.0
         ifo2 = copy.deepcopy(self.interferometer_2)
         ifo2.duration = 4.0
-        ifo2.sampling_frequency = 2048.0
+        ifo2.sampling_frequency = 512.0
         with self.assertRaises(AssertionError):
             base = baseline.Baseline("H1H2", ifo1, ifo2)
 
     def test_passed_duration_ifo_mismatch(self):
         ifo1 = copy.deepcopy(self.interferometer_1)
         ifo1.duration = 8.0
-        ifo1.sampling_frequency = 2048.0
+        ifo1.sampling_frequency = 512.0
         ifo2 = copy.deepcopy(self.interferometer_2)
         ifo2.duration = 8.0
-        ifo2.sampling_frequency = 2048.0
+        ifo2.sampling_frequency = 512.0
         with self.assertRaises(AssertionError):
             base = baseline.Baseline("H1H2", ifo1, ifo2, self.duration)
 
     def test_passed_duration_ifo1_mismatch(self):
         ifo1 = copy.deepcopy(self.interferometer_1)
         ifo1.duration = 8.0
-        ifo1.sampling_frequency = 2048.0
+        ifo1.sampling_frequency = 512.0
         with self.assertRaises(AssertionError):
             base = baseline.Baseline(
                 "H1H2",
@@ -100,7 +100,7 @@ class TestBaseline(unittest.TestCase):
     def test_passed_duration_ifo2_mismatch(self):
         ifo2 = copy.deepcopy(self.interferometer_2)
         ifo2.duration = 8.0
-        ifo2.sampling_frequency = 2048.0
+        ifo2.sampling_frequency = 512.0
         with self.assertRaises(AssertionError):
             base = baseline.Baseline(
                 "H1H2",
@@ -179,7 +179,7 @@ class TestBaseline(unittest.TestCase):
 
     def test_set_cross_and_power_spectral_density(self):
         pickled_base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247645038.pickle"
+            "test/test_data/H1L1_1247644138-1247644158.pickle"
         )
         pickled_ifo_1 = pickled_base.interferometer_1
         pickled_ifo_2 = pickled_base.interferometer_2
@@ -218,9 +218,9 @@ class TestBaseline(unittest.TestCase):
 
     def test_set_average_power_spectral_densities(self):
         pickled_base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247645038.pickle"
+            "test/test_data/H1L1_1247644138-1247644158.pickle"
         )
-        pickled_base.crop_frequencies_average_psd_csd(50, 500)
+        pickled_base.crop_frequencies_average_psd_csd(50, 200)
         pickled_ifo_1 = pickled_base.interferometer_1
         pickled_ifo_2 = pickled_base.interferometer_2
         PSD_1_test = pickled_ifo_1.average_psd
@@ -233,7 +233,7 @@ class TestBaseline(unittest.TestCase):
[…]
