### 2023-11-16 #5454 @maxlalleman
Dear reviewers,

Apologies again for the delay.
We just pushed through modifications to the code which should resolve most/all issues posted by @Sbozzolo and @cmbiwer here: [https://github.com/a-renzini/pygwb/issues.](https://github.com/a-renzini/pygwb/issues.).
If you have any more questions and/or additional comments, please let us know.
