### 2023-11-17 a-renzini/pygwb#1 @kevinturbang
Dear reviewers,

Apologies for the delay on our end. We have added the following to address the points above:

- Additional explanation on how to contribute for non-LIGO members.
- Additional information about the output of the various scripts in the documentation.
- We decided to keep the external links in the documentation, as we thought these are an added value for the users. Furthermore, the example of the LEGWORK documentation that you previously shared with us also has some external links.

Hopefully this addresses all your remaining comments. Please let us know if anything else needs modification.
