### 2023-11-18 f3c5514
checking whether we can unpin astropy version
