### 2023-11-20 bba98ef
fixing test configuration
