+++ test/baseline_test.py
@@ -4,6 +4,7 @@ import os
 import pickle
 import sys
 import unittest
+from test.conftest import testdir
 
 import bilby
 import gwpy.testing.utils
@@ -162,15 +163,15 @@ class TestBaseline(unittest.TestCase):
             self.interferometer_1,
             self.interferometer_2,
         )
-        base.save_to_pickle("test/test_data/baseline_pickle_test.pickle")
+        base.save_to_pickle(f"{testdir}/test_data/baseline_pickle_test.pickle")
 
     def test_load_from_pickle(self):
         base = baseline.Baseline.load_from_pickle(
-            "test/test_data/baseline_pickle_test.pickle"
+            f"{testdir}/test_data/baseline_pickle_test.pickle"
         )
 
     def test_from_parameters(self):
-        param_file = "test/test_data/parameters_baseline_test.ini"
+        param_file = f"{testdir}/test_data/parameters_baseline_test.ini"
         params = parameters.Parameters()
         params.update_from_file(param_file)
         base = baseline.Baseline.from_parameters(
@@ -179,7 +180,7 @@ class TestBaseline(unittest.TestCase):
 
     def test_set_cross_and_power_spectral_density(self):
         pickled_base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247644158.pickle"
+            f"{testdir}/test_data/H1L1_1247644138-1247644158.pickle"
         )
         pickled_ifo_1 = pickled_base.interferometer_1
         pickled_ifo_2 = pickled_base.interferometer_2
@@ -218,8 +219,11 @@ class TestBaseline(unittest.TestCase):
 
     def test_set_average_power_spectral_densities(self):
         pickled_base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247644158.pickle"
+            f"{testdir}/test_data/H1L1_1247644138-1247644158.pickle"
         )
+        notch_file = f"{testdir}/test_data/Official_O3_HL_notchlist.txt"
+
+        pickled_base.notch_list_path = notch_file
         pickled_base.crop_frequencies_average_psd_csd(50, 200)
         pickled_ifo_1 = pickled_base.interferometer_1
         pickled_ifo_2 = pickled_base.interferometer_2
@@ -231,6 +235,8 @@ class TestBaseline(unittest.TestCase):
         ifo_1.average_psd = None
         ifo_2.average_psd = None
         base = baseline.Baseline.from_interferometers([ifo_1, ifo_2])
+        base.notch_list_path = notch_file
+
         base.set_average_power_spectral_densities()
         base.frequencies = pickled_base.frequencies
         base.crop_frequencies_average_psd_csd(50, 200)
@@ -243,16 +249,21 @@ class TestBaseline(unittest.TestCase):
 
     def test_set_average_cross_spectral_density(self):
         pickled_base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247644158.pickle"
+            f"{testdir}/test_data/H1L1_1247644138-1247644158.pickle"
         )
+        notch_file = f"{testdir}/test_data/Official_O3_HL_notchlist.txt"
+
+        pickled_base.notch_list_path = notch_file
         pickled_base.crop_frequencies_average_psd_csd(50, 200)
         pickled_ifo_1 = pickled_base.interferometer_1
         pickled_ifo_2 = pickled_base.interferometer_2
         CSD_test = pickled_base.average_csd
 
         base = baseline.Baseline.load_from_pickle(
-            "test/test_data/H1L1_1247644138-1247644158.pickle"
[…]
