+++ test/conftest.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+testdir = Path(__file__).parent.absolute()
