### 2023-11-20 a-renzini/pygwb#1 @a-renzini
Dear @Sbozzolo and team, can we mark this issue as closed? Thanks!
