### 2023-11-20 a-renzini/pygwb#1 @Sbozzolo
Documentation is a never ending process, and I invite you to keep working on it as you expand PyGWB. For JOSS, feel free to close this issue.
