+++ JOSS_submission/paper.bib
@@ -222,3 +222,9 @@ archivePrefix = {arXiv},
     month = "3",
     year = "2023"
 }
+@article{stochasticm,
+    author = "LVK",
+    title = "{stochastic.m}",
+    url = "{ https://git.ligo.org/stochastic-public/stochastic}",
+    year = "{2020}"
+}
