+++ JOSS_submission/paper.md
@@ -107,7 +107,7 @@ where $d\rho_{\rm GW}$ is the energy density of GWs in the frequency band $f$ to
 
 # Statement of need
 
-Due to the considerable amount of data to analyze, and the vast panorama of GWB models to test, the detection and characterization of a GWB requires a community effort. Furthermore, data handling and model building entail a number of different choices, depending on specific analysis purposes. This exemplifies the need for an accessible, flexible, and user-friendly open-source codebase: `pygwb`. To fully cater to user needs, `pygwb` is modular and extensively customizable, and is accompanied by exhaustive documentation.
+Due to the considerable amount of data to analyze, and the vast panorama of GWB models to test, the detection and characterization of a GWB requires a community effort. Furthermore, data handling and model building entail a number of different choices, depending on specific analysis purposes. Up until the previous LIGO-Virgo-KAGRA Collaboration (LVK) observing run, O3, the collaboration has relied on an internal matlab-based pipeline available at [@stochasticm] to perform stochastic analyses. This pipeline lacks the ability to perform parameter estimation, as well as modularity and flexibility. This exemplifies the need for an accessible, flexible, and user-friendly open-source codebase for the current and upcoming LVK runs: `pygwb`. To fully cater to user needs, `pygwb` is modular and extensively customizable, and is accompanied by exhaustive documentation.
 
 # Method
 
@@ -141,7 +141,7 @@ The `pygwb` package is class-based and modular to facilitate the evolution of th
 
 The package is compatible with GW frame files in a variety of formats, relying on the I/O functionality of `gwpy` [@gwpy]. `NumPy` [@harris2020array] is heavily used within the `pygwb` code, as well as `matplotlib` [@Hunter:2007] for plotting purposes. Some of the frequency-related computations rely on functionalities of the `scipy` [@2020SciPy-NMeth] package. The `astropy` [@astropy] package is employed for cosmology-related computations. The parameter estimation module included in `pygwb` is based on `Bilby` [@Ashton_2019] and the  `dynesty` [@Speagle_2020] sampler package.
 
-A customizable pipeline script, `pygwb_pipe`, is provided with the package and can be run in default mode, which reproduces the methodology of the LIGO-Virgo-KAGRA Collaboration (LVK) isotropic analysis implemented on the most recent observation run [@Abbott_2021]. On the other hand, the modularity of the package allows users to develop custom `pygwb` pipelines to fit their needs. 
+A customizable pipeline script, `pygwb_pipe`, is provided with the package and can be run in default mode, which reproduces the methodology of the LVK isotropic analysis implemented on the most recent observation run [@Abbott_2021]. On the other hand, the modularity of the package allows users to develop custom `pygwb` pipelines to fit their needs. 
 A set of simple statistical checks can be performed on the data after a `pygwb` run by using the `statistical_checks` module.
 In addition, a parameter estimation script, `pygwb_pe`, is also included and allows to test a subset of default models with user-defined parameters. `pygwb_pe` is based on the `pygwb` parameter estimation module, `pe`, which allows the user to test both predefined and user-defined models and obtain posterior distributions on the parameters of interest. Users are encouraged to develop and test their own models within the `pe` module.
 The `pygwb` package also contains built-in support for running on `HTCondor`-supported servers using `dag` files to parallelize the analysis of long stretches of data. Using the dedicated `pygwb_combine` script, the output can be combined into an overall estimation of the GWB for the whole data set.
