### 2023-11-20 #5454 @a-renzini
@Sbozzolo excellent point - there are precise comparisons between `pygwb` and previous code(s) in the ApJ publication but nothing here. I added a (minimal) sentence and reference about this in the "statement of need".
