### 2023-12-05 1b36ed2
Adding combined model class
