### 2023-12-05 a-renzini/pygwb#8 @cmbiwer
Tutorial is missing import statments
This page is missing the import statements to make the code useful to the user: https://pygwb.docs.ligo.org/pygwb/simulator.html#injecting-a-power-spectrum-in-random-ligo-noise

I am finding myself, stumbling through finding all the missing import statements to do it. It would be much more helpful if these were provided. Especially since ``gwpy.frequencyseries`` cannot be imported as ``from gwpy import frequencyseries``.
