### 2023-12-05 a-renzini/pygwb#8 @cmbiwer
There is also a typo in this tutorial. There's an extra indentation in:
```
 net_HL = Network('HL', ifo_list)
```
