### 2023-12-05 a-renzini/pygwb#8 @cmbiwer
And now I am stuck at:
```
Traceback (most recent call last):
  File "/Users/cmbiwer/src/pygwb/test.py", line 55, in <module>
    H1 = Interferometer.get_empty_interferometer("H1") #LIGO Hanford detector
AttributeError: type object 'Interferometer' has no attribute 'get_empty_interferometer'
```

Looking at the link to that the tutorial gives: https://lscsoft.docs.ligo.org/bilby/api/bilby.gw.detector.html

It doesn't look like this function exists?

It seems this page needs a little bit more to help the user without debugging too much.
