### 2023-12-05 a-renzini/pygwb#9 @cmbiwer
PE demo is missing data file
It seems this demo is missing a data file: https://pygwb.docs.ligo.org/pygwb/run_pe.html

Where can someone get ``'/home/patrick.meyers/git_repos/o3-isotropic-final/magnetic/magnetic_data_product_for_pe.mat'``? It seems like there is a ``o3-isotropic-final`` repo somewhere?
