### 2023-12-05 #5454 @cmbiwer
I review it again. I think the authors addressed nearly everything.

A lot of the tutorial section is geared toward a LVC user or someone with access to the same environment. Its probably has limited usefulness to a non-LVC user.

But for the parts I was able to try out the functionality I opened two issues:
 * https://github.com/a-renzini/pygwb/issues/8
 * https://github.com/a-renzini/pygwb/issues/9
