### 2023-12-12 e61ebc3
Adjusted the calculations of the axis limits
