### 2023-12-12 a-renzini/pygwb#8 @kevinturbang
Dear @cmbiwer, thank you for going through the tutorial. Indeed, this tutorial could have used a bit more polishing. This has been addressed in the most recent merge request. The above comments have been addressed in the new version of the tutorial.
