### 2023-12-14 #5454 @cmbiwer
Well, the authors did address the prior two issues, but as I continue on these tutorials I run into the following errors now:
 * https://github.com/a-renzini/pygwb/issues/10
 * https://github.com/a-renzini/pygwb/issues/11
