### 2023-12-18 a-renzini/pygwb#11 @kevinturbang
I ran the code on my end and things seem to work fine for me. It seems like this might be a multiprocessing issues. In the last line of the cell, could you try setting the `npool` parameter equal to 1, instead of 10 to see if that fixes things? Thank you.
