### 2023-12-18 a-renzini/pygwb#10 @kevinturbang
This is strange, just ran this and things seem to work fine for me. It would seem the issue is related to the generation of the bilby dictionary.  Are you working in a conda environment with the most recent version of pygwb installed?
