### 2023-12-20 a-renzini/pygwb#10 @cmbiwer
I was able to get this to work now. I think the issue was I had a slightly older version of pygwb/dependencies than what's currently on ``master``. If I install from what's on ``master`` it works. Must have been some change.
