### 2023-12-20 a-renzini/pygwb#10 @cmbiwer
I think in part it may have been because of installing jupyter. Because if I install jupyter from conda, it breaks the example.
