### 2023-12-20 a-renzini/pygwb#11 @cmbiwer
How do you recommend installing jupyter or whatever app to run the notebook? It seems some of these issues stem from me just installing the latest jupyter off conda.
