### 2024-01-12 a-renzini/pygwb#11 @a-renzini
> How do you recommend installing jupyter or whatever app to run the notebook? It seems some of these issues stem from me just installing the latest jupyter off conda.

Hi @cmbiwer  - I've now looked into this, and indeed, the issue arises when using a jupyter notebook whereas it runs fine on jupyter lab. I've reproduced your error by running:

```
conda create -n pygwb_for_joss python=3.9
git clone git@git.ligo.org:pygwb/pygwb.git
cd pygwb
pip install .
conda install jupyter
pip install ipykernel ipython==8.12.0 jupyter
ipython kernel install --name "pygwb_for_joss" --user
```
and using the `pygwb_for_joss` environment. Whereas, if I run using `jupyter lab` as we usually do, I get the same results as on the tutorial in the docs.

I think this is coming from a known bug/issue with multiprocessing on jupyter, see
- https://stackoverflow.com/questions/41385708/multiprocessing-example-giving-attributeerror
- https://bugs.python.org/issue25053

I have now fixed the issue by just removing the multiprocessing call in the notebook, `npool=10`. The example runs fast and doesn't actually need multiprocessing. I don't think it's worth having the example break when using notebooks, but we can add a note about it in the tutorials.
