### 2024-01-15 6334f5e
Add function to adress git issue #18
