### 2024-01-15 cbb4194
Add label to coherence plots for threatical expected value and for the used outlierthreshold
