+++ pygwb/notch.py
@@ -290,9 +290,9 @@ class StochNotchList(list):
 
         if np.ndim(fmin) == 1:
             for i in range(len(fmin)):
-		if fmin[i] != self[i].minimum_frequency:
+	        if fmin[i] != self[i].minimum_frequency:
             		warnings.warn("The miniumum frequency of your notch does not agree with the notch from your file. Please note, this check assumes the ordering of the notchlist object and the file with which you check is identical. Different ordering will also trigger this error.")
-		if fmax[i] != self[i].maximum_frequency:
+	        if fmax[i] != self[i].maximum_frequency:
             		warnings.warn("The maximum frequency of your notch does not agree with the notch from your file. Please note, this check assumes the ordering of the notchlist object and the file with which you check is identical. Different ordering will also trigger this error.")
 
 def power_lines(fundamental=60, nharmonics=40, df=0.2):
