### 2024-01-15 051f646
print legend in coherence plots
