### 2024-01-16 a-renzini/pygwb#11 @cmbiwer
Thanks for looking into that. That is what I was doing.

> but we can add a note about it in the tutorials.

I think that would be helpful and I'd be happy to close the issue and approve the JOSS submission.
