+++ pygwb/parameters.py
@@ -26,6 +26,7 @@ A parameter file must be structured specifically to be read by the ``Parameters`
     channel:
     frametype:
     time_shift:
+    random_timeshift:
     [preprocessing]
     new_sample_rate: 
     input_sample_rate: 
@@ -196,6 +197,10 @@ class Parameters:
         Suffix for the output data file. Options are hdf5, npz, json, pickle. Default is json.
     time_shift: ``int``
         Seconds to timeshift the data by in preprocessing. Default is 0.
+    random_timeshift: ``bool``, optional
+        Adding a random small extra shift to the time_shift.
+        This small shift is bigger than zero, but smaller than one.
+        Default not applied.
     gate_data: ``bool``
         Whether to apply self-gating to the data in preprocessing. Default is False.
     gate_tzero: ``float``
@@ -246,6 +251,7 @@ class Parameters:
     alphas_delta_sigma_cut: List = field(default_factory=lambda: [-5, 0, 3])
     save_data_type: str = "npz"
     time_shift: int = 0
+    random_timeshift: bool = False
     path_gate_data: str = ""
     gate_data: bool = False
     gate_tzero: float = 1.0
@@ -402,6 +408,7 @@ class Parameters:
         data_specs_dict["channel"] = param_dict["channel"]
         data_specs_dict["frametype"] = param_dict["frametype"]
         data_specs_dict["time_shift"] = param_dict["time_shift"]
+        data_specs_dict["random_timeshift"] = param_dict["random_timeshift"]
         param["data_specs"] = data_specs_dict
 
         preprocessing_dict = {}
@@ -571,6 +578,11 @@ class ParametersHelp(enum.Enum):
     alphas_delta_sigma_cut = "List of spectral indexes to use in delta sigma cut calculation. Default is [-5, 0, 3]."
     save_data_type = "Suffix for the output data file. Options are hdf5, npz, json, pickle. Default is json."
     time_shift = "Seconds to timeshift the data by in preprocessing. Default is 0."
+    random_timeshift = (
+        " ``bool``, optional. Adding a random small extra shift to the time_shift. "
+        "This small shift is bigger than zero, but smaller than one. "
+        "Default not applied."
+    )
     path_gate_data = (
         "Path to the pygwb output containing information about gates. "
         "If loading a single file, it has to be an .npzfile "
