+++ pygwb_pipe/parameters.ini
@@ -9,7 +9,9 @@ data_type: public
 ; Detector name will be added later
 channel: GWOSC-16KHZ_R1_STRAIN
 ; Time shift (in s)
-time_shift: 0
+time_shift: H1:0,L1:0
+; Adding random shift to time shift
+random_timeshift: False
 [preprocessing]
 ; Sample rate after resampling
 new_sample_rate: 4096
