### 2024-01-18 374d7e4
Fixing detector test due to new parameters.
