### 2024-01-18 58868ee
Fixing detector test due to new parameters.
