### 2024-01-18 #5454 @plaplant
@cmbiwer thanks for your review!

@a-renzini now that both reviewers have approved the submission, we can move forward with the final acceptance checks. If the software has changed in the course of the review (excluding changes to the paper/manuscript itself), please make a new tagged release of the repository. After that, please archive it (on [Zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or some other long-term hosting platform). Once these are done, please post the new version number and DOI of the archive in this thread. Let me know if you have any questions!
