### 2024-01-19 #5454 @a-renzini
Thanks @plaplant , and all!

I've created a pypi release of **v1.4.0** : https://pypi.org/project/pygwb/1.4.0/
And created an updated zenodo entry: https://zenodo.org/records/10532825
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10532825.svg)](https://doi.org/10.5281/zenodo.10532825)
