### 2024-01-22 #5454 @plaplant
@a-renzini there are a few (small!) final issues that need to be addressed before accepting this. There is a checklist in my comment [here](https://github.com/openjournals/joss-reviews/issues/5454#issuecomment-1900884201) that lists the last few things to do. The biggest one is that the Zenodo author list does not seem to match the JOSS paper author list. Please adjust the Zenodo authors to reflect the JOSS paper submission, and make sure that all author affiliations/ORCIDs are correct. Thanks!
