+++ JOSS_submission/paper.md
@@ -139,7 +139,7 @@ where $\Omega_{\rm M}(f|\lambda)$ is the GWB model and $\lambda$ are its paramet
 
 The `pygwb` package is class-based and modular to facilitate the evolution of the code and to increase flexibility of the analysis pipeline. The advantage of the Python language lies in rapid code execution, while maintaining a certain level of user-friendliness, which results in a shallow learning curve and will encourage future contributions to the code from the whole GW community. A summary of all `pygwb` modules and its main external dependencies can be found in the `pygwb` schema \autoref{fig:schema}.
 
-The package is compatible with GW frame files in a variety of formats, relying on the I/O functionality of `gwpy` [@gwpy]. `NumPy` [@harris2020array] is heavily used within the `pygwb` code, as well as `matplotlib` [@Hunter:2007] for plotting purposes. Some of the frequency-related computations rely on functionalities of the `scipy` [@2020SciPy-NMeth] package. The `astropy` [@astropy] package is employed for cosmology-related computations. The parameter estimation module included in `pygwb` is based on `Bilby` [@Ashton_2019] and the  `dynesty` [@Speagle_2020] sampler package.
+The package is compatible with GW frame files in a variety of formats, relying on the I/O functionality of `gwpy` [@gwpy]. `NumPy` [@harris2020array] is heavily used within the `pygwb` code, as well as `matplotlib` [@Hunter:2007] for plotting purposes. Some of the frequency-related computations rely on functionalities of the `scipy` [@2020SciPy-NMeth] package. The `astropy` [@Collaboration2022TheAP] package is employed for cosmology-related computations. The parameter estimation module included in `pygwb` is based on `Bilby` [@Ashton_2019] and the  `dynesty` [@Speagle_2020] sampler package.
 
 A customizable pipeline script, `pygwb_pipe`, is provided with the package and can be run in default mode, which reproduces the methodology of the LVK isotropic analysis implemented on the most recent observation run [@Abbott_2021]. On the other hand, the modularity of the package allows users to develop custom `pygwb` pipelines to fit their needs. 
 A set of simple statistical checks can be performed on the data after a `pygwb` run by using the `statistical_checks` module.
