### 2024-02-10 #5454 @plaplant
@a-renzini apologies for the delay. Thanks for addressing the last few issues. I made [a small PR](https://github.com/a-renzini/pygwb/pull/12) that fixes a few small citation issues in the JOSS paper. Once that's merged in, I think this is good to go!
