### 2024-02-12 477416b
Adding test for combined model
