### 2024-02-12 3b65bae
Changing documentation for combined model class
