### 2024-02-14 #5454 @a-renzini
@plaplant done - and actually I caught another missing ref. in the draft (we now have the AAS doi for the companion paper) so I've added that in now; going to re-generate
