### 2024-02-14 a1809cb
Update detector.py
