+++ pygwb/detector.py
@@ -40,7 +40,7 @@ default value.
     window_downsampling="hamming",
     ftype="fir",
     time_shift=0,
-    random_timeshift=False,
+    random_time_shift=False,
     )
     
 To illustrate one of the features of the module, we gate the detector data. This procedure applies a window to stretches of data to get rid of
@@ -235,7 +235,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
             window_downsampling=parameters.window_downsampling,
             ftype=parameters.ftype,
             time_shift=parameters.time_shift,
-            random_timeshift=parameters.random_timeshift,
+            random_time_shift=parameters.random_time_shift,
             input_sample_rate=parameters.input_sample_rate,
         )
         return ifo
@@ -268,7 +268,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
         window_downsampling = kwargs.pop("window_downsampling")
         ftype = kwargs.pop("ftype")
         time_shift = kwargs.pop("time_shift")
-        random_timeshift = kwargs.pop("random_timeshift")
+        random_time_shift = kwargs.pop("random_time_shift")
         self.duration = segment_duration
         self.timeseries = preprocessing_data_channel_name(
             IFO=self.name,
@@ -285,7 +285,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
             window_downsampling=window_downsampling,
             ftype=ftype,
             time_shift=time_shift,
-            random_timeshift = random_timeshift,
+            random_time_shift = random_time_shift,
             input_sample_rate=input_sample_rate,
         )
         self._check_timeseries_channel_name(channel)
@@ -319,7 +319,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
         window_downsampling = kwargs.pop("window_downsampling")
         ftype = kwargs.pop("ftype")
         time_shift = kwargs.pop("time_shift")
-        random_timeshift = kwargs.pop("random_timeshift")
+        random_time_shift = kwargs.pop("random_time_shift")
         self.duration = segment_duration
         self.timeseries = preprocessing_data_timeseries_array(
             array=timeseries_array,
@@ -333,7 +333,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
             window_downsampling=window_downsampling,
             ftype=ftype,
             time_shift=time_shift,
-            random_timeshift=random_timeshift,
+            random_time_shift=random_time_shift,
         )
         self.timeseries.channel = kwargs.pop("channel")
         self._check_timeseries_sample_rate(new_sample_rate)
@@ -362,7 +362,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
         window_downsampling = kwargs.pop("window_downsampling")
         ftype = kwargs.pop("ftype")
         time_shift = kwargs.pop("time_shift")
-        random_timeshift = kwargs.pop("random_timeshift")
+        random_time_shift = kwargs.pop("random_time_shift")
         self.duration = segment_duration
         self.timeseries = preprocessing_data_gwpy_timeseries(
             gwpy_timeseries=gwpy_timeseries,
@@ -372,7 +372,7 @@ class Interferometer(bilby.gw.detector.Interferometer):
             window_downsampling=window_downsampling,
             ftype=ftype,
             time_shift=time_shift,
-            random_timeshift=random_timeshift,
+            random_time_shift=random_time_shift,
         )
         self.timeseries.channel = kwargs.pop("channel")
         self._check_timeseries_sample_rate(new_sample_rate)
