+++ pygwb/preprocessing.py
@@ -419,7 +419,7 @@ def self_gate_data(
     gated = time_series_data.mask(deadtime=deadtime, const=0, tpad=tpad)
     return gated, deadtime
 
-def shift_timeseries(time_series_data: timeseries.TimeSeries, time_shift: int = 0, random_timeshift = False):
+def shift_timeseries(time_series_data: timeseries.TimeSeries, time_shift: int = 0, random_time_shift = False):
 
     """
     Function that shifts a timeseries by an amount ``time_shift``
@@ -434,17 +434,17 @@ def shift_timeseries(time_series_data: timeseries.TimeSeries, time_shift: int = 
         Value of the time shift (in seconds).
         Default value is 0.
 
-    random_timeshift: ``bool``, optional
-        Adding a random small extra shift to the time_shift.
-        This small shift is bigger than zero, but smaller than one.
-        Default not applied.
+    random_time_shift: ``bool``, optional
+        If True, a random extra shift between 0 and 1 is added to the time_shift.
+        Default is False.
+
 
     Returns
     =======
     shifted_data: ``gwpy.timeseries.TimeSeries``
         TimeSeries containing the shifted_data.
     """
-    if random_timeshift:
+    if random_time_shift:
         extra_shift = random.random()
         if extra_shift == 0:
             extra_shift = random.random()
@@ -477,7 +477,7 @@ def preprocessing_data_gwpy_timeseries(
     window_downsampling: str = "hamming",
     ftype: str = "fir",
     time_shift: int = 0,
-    random_timeshift: bool = False,
+    random_time_shift: bool = False,
 ):
     """
     Function doing the pre-processing of a gwpy timeseries to be used in the remainder of the code.
@@ -508,10 +508,10 @@ def preprocessing_data_gwpy_timeseries(
     time_shift: ``int``, optional
         Value of the time shift (in seconds). Default is 0.
 
-    random_timeshift: ``bool``, optional
-        Adding a random small extra shift to the time_shift.
-        This small shift is bigger than zero, but smaller than one.
-        Default not applied.
+    random_time_shift: ``bool``, optional
+        If True, a random extra shift between 0 and 1 is added to the time_shift.
+        Default is False.
+
 
     Returns
     =======
@@ -528,7 +528,7 @@ def preprocessing_data_gwpy_timeseries(
         ftype=ftype,
     )
     if time_shift > 0:
-        return shift_timeseries(time_series_data=filtered, time_shift=time_shift, random_timeshift = random_timeshift)
+        return shift_timeseries(time_series_data=filtered, time_shift=time_shift, random_time_shift = random_time_shift)
     elif time_shift < 0:
         warnings.warn(
             f"The provided time shift parameter {time_shift} is negative."
@@ -552,7 +552,7 @@ def preprocessing_data_channel_name(
     window_downsampling: str = "hamming",
     ftype: str = "fir",
     time_shift: int = 0,
-    random_timeshift: bool = False,
+    random_time_shift: bool = False,
     local_data_path: str = "",
     frametype: str = "",
     input_sample_rate: int = 16384,
@@ -604,10 +604,9 @@ def preprocessing_data_channel_name(
     time_shift: ``int``, optional
         Value of the time shift (in seconds). Default is 0.
 
-    random_timeshift: ``bool``, optional
-        Adding a random small extra shift to the time_shift.
-        This small shift is bigger than zero, but smaller than one.
-        Default not applied.
+    random_time_shift: ``bool``, optional
+        If True, a random extra shift between 0 and 1 is added to the time_shift.
+        Default is False.
 
     local_data_path: ``str``, optional
         Path where local gwf frame file is stored. Default is an empty string.
[…]
