+++ pygwb_pipe/parameters.ini
@@ -11,7 +11,7 @@ channel: GWOSC-16KHZ_R1_STRAIN
 ; Time shift (in s)
 time_shift: H1:0,L1:0
 ; Adding random shift to time shift
-random_timeshift: False
+random_time_shift: False
 [preprocessing]
 ; Sample rate after resampling
 new_sample_rate: 4096
