### 2024-02-14 a77c750
Update parameters_detector_test.ini
