+++ test/test_data/parameters_detector_test.ini
@@ -5,7 +5,7 @@ tf: 1247644158
 data_type: public
 channel: GWOSC-16KHZ_R1_STRAIN
 time_shift: 0
-random_timeshift: False
+random_time_shift: False
 [preprocessing]
 new_sample_rate: 512
 ; Sample rate of input data
