### 2024-02-14 195fe52
Update detector_test.py
