+++ test/preprocessing_test.py
@@ -20,7 +20,7 @@ class Test(unittest.TestCase):
         self.number_cropped_seconds = 2
         self.data_type = "public"
         self.time_shift = 0
-        self.random_timeshift = False
+        self.random_time_shift = False
         self.sample_rate = 4096
         self.input_sample_rate = 16384
         self.local_data_path = f"{testdir}/test_data/data_gwf_preproc_testing.gwf"
@@ -68,7 +68,7 @@ class Test(unittest.TestCase):
             window_downsampling="hamming",
             ftype="fir",
             time_shift=self.time_shift,
-            random_timeshift=self.random_timeshift,
+            random_time_shift=self.random_time_shift,
             local_data_path=self.local_data_path,
             input_sample_rate=self.input_sample_rate,
         )
@@ -88,7 +88,7 @@ class Test(unittest.TestCase):
             window_downsampling="hamming",
             ftype="fir",
             time_shift=self.time_shift,
-            random_timeshift=self.random_timeshift,
+            random_time_shift=self.random_time_shift,
         )
         self.assertEqual(len(timeseries_output2), 2031616)
         self.assertEqual(timeseries_output2.sample_rate.value, 4096.0) 
@@ -101,7 +101,7 @@ class Test(unittest.TestCase):
             window_downsampling="hamming",
             ftype="fir",
             time_shift=self.time_shift,
-            random_timeshift=self.random_timeshift,
+            random_time_shift=self.random_time_shift,
         )
 
         self.assertEqual(len(timeseries_output3), 2031616)
@@ -121,16 +121,16 @@ class Test(unittest.TestCase):
             time_shifted_data.value[int(1/self.timeseries_data.dt.value)],
         )
         
-        random_timeshifted_data = preprocessing.shift_timeseries(
+        random_time_shifted_data = preprocessing.shift_timeseries(
             time_series_data = self.timeseries_data,
             time_shift = 1,
-            random_timeshift = True
+            random_time_shift = True
         )
-        val1 = random_timeshifted_data.value[0]
+        val1 = random_time_shifted_data.value[0]
         if val1 in self.timeseries_data.value:
             index2 = np.where(self.timeseries_data.value == val1)[0][0]
         self.assertEqual(
-            random_timeshifted_data[0],
+            random_time_shifted_data[0],
             self.timeseries_data.value[index2],
         )            
 
