### 2024-02-20 60ae2ea
adding a test for omega_gw calculation
