+++ pygwb/background.py
@@ -1,3 +1,5 @@
+import warnings
+
 import bilby
 import numpy as np
 from loguru import logger
@@ -78,13 +80,16 @@ def compute_Omega_from_CBC_dictionary(injection_dict, sampling_frequency, T_obs,
     )
     freqs_psd = waveform_generator.frequency_array
     
+    key_1 = list(injection_dict.keys())[0]
     try:
-        N_inj = len(injection_dict['geocent_time']['content'])
+        N_inj = len(injection_dict[key_1]['content'])
     except:
-        N_inj = len(injection_dict['geocent_time'])
+        N_inj = len(injection_dict[key_1])
 
-    Omega_GW_freq = np.zeros(len(freqs_psd))
+    psd_evs = np.zeros(len(freqs_psd))
     logger.info('Compute the total injected Omega for ' + str(N_inj) + ' injections')
+
+    check_injection_dictionary(injection_dict)
 
     # Loop over injections
     for i in tqdm(range(N_inj)):
@@ -98,13 +103,9 @@ def compute_Omega_from_CBC_dictionary(injection_dict, sampling_frequency, T_obs,
         # Get frequency domain waveform
         polarizations = waveform_generator.frequency_domain_strain(inj_params)
         
-        # Final PSD of the injection
-        psd = np.abs(polarizations['plus'])**2 + np.abs(polarizations['cross'])**2 
-        
-        # Add to Omega_spectrum
-        Omega_GW_freq += 2* np.pi**2 * freqs_psd**3 * psd / (3 * H0.si.value**2)
-        
-    Omega_GW_freq *= 2 / T_obs
+        # Add PSDs of individual events
+        psd_evs += np.abs(polarizations['plus'])**2 + np.abs(polarizations['cross'])**2         
+    Omega_GW_freq = 4.* np.pi**2 * freqs_psd**3 * psd_evs / (3 * H0.si.value**2) / T_obs 
 
     logger.debug('Compute Omega_ref at f_ref=' + format(f_ref, '.0f') + ' Hz')
     df = freqs_psd[1] - freqs_psd[0]
@@ -118,8 +119,33 @@ def compute_Omega_from_CBC_dictionary(injection_dict, sampling_frequency, T_obs,
 
     
     if return_spectrum == True:
-        return freqs_psd, Omega_GW_freq
+        return np.array(freqs_psd), np.array(Omega_GW_freq)
     
     else:
         return Omega_ref
 
+def check_injection_dictionary(injection_dict):
+    '''
+    Check injection dict content
+    '''
+    # check for missing keys and set values to defaults
+    if not 'phase' in injection_dict:
+        warnings.warn("'phase' parameter not set in injection set; setting to random value.")
+        injection_dict['phase']=2*np.pi*np.random.rand()
+    if not 'theta_jn' in injection_dict:
+        warnings.warn("'theta_jn' parameter not set in injection set; setting to random value.")
+        cos_inc = np.random.rand()*2.-1.
+        injection_dict['theta_jn']=np.arccos(cos_inc)
+    if not 'a_1' in injection_dict:
+        warnings.warn("'a_1' spin parameter not set in injection set; setting to 0.")
+        injection_dict['a_1']=0
+    if not 'a_2' in injection_dict:
+        warnings.warn("'a_2' spin parameter not set in injection set; setting to 0.")
+        injection_dict['a_2']=0
+    if not 'tilt_1' in injection_dict:
+        warnings.warn("'tilt_1' spin parameter not set in injection set; setting to 0.")
+        injection_dict['tilt_1']=0
+    if not 'tilt_2' in injection_dict:
+        warnings.warn("'tilt_2' spin parameter not set in injection set; setting to 0.")
+        injection_dict['tilt_2']=0
+    
