### 2024-02-22 b8a770d
setting more customizable waveform arguments
