### 2024-02-28 #5454 @dfm
@a-renzini — Can you update the manuscript to use `\mathrm` instead of `\rm`? I think that might fix [the metadata issues above](https://github.com/openjournals/joss-reviews/issues/5454#issuecomment-1967998421).
