+++ JOSS_submission/paper.md
@@ -79,7 +79,7 @@ affiliations:
    index: 11
  - name: Université Côte d’Azur, Observatoire Côte d’Azur, ARTEMIS, Nice, France
    index: 12
- - name: Max Planck Institute for Gravitational Physics (Albert Einstein Institute), D-14476 Potsdam, Germany
+ - name: Max Planck Institute for Gravitational Physics (Albert Einstein Institute), D-14476 Potsdam, Geany
    index: 13
  - name: OzGrav, Charles Sturt University, Wagga Wagga, New South Wales 2678, Australia
    index: 14
@@ -97,13 +97,13 @@ aas-journal: Astrophysical Journal
 
 # Introduction
 
-A gravitational-wave background (GWB) is expected from the superposition of all gravitational waves (GWs) too faint to be detected individually, or by the incoherent overlap of a large number of signals in the same band [@AIReview]. A GWB is primarily characterized by its spectral emission, usually parameterized by the GW fractional energy density spectrum $\Omega_{\rm GW}(f)$, which is the target for stochastic GW searches [@AllenRomano],
+A gravitational-wave background (GWB) is expected from the superposition of all gravitational waves (GWs) too faint to be detected individually, or by the incoherent overlap of a large number of signals in the same band [@AIReview]. A GWB is primarily characterized by its spectral emission, usually parameterized by the GW fractional energy density spectrum $\Omega_{\mathrm {GW}}(f)$, which is the target for stochastic GW searches [@AllenRomano],
 
 $$
-\Omega_{\rm GW}(f) = \frac{1}{\rho_c}\frac{d\rho_{\rm GW}(f)}{d\ln f},
+\Omega_{\mathrm{GW}}(f) = \frac{1}{\rho_c}\frac{d\rho_{\mathrm{GW}}(f)}{d\ln f},
 $$
 
-where $d\rho_{\rm GW}$ is the energy density of GWs in the frequency band $f$ to $f + df$, and $\rho_c$ is the critical energy density of the Universe. Different categories of GW sources may be identified by the unique spectral shape of their background emission; hence, the detection of a GWB will provide invaluable information about the evolution of the Universe and the population of GW sources within it.
+where $d\rho_{\mathrm{GW}}$ is the energy density of GWs in the frequency band $f$ to $f + df$, and $\rho_c$ is the critical energy density of the Universe. Different categories of GW sources may be identified by the unique spectral shape of their background emission; hence, the detection of a GWB will provide invaluable information about the evolution of the Universe and the population of GW sources within it.
 
 # Statement of need
 
@@ -114,13 +114,13 @@ Due to the considerable amount of data to analyze, and the vast panorama of GWB 
 The GWB spectrum estimation implemented in `pygwb` is based on the unbiased minimum variance cross-correlation estimator [@RomanoRev],
 
 $$
-\hat{\Omega}_{{\rm GW}, f} = \frac{{\rm Re}[C_{IJ, f}]}{\gamma_{IJ}(f) S_0(f)}.
+\hat{\Omega}_{{\mathrm{GW}}, f} = \frac{{\mathrm{Re}}[C_{IJ, f}]}{\gamma_{IJ}(f) S_0(f)}.
 $$
 
[…]
