+++ _sources/pipeline.rst.txt
@@ -1,0 +1,291 @@
+=========================================
+Using ``pygwb_pipe``: a quickstart manual
+=========================================
+
+The various modules of the ``pygwb`` package can be combined into a pipeline, as done in the ``pygwb_pipe`` script. This script 
+takes the data as input and outputs an estimator of the point estimate and variance of the gravitational-wave background (GWB) for these
+data. More information on how the various modules interact and are combined into a pipeline can be found in the `pygwb paper <https://arxiv.org/pdf/2303.15696.pdf>`_.
+
+.. note::
+  The proposed ``pygwb_pipe`` pipeline is only one of the many ways to assemble the ``pygwb`` modules, and users should
+  feel free to create their own pipeline, that addresses their needs.
+
+**1. Script parameters**
+========================
+
+The parameters of the ``pygwb_pipe`` script can be visualized by running the following command:
+
+.. code-block:: shell
+
+   pygwb_pipe --help
+
+This will display the following set of parameters, which can be passed to the pipeline:
+
+.. code-block:: shell
+
+  --param_file PARAM_FILE                                                                                                                                                     
+                        Parameter file to use for analysis.                                                                                                                   
+  --output_path OUTPUT_PATH                                                                                                                                                   
+                        Location to save output to.         
+  --calc_coh CALC_COH	 
+                        Calculate coherence spectrum from data.                                                                                                                                      
+  --calc_pt_est CALC_PT_EST                                                                                                                                                   
+                        Calculate omega point estimate and sigma from data.                                                                                                   
+  --apply_dsc APPLY_DSC                                                                                                                                                       
+                        Apply delta sigma cut when calculating final output.                                                                                                  
+  --pickle_out PICKLE_OUT                                                                                                                                                     
+                        Pickle output Baseline of the analysis.                                                                                                               
+  --wipe_ifo WIPE_IFO   Wipe interferometer data to reduce size of pickled Baseline.                                                                                          
+  --t0 T0               Initial time.                                                                                                                                         
+  --tf TF               Final time.                                                                                                                                           
+  --data_type DATA_TYPE                                                                                                                                                       
+                        Type of data to access/download; options are private,                                                                                                 
+                        public, local. Default is public.                                                                                                                     
+  --channel CHANNEL     Channel name; needs to match an existing channel. Default is                                                                                          
+                        "GWOSC-16KHZ_R1_STRAIN"                                                                                                                               
+  --new_sample_rate NEW_SAMPLE_RATE                                                                                                                                           
+                        Sample rate to use when downsampling the data (Hz). Default                                                                                           
+                        is 4096 Hz.                                                                                                                                           
+  --input_sample_rate INPUT_SAMPLE_RATE                                                                                                                                       
[…]
